% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecodata.R
\name{abbreviated_units}
\alias{abbreviated_units}
\title{abbreviated_units}
\usage{
abbreviated_units(x)
}
\arguments{
\item{x}{Numeric, number to abbreviate}
}
\value{
String describing the number, with the appropriate abbreviation
}
\description{
Convert a number to an abbreviated string, using "th", "ml", or "bn" for "thousand", "million", or "billion", as appropriate
}
