% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecodata.R
\name{abbreviated_units_dollar}
\alias{abbreviated_units_dollar}
\title{abbreviated_units_dollar}
\usage{
abbreviated_units_dollar(x)
}
\arguments{
\item{x}{Numeric, number to abbreviate}
}
\value{
String describing the number as a dollar amount, with a dollar sign, commas, and with the appropriate abbreviation
}
\description{
Convert a dollar amount to an abbreviated string, using "th", "ml", or "bn" for "thousand", "million", or "billion", as appropriate
}
