% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc_acl.R
\docType{data}
\name{abc.acl}
\alias{abc.acl}
\title{MAMFC ABC or ACL for Managed Stocks}
\format{
A data set containing 321 rows and 4 columns.

\itemize{
\item Time: Management year.
\item Value: Value of variable \code{Var}
\item Var: Either ABC or ACL depending on what is available for that stock and Catch.
\item EPU: Ecological Production Unit (EPU) where data originated. Here "MAB" refers to the Mid-Atlantic Bight.
}
}
\usage{
abc.acl
}
\description{
These data contain times series of ABC or ACL and Catch for Mid-Atlantic managed
stocks
}
\details{
More information available at \url{https://noaa-edab.github.io/tech-doc/mafmc-abcacl-and-catch.html}
}
\keyword{datasets}
