% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecodata.R
\name{add_ecodata}
\alias{add_ecodata}
\title{`add_ecodata()`
Add data for one or more variables from FRED and/or World Bank Data, for given URLs or variable codes. The function will figure out whether the data is available from FRED or World Bank Data.}
\usage{
add_ecodata(data, varcodes, varnames = NULL, recessions = FALSE)
}
\arguments{
\item{data}{Data frame with existing ecodata, data from FRED and/or World Bank Data}

\item{varcodes}{String for vector of strings that identifies the variable codes or URLs for the data.}

\item{varnames}{Optional, string or vector of strings for the variable names. Default is the code given by the source.}

\item{recessions}{Logical for whether or not to include a dummy variable identifying a NBER U.S. Recessions. Default = FALSE.}
}
\value{
Data frame merged with original data frame that includes the new variables requested. The data frame will also include all relevant meta data describing the data and citing its source.
}
\description{
`add_ecodata()`
Add data for one or more variables from FRED and/or World Bank Data, for given URLs or variable codes. The function will figure out whether the data is available from FRED or World Bank Data.
}
