% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecodata.R
\name{add_ecodata_class}
\alias{add_ecodata_class}
\title{Add 'ecodata' to the class of a date frame}
\usage{
add_ecodata_class(df)
}
\arguments{
\item{df}{A data frame that has the features of an ecodata data frame}
}
\value{
The same data frame, but with the class 'ecodata' added
}
\description{
Add 'ecodata' to the class of a date frame
}
