% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advection.R
\docType{data}
\name{advection}
\alias{advection}
\title{shelf break advection index}
\format{
Data contain 744 rows and 5 columns.
\itemize{
\item Var: Specifies variable of interest.
\item Value: Value of variable \code{Var}.
\item Units: Units of variable \code{Var}.
\item EPU: Ecological Production Unit (EPU) from which data were collected or aggregated. An EPU of \code{All} means that the scope of the data set is the entire Northeast US Continental Shelf.
\item Time: Year.Month
}
}
\usage{
advection
}
\description{
Integrated mass transport through transects along the shelf break from HYCOM
}
\details{
Links to source data, further data definitions, and extended methodology are available at
\url{https://noaa-edab.github.io/tech-doc/advection.html}
}
\keyword{datasets}
