% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_biomass.R
\docType{data}
\name{aggregate_biomass}
\alias{aggregate_biomass}
\title{NEFSC survey data broken into season and aggregate group biomass}
\format{
A data set containing 38008 rows and 6 columns.

\itemize{
\item Time: Survey year.
\item Value: Value of variable \code{Var}
\item Var: Includes SOE species grouping, season that survey occurred (fall or spring), and variable calculated. The specific variables included are stratified mean biomass with associated confidence intervals, coefficients of variation, and standard errors.
\item EPU: Ecological Production Unit (EPU) where data originated. Here "MAB" refers to the Mid-Atlantic Bight.
\item Units: Units of variable \code{Var}.
\item Source: Survdat.
}
}
\usage{
aggregate_biomass
}
\description{
These data contain times series of CPUE, kg tow ^-1, of aggregated species groups from NEFSC surveys.
}
\details{
More information available at \url{https://noaa-edab.github.io/tech-doc/survdat.html}
}
\keyword{datasets}
