% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anolisTree}
\alias{anolisTree}
\title{Feeding of bird nestlings}
\format{
A 'data.frame': 25 obs. of  2 variables
\describe{
  \item{feeding}{items fed to a wird}
  \item{attractiveness}{estimated attractiveness of a bird}
  ...
}
}
\source{
\url{http://biometry.github.io/APES/}
}
\usage{
anolisTree
}
\description{
A dataset containing observationso items fed to bird nestlings,
depending on their attractiveness.
}
\author{
Florian Hartig
}
\keyword{datasets}
