% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquaculture.R
\docType{data}
\name{aquaculture}
\alias{aquaculture}
\title{Oyster aquaculture}
\format{
A data set containing 165 rows and 4 columns.

\itemize{
\item Var: Category of oyster data, production (Pieces), Shellfish lease Acres (Area), Production/Area.
\item Value: Value of variable \code{Var}.
\item Region: Specific state of oyster production.
\item Time: Year.
}
}
\usage{
aquaculture
}
\description{
These data contain time series of number of oysters, the area of oyster lease areas and number of oysters per area.
}
\details{
More information available at \url{https://noaa-edab.github.io/tech-doc/aquaculture.html}
}
\keyword{datasets}
