% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{barbetTree}
\alias{barbetTree}
\title{barbetTree}
\format{
A 'list': List of 5.
\describe{
 \item{edge}{Edges of the tree. First, all leaves are counted like internal nodes, then the tree is traversed from the root edge, going right before left. Each row corresponds to an edge, the first column is the starting node, the second is the target. (2 column matrix, integer).}
 \item{edge.length}{Length of the corresponding edge (same number of elements as rows in tree$edge) (numeric).}
 \item{Nnode}{Number of internal nodes (edges between species); Number of tips / leaves - 1 (integer).}
 \item{root.edge}{Length of the leftmost edge, resp the stem (not displayed every time in plots) (numeric).}
 \item{tip.label}{Names of the species (character).}
}
}
\source{
\url{http://www.phytools.org/Cordoba2017/data/BarbetTree.nex}
}
\usage{
barbetTree
}
\description{
This data set comes from \url{http://www.phytools.org/Cordoba2017} and is about fictional birds and their attributes.
}
\details{
This is a phylogenetic tree, resp. dendrogram belonging to the "barbetData" data set.
Hint: There are some species in the tree for which we have no observations.
}
\examples{
library(ape)
library(effects)
library(lme4)


dat = read.csv(url("http://www.phytools.org/Cordoba2017/data/Barbetdata.csv"),
               header = TRUE, row.names = 1)
tree = read.nexus(url("http://www.phytools.org/Cordoba2017/data/BarbetTree.nex"))

dat$species = as.factor(row.names(dat))
plot(tree)

tree$tip.label \%in\% rownames(dat)


model = lm(log(Lnote) ~ Lnalt + I(Lnalt^.5) + I(Lnalt^2), data = dat)
summary(model)
plot(model)

#plot(allEffects(model, partial.residuals = T), selection = 1)
plot(allEffects(model, partial.residuals = T))

res = simulateResiduals(model)
plot(res)
testDispersion(res)

}
\author{
Florian Hartig
}
\keyword{datasets}
