% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bees}
\alias{bees}
\title{Bees}
\format{
A 'data.frame': 72 obs. of 7 variables:
\describe{
\item{Spobee}{Response variable. Density of spores (numeric).}
\item{Hive}{ID of the respective hive. (integer, 1 - 24).}
\item{X}{Assuming coordinates (integer).}
\item{Y}{Assuming coordinates (numeric).}
\item{Infection}{Degree of infection. (integer: 0 (none), 1 (minor), 2 (moderate), 3 (major)).}
\item{BeesN}{Bees per hive (integer).}
}
}
\usage{
bees
}
\description{
Honeybee data
}
\details{
American Foulbrood (AFB) is an infectious disease for honeybees (Apis mellifera). The spores of the bacterium "Paenibacillus larvae" infect honeybee larvae. These spores are highly infective and one dead infested honeybee larva might spread billions of new spores.
Note: This is nested data, because there are several observations from each hive!
}
\examples{
library(EcoData)
fit <- lm(log(Spobee + 1) ~ Infection + BeesN, data = bees)
summary(fit)



}
\author{
Florian Hartig
}
\keyword{datasets}
