% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benthic_index.R
\docType{data}
\name{benthos_index}
\alias{benthos_index}
\title{Benthic Index}
\format{
n rows and 5 columns.

\itemize{
\item Value: Value of variable \code{Var}.
\item Var: Variable calculated.
\item Time: Year.
\item EPU: Ecological Production Unit (EPU) where sampling occurred.
\item Units: grams per stomach.
}
}
\usage{
benthos_index
}
\description{
Data include Megabenthos and Macrobenthos biomass and center of gravity indices from predator diet data from NEFSC bottom trawl survey.
}
\details{
Methods used to calculate data are available at \url{https://noaa-edab.github.io/tech-doc/benthos_index.html}.
}
\keyword{datasets}
