% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{birdabundance}
\alias{birdabundance}
\title{Bird Abundance}
\format{
A 'data.frame': 56 obs. of 8 variables:
\describe{
 \item{SITE}{Site ID.}
 \item{ABUND}{Bird abundance.}
 \item{AREA}{Area.}
 \item{DIST}{Distance to the next forest fragment.}
 \item{LDIST}{Distance to the next large forest fragment}
 \item{YR.ISOL}{Years since the site was isolated / fragmented.}
 \item{GRAZE}{Grazing intensity in the surrounding.}
 \item{ALT}{Altitude.}
}
}
\source{
\url{http://biometry.github.io/APES/}
}
\usage{
birdabundance
}
\description{
Abundance of birds in forest fragments in Australia
}
\examples{
birdabundance_df <- birdabundance
head(birdabundance_df)

str(birdabundance_df)
plot(birdabundance_df$ABUND, c(1:56))
plot(birdabundance_df$ABUND, birdabundance_df$GRAZE)


library(glmmTMB)
# gaussian model -------------------------------------------------
fit <-  glmmTMB(ABUND ~ AREA + DIST ,family = gaussian, data = birdabundance_df)
summary(fit)



library(DHARMa)
res <- simulateResiduals(fit)
plot(res)


DHARMa::testQuantiles(res)
DHARMa::testOutliers(res)
}
\author{
Florian Hartig
}
\keyword{datasets}
