% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{birdfeeding}
\alias{birdfeeding}
\title{Feeding of Bird Nestlings}
\format{
A 'data.frame': 25 obs. of 2 variables
\describe{
 \item{feeding}{Items fed to a bird.}
 \item{attractiveness}{Estimated attractiveness of a bird.}
  ...
}
}
\source{
\url{http://biometry.github.io/APES/}
}
\usage{
birdfeeding
}
\description{
A data set containing observations of items fed to bird nestlings,
depending on their attractiveness.
}
\examples{
birdfeeding_df <- birdfeeding
head(birdfeeding_df)

str(birdfeeding_df)
plot(birdfeeding_df$feeding, birdfeeding_df$attractiveness)



library(glmmTMB)
# poisson model -------------------------------------------------
fit <-  glmmTMB(feeding ~ attractiveness ,family = poisson, data = birdfeeding_df)
summary(fit)



library(DHARMa)
res <- simulateResiduals(fit)
plot(res)
}
\author{
Florian Hartig
}
\keyword{datasets}
