% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blue_runner.R
\docType{data}
\name{blue_runner}
\alias{blue_runner}
\title{SAFMC managed Blue Runner presence on NES}
\format{
Data set includes 1152 rows and 6 columns.

\itemize{
\item Var: Either lat or long of Blue runner present on NES broken into 3 time catgories (prior to 2001, 2001-2010, 2010-pres)
\item Value: Value of variable \code{Var}.
\item Time: unused
\item Units: decimal degrees
\item EPU: Ecological Production Unit (EPU) where sampling occurred.
\item Source: NEFSC bottom trawl survey.
}
}
\usage{
blue_runner
}
\description{
Data include spatial, presence/absence data of South Atlantic FMC managed species
blue runner (Caranx crysos) found on the NES.
}
\details{
Links to source data, further data definitions, and extended methodology are available at
\url{https://noaa-edab.github.io/tech-doc/safmc-managed-spp.html}
}
\keyword{datasets}
