% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bottom_temp.R
\docType{data}
\name{bottom_temp}
\alias{bottom_temp}
\title{Bottom temperature}
\format{
Data contain 648 rows and 5 columns.
\itemize{
\item Var: Specifies variable of interest.
\item Value: Value of variable \code{Var}.
\item Units: Units of variable \code{Var}.
\item EPU: Ecological Production Unit (EPU) from which data were collected or aggregated. An EPU of \code{All} means that the scope of the data set is the entire Northeast US Continental Shelf.
\item Time: Year.
}
}
\usage{
bottom_temp
}
\description{
Time series of annual in situ bottom temperatures on the Northeast Continental
Shelf derived from NEFSC surveys 1977-pres.
}
\details{
Links to source data, further data definitions, and extended methodology are available at
\url{https://noaa-edab.github.io/tech-doc/bottom-temperatures.html}
}
\keyword{datasets}
