% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bottom_temp_comp.R
\docType{data}
\name{bottom_temp_comp}
\alias{bottom_temp_comp}
\title{Bottom temperature - High Resolution}
\format{
Data contain 1280 rows and 5 columns.
\itemize{
\item Var: Specifies variable of interest.
\item Value: Value of variable \code{Var}.
\item Units: Units of variable \code{Var}.
\item EPU: Ecological Production Unit (EPU) from which data were collected or aggregated. An EPU of \code{All} means that the scope of the data set is the entire Northeast US Continental Shelf.
\item Time: Year.
}
}
\usage{
bottom_temp_comp
}
\description{
Time series of annual modeled bottom temperatures on the Northeast Continental
Shelf from 1959-pres.
}
\details{
Links to source data, further data definitions, and extended methodology are available at
\url{https://noaa-edab.github.io/tech-doc/bottom-temperature---high-resolution.html}
}
\keyword{datasets}
