% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bottom_temp_glorys.R
\docType{data}
\name{bottom_temp_glorys}
\alias{bottom_temp_glorys}
\title{Bottom temperature - GLORYS}
\format{
Data contain 104 rows and 5 columns.
\itemize{
\item Var: Specifies variable of interest.
\item Value: Value of variable \code{Var}.
\item Units: Units of variable \code{Var}.
\item EPU: Ecological Production Unit (EPU) from which data were collected or aggregated. An EPU of \code{All} means that the scope of the data set is the entire Northeast US Continental Shelf.
\item Time: Year.
}
}
\usage{
bottom_temp_glorys
}
\description{
Time series of annual modelled bottom temperatures on the Northeast Continental
Shelf.
}
\details{
Links to source data, further data definitions, and extended methodology are available at
\url{https://noaa-edab.github.io/tech-doc/bottom-temperature-glorys.html}
}
\keyword{datasets}
