% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bottom_temp_model_gridded.R
\docType{data}
\name{bottom_temp_model_gridded}
\alias{bottom_temp_model_gridded}
\title{Bottom temperature - Seasonal Gridded Product}
\format{
Data contain raster files for each season.
\itemize{
\item Raster stack of 4 layers containing bottom temperature anomalies.
}
}
\usage{
bottom_temp_model_gridded
}
\description{
Spatially modeled bottom temperatures by season on the Northeast Continental
Shelf from most recent year.
}
\details{
Links to source data, further data definitions, and extended methodology are available at
\url{https://noaa-edab.github.io/tech-doc/bottom-temperature---high-resolution.html}
}
\keyword{datasets}
