% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_bay_temp.R
\docType{data}
\name{ch_bay_temp}
\alias{ch_bay_temp}
\title{Chesapeake Bay temperature with historical trends for comparison}
\format{
A data set containing 1376 rows and 5 columns.

\itemize{
\item Time: Day
\item Value: Value of variable \code{Var}
\item Var: Temperature measurement
\item EPU: Ecological Production Unit (EPU) where data originated. Here "MAB" refers to the Mid-Atlantic Bight.
\item Units: Units of variable \code{Var}.

}
}
\usage{
ch_bay_temp
}
\description{
Chesapeake Bay temperature with historical trends for comparison
}
\details{
More information available at \url{https://noaa-edab.github.io/tech-doc/https://noaa-edab.github.io/tech-doc/chesapeake-bay-salinity-and-temperature.html}
}
\keyword{datasets}
