% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ches_bay_sf.R
\docType{data}
\name{ches_bay_sf}
\alias{ches_bay_sf}
\title{Chesapeake Bay shapefiles}
\format{
Spatial feature outlining Chesapeake Bay

\itemize{
\item Spatial polygons

}
}
\usage{
ches_bay_sf
}
\description{
Chesapeake Bay shapefiles
}
\details{
More information available at \url{https://noaa-edab.github.io/tech-doc/https://noaa-edab.github.io/tech-doc/chesapeake-bay-salinity-and-temperature.html}
}
\keyword{datasets}
