% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ches_bay_sst.R
\docType{data}
\name{ches_bay_sst}
\alias{ches_bay_sst}
\title{Chesapeake Bay Gridded Sea Surface Temperature}
\format{
A data set containing 119577 rows and 4 columns.

\itemize{
\item Latitude: Latitude of sea surface grid.
\item Longitude: Longitude of sea surface grid.
\item Value: Tempurature in Celsius.
\item Season: Annual Season.

}
}
\usage{
ches_bay_sst
}
\description{
Chesapeake Bay Gridded Sea Surface Temperature
}
\details{
More information available at \url{https://noaa-edab.github.io/tech-doc/chesapeake-bay-seasonal-sst-anomalies.html}
}
\keyword{datasets}
