% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coast.R
\docType{data}
\name{coast}
\alias{coast}
\title{Coast}
\format{This data set is a spatial (sf) object.}
\usage{
coast
}
\description{
Coast is a spatial dataset that outlines the east coast of the United States.
To be used in plotting the NES
}
\keyword{datasets}
