% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cold_pool.R
\docType{data}
\name{cold_pool}
\alias{cold_pool}
\title{Cold pool index}
\format{
A data set containing 41 rows and 5 columns.

\itemize{
\item Time: Year
\item Value: Value of variable.
\item EPU: Ecological Production Unit (EPU) where data originated. Here "MAB" refers to the Mid-Atlantic Bight.
\item Units: Anomaly.
\item Source: NEFSC Hydro Data (OCDBS).

}
}
\usage{
cold_pool
}
\description{
Cold pool index
}
\details{
More information available at \url{https://noaa-edab.github.io/tech-doc/cold-pool-index.html}
}
\keyword{datasets}
