% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cold_pool_sf.R
\docType{data}
\name{cold_pool_sf}
\alias{cold_pool_sf}
\title{Cold pool sf index}
\format{
Simple feature collection with 26 features and 0 fields

\itemize{
\item geometry: yearly spatial feature for the cold pool.

}
}
\usage{
cold_pool_sf
}
\description{
Cold pool sf index
}
\details{
More information available at \url{https://noaa-edab.github.io/tech-doc/cold-pool-index.html}
}
\keyword{datasets}
