% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community_climate_vulnerability.R
\docType{data}
\name{community_climate_vulnerability}
\alias{community_climate_vulnerability}
\title{Community climate vulnerability indicators}
\format{
273424 rows and 5 columns.

\itemize{
\item Time: Year.
\item Var: Community, State-climate vulnerability indicator.
\item Value: Value of climate vulnerability indicator.
\item EPU: Region where sampling occurred: NE New England, MAB Mid-Atlantic.
\item Units: Units of climate vulnerability indicator.
}
}
\usage{
community_climate_vulnerability
}
\description{
Data include time series of community scores for community sensitivity to temperature,
ocean acidification, stock size/status, total sensitivity, and total vulnerability
based on Hare et al. 2016 species vulnerability and community dependency for
coastal communities in the Northeast US.
}
\details{
Reciprocal Simpson's Diversity scores for each community, Regional Quotients,
and Regional climate vulnerability scores are also included. More information
including processing and indicator derivation steps are available at
\url{https://noaa-edab.github.io/tech-doc/community-climate-vulnerability.html}.
}
\keyword{datasets}
