% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community_risks.R
\docType{data}
\name{community_risks}
\alias{community_risks}
\title{Community risk indicators (CEVRI)}
\format{
273424 rows and 5 columns.

\itemize{
\item Time: Year.
\item Var: Community, State-climate vulnerability indicator.
\item Value: Value of climate vulnerability indicator.
\item EPU: Region where sampling occurred: NE New England, MAB Mid-Atlantic.
\item Units: Units of climate vulnerability indicator.
}
}
\usage{
community_risks
}
\description{
Data include time series of Northeast U.S. community scores for community commercial fishing
revenue sensitivity to temperature, ocean acidification, stock size/status,
total sensitivity, and total vulnerability based on Hare et al. 2016. species
vulnerability.
}
\details{
Reciprocal Simpson's Diversity scores for each community, Regional Quotients,
and Regional climate vulnerability scores are also included. More information
including processing and indicator derivation steps are available at
\url{https://noaa-edab.github.io/tech-doc/community-climate-vulnerability.html}.
}
\keyword{datasets}
