% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition.R
\docType{data}
\name{condition}
\alias{condition}
\title{Fish Condition}
\format{
These data include 1612 rows and 4 columns.

\itemize{
\item Var: Species ID code
\item Value: Value of variable \code{Var}.
\item Units: Units of variable \code{Var}.
\item Time: Year.
}
}
\usage{
condition
}
\description{
Mean relative condition calculated from the NEFSC bottom trawl survey.
}
\details{
These data are fully documented at \url{https://noaa-edab.github.io/tech-doc/fish-condition-indicator.html}.
}
\keyword{datasets}
