% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_all_plots.R
\name{create_all_plots}
\alias{create_all_plots}
\title{Create all variations of an ecodata plot function}
\usage{
create_all_plots(ecodata_name = NULL, write_only = FALSE, print = TRUE, n = 0)
}
\arguments{
\item{ecodata_name}{string. the data name of the target ecodata dataset or plot function (default = NULL)}

\item{write_only}{boolean. should the function output the plotting code, as text, for all plot variations (default = FALSE)}

\item{print}{boolean. should the function print output to the session window (default = TRUE)}

\item{n}{numeric. number of data points for short term trend (default = 0)}
}
\value{
list containing plots when write_only = FALSE and text when write_only = TRUE. plots are also exported to the IDE plot pane.
}
\description{
This function can flexibly plot all possible variations given the arguments and attributes of an ecodata plot function
This function can also print all of the code needed to run all possible variations
This may be useful for reviewing data submissions, creating catalog pages, automating QA/QC protocols and comparing against other ecodata versions.
}
