% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crew_survey.R
\docType{data}
\name{crew_survey}
\alias{crew_survey}
\title{Crew Survey}
\format{
A data set containing 3048 rows and 7 columns.

\itemize{
\item Time: Year
\item Value: Value of Metric
\item Var: Short-hand survey question name.
\item Var.long: original survey question name.
\item Var.Answer: categorical answer to question
\item Metric: metric evaluated for each question
\item EPU: Ecological Production Unit (EPU) where data originated. Here "MAB" refers to the Mid-Atlantic Bight.

}
}
\usage{
crew_survey
}
\description{
The crew survey is an intercept survey that evaluates demographic and job satisfaction information of crew and hired captains
}
\details{
More information available at \url{https://noaa-edab.github.io/tech-doc/crew_survey.html}
}
\keyword{datasets}
