% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissolved_oxygen.R
\docType{data}
\name{dissolved_oxygen}
\alias{dissolved_oxygen}
\title{Dissolved Oxygen from FishBOT}
\format{
A data set containing 14736 rows and 5 columns.

\itemize{
\item Time: Date
\item Lon: Longitude
\item Lat: Latitude
\item Value: Value of variable.
\item Var: Dissolved Oxygen obervations: (measured = 'dissolved_oxygen', count = 'dissolved_oxygen_count', minimum = 'dissolved_oxygen_min', or maximum = 'dissolved_oxygen_max')

}
}
\usage{
dissolved_oxygen
}
\description{
Gridded observations of dissolved oxygen from FishBOT
}
\details{
More information available at \url{https://noaa-edab.github.io/tech-doc/dissolved_oxygen.html}
}
\keyword{datasets}
