% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecodata.R
\name{ecodata_cite}
\alias{ecodata_cite}
\title{`ecodata_cite(data)`
Get citation information for every variable in the given ecodata data frame}
\usage{
ecodata_cite(data)
}
\arguments{
\item{data}{Data frame from `get_ecodata()` that includes units information in the meta data}
}
\value{
Returns a data frame with a row for each variable in the given data frame, with a column for how to cite each variable
}
\description{
`ecodata_cite(data)`
Get citation information for every variable in the given ecodata data frame
}
\seealso{
[ecodata_cite_table()]
}
