% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecodata.R
\name{ecodata_colorscale}
\alias{ecodata_colorscale}
\title{ecodata_colorscale}
\usage{
ecodata_colorscale(n)
}
\arguments{
\item{n}{Number of colors to return, n must be between 1 and 7}
}
\value{
Returns a vector of color values
}
\description{
Return up to 6 colors to use for a categorical color scale in a plot
}
