% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecodata.R
\name{ecodata_description}
\alias{ecodata_description}
\title{`ecodata_description(data)`
Get a description of every variable in the given ecodata data frame}
\usage{
ecodata_description(data)
}
\arguments{
\item{data}{Data frame from `get_ecodata()` that includes units information in the meta data}
}
\value{
Returns a data frame with a row for each variable in the given data frame, and several columns describing attributes of each variable
}
\description{
`ecodata_description(data)`
Get a description of every variable in the given ecodata data frame
}
\seealso{
[ecodata_description_table()]
}
