% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecodata.R
\name{ecodata_fred_hasapi}
\alias{ecodata_fred_hasapi}
\title{Test FRED API}
\usage{
ecodata_fred_hasapi()
}
\description{
Check to see if the FRED API key already exists and is loaded. If it exists as an environmental variable, but is not loaded, it will attempt to load it. Halts execution if the API key is not found.
}
