% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecodata.R
\name{ecodata_get_attributes}
\alias{ecodata_get_attributes}
\title{Get a Given Attribute for All Columns}
\usage{
ecodata_get_attributes(df, attribute)
}
\arguments{
\item{df}{Data frame from `get_ecodata()` that includes units information in the meta data}

\item{attribute}{Character string, attribute to retrieve. Eg: "Units", "Frequency", "Source", etc.}
}
\value{
A vector of unique source descriptions from the data frame
}
\description{
Get unique values for an attribute for the all the variables in the ecodata data frame
}
