% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecodata.R
\name{ecodata_get_fredkey}
\alias{ecodata_get_fredkey}
\title{Get FRED API key}
\usage{
ecodata_get_fredkey()
}
\value{
String that is the FRED API key. Empty string if it doesn't exist
}
\description{
Get the FRED API key if it exists. Returns empty string if it doesn't exist
}
