% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecodata.R
\name{ecodata_get_frequencies}
\alias{ecodata_get_frequencies}
\title{Get all frequencies}
\usage{
ecodata_get_frequencies(df)
}
\arguments{
\item{df}{Data frame from `get_ecodata()` that includes units information in the meta data}
}
\value{
A vector of unique unit descriptions from the data frame
}
\description{
Get unique frequencies for the variables in the given data frame
}
