% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecodata.R
\name{ecodata_get_sources}
\alias{ecodata_get_sources}
\title{Get Data Sources}
\usage{
ecodata_get_sources(df)
}
\arguments{
\item{df}{Data frame from `get_ecodata()` that includes units information in the meta data}
}
\value{
A vector of unique source descriptions from the data frame
}
\description{
Get unique sources for the variables in the ecodata data frame
}
