% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecodata.R
\name{ecodata_get_units}
\alias{ecodata_get_units}
\title{`ecodata_get_units(df)`
Get unique units for the variables in the given data frame}
\usage{
ecodata_get_units(df)
}
\arguments{
\item{df}{Data frame from `get_ecodata()` that includes units information in the meta data}
}
\value{
A vector of unique unit descriptions from the data frame
}
\description{
`ecodata_get_units(df)`
Get unique units for the variables in the given data frame
}
