% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecodata.R
\name{ecodata_join}
\alias{ecodata_join}
\title{Full join multiple ecodata data frames}
\usage{
ecodata_join(...)
}
\arguments{
\item{...}{Data frames to join.}
}
\value{
An ecodata data frame containing the full join of all input ecodata data frames
}
\description{
This function takes any number of data frames, checks that they all contain
a "Date" column, and joins them using `dplyr::full_join()`.
}
