% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecodata.R
\name{ecodata_linear_interpolate}
\alias{ecodata_linear_interpolate}
\title{Perform a linear interpolation for missing values}
\usage{
ecodata_linear_interpolate(df, variable_names = NULL)
}
\arguments{
\item{df}{An ecodata data frame}

\item{variable_names}{Optional, vector of strings for the names of the variables in `df` to perform the linear interpolation. Default is to perform the operation on all the variables.}
}
\value{
An ecodata data frame with the same structure as `df`, with NA values replaced with linear interpolations.
}
\description{
Perform a linear interpolation for any missing values for an ecodata data frame. This is useful for plotting time series with mixed frequencies.
}
