% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecodata.R
\name{ecodata_scale_dollar}
\alias{ecodata_scale_dollar}
\title{Modify the y-axis scale to use dollar formatting while keeping other settings intact.}
\usage{
ecodata_scale_dollar(axis = "y")
}
\arguments{
\item{axis}{Optional, character "x" or "y" for which axis to set}
}
\value{
A scale to add on to a gggplot
}
\description{
Modify the y-axis scale to use dollar formatting while keeping other settings intact.
}
