% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecodata.R
\name{ecodata_set_fredkey}
\alias{ecodata_set_fredkey}
\title{Set the FRED API key}
\usage{
ecodata_set_fredkey(API_Key)
}
\arguments{
\item{API_Key}{String that is the API key for FRED. Go to https://fredaccount.stlouisfed.org/apikeys to request a key.}
}
\description{
Set the FRED API key as a permanent environmental variable, and make it available immediately
This will append the ~/.Renviron file
}
