% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecodata.R
\name{ecodata_theme}
\alias{ecodata_theme}
\title{`ecodata_theme()`
A ggplot theme that is simple and clean and has a large text size}
\usage{
ecodata_theme()
}
\value{
Returns a ggplot theme that can be added to a `ggplot()` call
}
\description{
`ecodata_theme()`
A ggplot theme that is simple and clean and has a large text size
}
