% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecodata.R
\name{ecodata_theme_large}
\alias{ecodata_theme_large}
\title{Theme with large text}
\usage{
ecodata_theme_large()
}
\value{
Returns a ggplot theme that can be added to a `ggplot()` call
}
\description{
A ggplot theme that is simple and clean and has a large text size
}
