% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{elemental}
\alias{elemental}
\title{Elemental}
\format{
An object of class \code{data.frame} with 524 rows and 27 columns.
}
\usage{
elemental
}
\description{
Dataset on N composition
}
\details{
Detailed explanation at https://esajournals.onlinelibrary.wiley.com/doi/full/10.1002/ecs2.3940
}
\examples{
elemental

# traditional lm, arc sine transformed response of proportions, see critique in https://esajournals.onlinelibrary.wiley.com/doi/10.1890/10-0340.1
m1 <- lm(N_arc ~ Year + Site + Species, data = elemental)
summary(m1)

# beta regression
library(glmmTMB)
m2 <- glmmTMB(N_dec ~ Year + Site + Species, family = beta_family, data = elemental )
summary(m2)

}
\author{
Florian Hartig
}
\keyword{datasets}
