% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{elk}
\alias{elk}
\title{Elk}
\format{
A data frame with 3848 rows and 6 variables::
\describe{
\item{dist_roads}{distance of the observation to the next road in m}
\item{NDVI}{Normalised Difference Vegetation Index = greeness of the vegetation, proxy for vegetation health and productivity}
\item{ruggedness}{unevenness of the terrain}
\item{dem}{digitale elevation model = elevation}
\item{presence}{0 = absence / 1 = presence of elks at the location}
\item{habitat}{factor with two levels: "forest","open"}
}
}
\usage{
elk
}
\description{
Data of elk presence in Canada
}
\details{
The dataset contains presence / absence data of elks in Canada.

An interesting goal for an analysis is to examine the effect of roads on elk presence, where one can find effect sizes switching (Simpson's paradox) depending on what covariates are adjusted for.
}
\examples{

plot(as.factor(elk$presence) ~ elk$dist_roads)
m0 = glm(presence ~ dist_roads, data = elk)
summary(m0)
}
\author{
Florian Hartig
}
\keyword{datasets}
