% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_density.R
\docType{data}
\name{energy_density}
\alias{energy_density}
\title{Forage fish energy density}
\format{
A data set containing 264 rows and 4 columns.

\itemize{
\item Year: Time.
\item Var: Species / Season / Variable
\item Value: Value of Variable
\item EPU: Ecological Production Unit (NA).
}
}
\usage{
energy_density
}
\description{
Forage fish energy density comparison between studies in the 80s and 90s.
}
\details{
More information available at \url{https://noaa-edab.github.io/tech-doc/forage-fish-energy-density.html}
}
\keyword{datasets}
