% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epu_sf.R
\docType{data}
\name{epu_sf}
\alias{epu_sf}
\title{Ecological Production Units}
\format{
This data set is a \code{sf} object with 4 features and 3 fields.
}
\usage{
epu_sf
}
\description{
Ecological Production Units (EPUs) delineate regions unique in physiographic, oceanographic and biotic aspects
on the Northeast U.S. Continental Shelf, representing an area of approximately 264,000 km within the 200 m
isobath. There are four EPUs on the Shelf, including the Mid-Atlantic Bight (MAB), Georges Bank (GB), Gulf of
Maine (GOM), and Scotian Shelf (SS). The Scotian Shelf lies outside the management domain of US Fishery Management
Councils and is generally not considered in State of the Ecosystem Reports.
}
\details{
More information about the delineation of Ecological Production Units can be found at
\url{https://noaa-edab.github.io/tech-doc/epu.html}.
}
\keyword{datasets}
