% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forage_anomaly.R
\docType{data}
\name{forage_anomaly}
\alias{forage_anomaly}
\title{Forage Anomaly}
\format{
240 rows and 4 columns.

\itemize{
\item Value: Value of variable \code{Var}.
\item Var: Variable calculated.
\item Time: Time step of \code{Var}.
\item EPU: Ecological Production Unit (EPU) where sampling occurred.

}
}
\usage{
forage_anomaly
}
\description{
Data include expected mean and uncertainty of forage anomaly from the ecomon survey.
}
\details{
Methods used to calculate data are available at \url{https://noaa-edab.github.io/tech-doc/plankton-diversity.html}.
}
\keyword{datasets}
