% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forage_index.R
\docType{data}
\name{forage_index}
\alias{forage_index}
\title{Forage Index}
\format{
888 rows and 5 columns.

\itemize{
\item Value: Value of variable \code{Var}.
\item Var: Variable calculated.
\item Time: Year.
\item EPU: Ecological Production Unit (EPU) where sampling occurred.
\item Units: grams per stomach.
}
}
\usage{
forage_index
}
\description{
Data include forage fish biomass index from predator diet data from NEFSC bottom trawl survey.
}
\details{
Methods used to calculate data are available at \url{https://noaa-edab.github.io/tech-doc/forage-fish-biomass-indices.html}.
}
\keyword{datasets}
