% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fram_pfmc_2019}
\alias{fram_pfmc_2019}
\title{FRAM data from 2019 PFMC working group for salmon dashboard project}
\format{
A data frame.
}
\usage{
fram_pfmc_2019
}
\description{
FRAM data from 2019 PFMC working group for salmon dashboard project
}
\keyword{datasets}
