% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecodata.R
\name{geom_recession}
\alias{geom_recession}
\title{`geom_recession()`
Return a geom_rect() that identifies recessions in a time series plot}
\usage{
geom_recession(
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  alpha = 0.2,
  fill = "dodgerblue3",
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{data}{Data frame}

\item{alpha}{Numeric transparency parameter. Default = 0.2.}

\item{fill}{String identifying the fill color, default is "dodgerblue3"}
}
\value{
A geom_rect() that can be added to a ggplot time series plot
}
\description{
`geom_recession()`
Return a geom_rect() that identifies recessions in a time series plot
}
