% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_regime.R
\name{geom_regime}
\alias{geom_regime}
\title{A geom to fit a regime shifts to a time series}
\usage{
geom_regime(
  data = NULL,
  mapping = NULL,
  stat = "REGIME",
  geom = "vline",
  position = "identity",
  inherit.aes = TRUE,
  na.rm = FALSE,
  color = "red",
  show.legend = NA
)
}
\arguments{
\item{data}{Input series to be analyzed. If NULL, data is inherited from previous layer or \code{ggplot} call.}

\item{mapping}{Set of aesthetic mappings created by \code{aes()}. By default \code{inherit.aes = TRUE}, which
assigns the top-level plotting \code{aes()} to the GLS geom.}

\item{stat}{stat}

\item{geom}{geom}

\item{position}{position}

\item{inherit.aes}{inherit aesthetics}

\item{na.rm}{remove NAs}

\item{color}{color}

\item{show.legend}{show legend}

\item{...}{Other arguments may be passed to the stat, including fixed aesthetics.}
}
\value{
If regime shifts exists, then xintercept line(s) is returned. Otherwise output is \code{NULL}.
}
\description{
A geom to fit a regime shifts to a time series
}
\examples{
\dontrun{
library(ggplot2)

#Generate series

m <- 0.1
x <- 1:30
y <-  m*x + rnorm(30, sd = 0.35)

data <- data.frame(x = x,
                  y = y)

#Plot series with trend
ggplot(data = data) +
  geom_line(aes(x = x, y = y)) +
  geom_regime()
}
}
