% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecodata.R
\docType{data}
\name{GeomRecession}
\alias{GeomRecession}
\title{Definition of the GeomRecession}
\format{
An object of class \code{GeomRecession} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 4.
}
\usage{
GeomRecession
}
\description{
Definition of the GeomRecession
}
\keyword{datasets}
