% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bluebook.R
\name{get_bluebook}
\alias{get_bluebook}
\title{get_bluebook grabs processed bluebook data off of Github, and joins in the OBJECTID}
\usage{
get_bluebook()
}
\value{
null, data file written to inst/ folder
}
\description{
get_bluebook grabs processed bluebook data off of Github, and joins in the OBJECTID
}
\examples{
\dontrun{
get_bluebook()
}
}
