% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_buoy.R
\name{get_buoy}
\alias{get_buoy}
\title{get_buoy is a wrapper for rnoaa::buoy, to access the National Buoy Data Center}
\usage{
get_buoy(years = NULL)
}
\value{
null, data file written to inst/ folder
}
\description{
get_buoy is a wrapper for rnoaa::buoy, to access the National Buoy Data Center
}
\examples{
\dontrun{
get_buoy()
}
}
