% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_calcofi.R
\name{get_calcofi}
\alias{get_calcofi}
\title{fit_bycatch is the primary function for fitting bycatch models to time series of takes and effort}
\usage{
get_calcofi(min_year = 1985, min_years = 30, min_n = 300)
}
\arguments{
\item{min_year}{The minimum year used, defaults to 1985}

\item{min_years}{The minimum number of years with positive observations that a species needs to have to be included, defaults to 30}

\item{min_n}{The minimum number of occurrences a species needs to have across all years, defaults to 300}
}
\value{
null, data file written to inst/ folder
}
\description{
fit_bycatch is the primary function for fitting bycatch models to time series of takes and effort
}
\examples{
\dontrun{
get_calcofi(min_year = 1990, min_years = 10, min_n = 100)
}
}
