% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_calcofi_indices.R
\name{get_calcofi_indices}
\alias{get_calcofi_indices}
\title{get_calcofi_indices grabs processed indices off of Github, and joins in the OBJECTID}
\usage{
get_calcofi_indices()
}
\value{
null, data file written to inst/ folder
}
\description{
get_calcofi_indices grabs processed indices off of Github, and joins in the OBJECTID
}
\examples{
\dontrun{
get_calcofi_indices()
}
}
