% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dart.R
\name{get_dart}
\alias{get_dart}
\title{get_dart grabs processed DART data off of Github, and joins in the OBJECTID}
\usage{
get_dart()
}
\value{
null, data file written to inst/ folder
}
\description{
get_dart grabs processed DART data off of Github, and joins in the OBJECTID
}
\examples{
\dontrun{
get_dart()
}
}
