% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecodata.R
\name{get_ecodata}
\alias{get_ecodata}
\title{`get_ecodata()`
Get data for one or more variables from FRED and/or World Bank Data, for given URLs or variable codes. The function will figure out whether the data is available from FRED or World Bank Data.}
\usage{
get_ecodata(varcodes, varnames = NULL, recessions = FALSE)
}
\arguments{
\item{varcodes}{String for vector of strings that identifies the variable codes or URLs for the data.}

\item{varnames}{Optional, string or vector of strings for the variable names. Default is the code given by the source.}

\item{recessions}{Logical for whether or not to include a dummy variable identifying a NBER U.S. Recessions. Default = FALSE.}
}
\value{
Data frame time series that includes the date and the variable requested. The data frame will also include all relevant meta data describing the data and citing its source.
}
\description{
`get_ecodata()`
Get data for one or more variables from FRED and/or World Bank Data, for given URLs or variable codes. The function will figure out whether the data is available from FRED or World Bank Data.
}
