% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecodata.R
\name{get_ecodata_allcountries_wb}
\alias{get_ecodata_allcountries_wb}
\title{Retrieve data from World Bank Data for all countries}
\usage{
get_ecodata_allcountries_wb(varcode, varname = NULL)
}
\arguments{
\item{varcode}{String that is the URL for the data, variable code, or the partial URL including the variable code. If a location is included in the `varcode`, it will be ignored.}

\item{varname}{Optional, string to give the variable name. By default, the variable will be named the variable code given in World Bank Data.}
}
\value{
Data frame for the single variable from World Bank Data. The data frame will include a column for every country. The data frame will also include all relevant meta data describing the data and citing its source.
}
\description{
Retrieve data for a single variable from World Bank Data for all available countries, given the URL for the data location, variable code, or the partial URL including the variable code.
}
