% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecodata.R
\name{get_ecodata_variable}
\alias{get_ecodata_variable}
\title{`get_ecodata_variable()`
Get data for a single variable from either FRED or World Bank Data, for a given URL or variable code. The function will figure out whether the data is available from FRED or World Bank Data.}
\usage{
get_ecodata_variable(varcode, varname = NULL, recessions = FALSE)
}
\arguments{
\item{varcode}{String that identifies the variable code or URL for the data}

\item{varname}{Optional, string for the variable name. Default is the code given by the source.}

\item{recessions}{Logical for whether or not to include a dummy variable identifying a NBER U.S. Recessions. Default = FALSE.}
}
\value{
Data frame time series that includes the date and the variable requested. The data frame will also include all relevant meta data describing the data and citing its source.
}
\description{
`get_ecodata_variable()`
Get data for a single variable from either FRED or World Bank Data, for a given URL or variable code. The function will figure out whether the data is available from FRED or World Bank Data.
}
