% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecodata.R
\name{get_ecodata_variable_country_wb}
\alias{get_ecodata_variable_country_wb}
\title{`get_ecodata_variable_wb()`
Downloads data from World Bank for a given varcode and countrycode, sets the variable equal to varname}
\usage{
get_ecodata_variable_country_wb(varcode, countrycode, varname = NULL)
}
\arguments{
\item{varcode}{String that is the variable code}

\item{countrycode}{String that is the country code}

\item{varname}{Optional, string for the variable name}
}
\value{
Data frame time series with two variables, Date and the variable requested. The data frame will also include all relevant meta data describing the data and citing its source.
}
\description{
`get_ecodata_variable_wb()`
Downloads data from World Bank for a given varcode and countrycode, sets the variable equal to varname
}
