% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fram.R
\name{get_fram}
\alias{get_fram}
\title{get_fram grabs processed DART data off of Github, and joins in the OBJECTID}
\usage{
get_fram()
}
\value{
null, data file written to inst/ folder
}
\description{
get_fram grabs processed DART data off of Github, and joins in the OBJECTID
}
\examples{
\dontrun{
get_fram()
}
}
