% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecodata.R
\name{get_frequency_types}
\alias{get_frequency_types}
\title{Function to return standard frequency types}
\usage{
get_frequency_types()
}
\value{
Returns as an ordered vector, c("Annual", "Semiannual", "Quarterly", "Monthly", "Biweekly", "Weekly", "Daily")
}
\description{
Function to return standard frequency types
}
