% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lighthouse.R
\name{get_lighthouse}
\alias{get_lighthouse}
\title{get_lighthouse is downloads and processes data from the BC lighthouse archive}
\usage{
get_lighthouse(years = NULL)
}
\arguments{
\item{years}{A single year, or vector of years -- if not included, all data is pulled}
}
\value{
null, data file written to inst/ folder
}
\description{
get_lighthouse is downloads and processes data from the BC lighthouse archive
}
\examples{
\dontrun{
get_lighthouse()
}
}
