% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecodata.R
\name{get_recessions}
\alias{get_recessions}
\title{`get_recessions(data)`
Get NBER recession dates that line up with data in the data frame}
\usage{
get_recessions(data)
}
\arguments{
\item{data}{Data frame with a time series}
}
\value{
A copy of the data frame, `data`, with a new variable `Recession` equal to TRUE or FALSE if there was a recession in the time period
}
\description{
`get_recessions(data)`
Get NBER recession dates that line up with data in the data frame
}
