% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rreas_indices.R
\name{get_rreas_indices}
\alias{get_rreas_indices}
\title{get_rreas_indices grabs processed indices off of Github, and joins in the OBJECTID}
\usage{
get_rreas_indices()
}
\value{
null, data file written to inst/ folder
}
\description{
get_rreas_indices grabs processed indices off of Github, and joins in the OBJECTID
}
\examples{
\dontrun{
get_rreas_indices()
}
}
