% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecodata.R
\name{get_state_fips}
\alias{get_state_fips}
\title{Get the FIPS code for a U.S. state}
\usage{
get_state_fips(state)
}
\arguments{
\item{state}{String that is either the full name of the state or the two-letter abbreviation}
}
\value{
String that is the FIPS code for the state
}
\description{
Get the FIPS code for a given U.S. state
}
