% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecodata.R
\name{get_state_fips_all}
\alias{get_state_fips_all}
\title{`get_state_fips_all()`
Get a data frame with all the U.S. states and their FIPS codes. Data comes from `usmaps::fips_info()`}
\usage{
get_state_fips_all()
}
\value{
A data frame with all the U.S. states and their FIPS codes
}
\description{
`get_state_fips_all()`
Get a data frame with all the U.S. states and their FIPS codes. Data comes from `usmaps::fips_info()`
}
\seealso{
[usmaps::fips_info()]
}
