% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecodata.R
\name{get_state_name}
\alias{get_state_name}
\title{`get_state_name(state)`
Get the state name for the given abbreviation or FIPS code}
\usage{
get_state_name(state)
}
\arguments{
\item{state}{String that is a FIPS code or two-letter abbreviation}
}
\value{
String that is the name of the state
}
\description{
`get_state_name(state)`
Get the state name for the given abbreviation or FIPS code
}
