% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecodata.R
\name{get_varcode_url}
\alias{get_varcode_url}
\title{Get a variable code from a FRED or World Bank Data URL}
\usage{
get_varcode_url(url)
}
\arguments{
\item{url}{String that is a URL to a FRED or World Bank data series}
}
\value{
String that is the variable code
}
\description{
Get a variable code from a FRED or World Bank Data URL
}
