% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecodata.R
\name{get_varcode_url_fred}
\alias{get_varcode_url_fred}
\title{`get_varcode_url_fred()`
Extract a variable code from a FRED URL}
\usage{
get_varcode_url_fred(url)
}
\arguments{
\item{url}{String that is a URL to a FRED data series}
}
\value{
String that is the variable code
}
\description{
`get_varcode_url_fred()`
Extract a variable code from a FRED URL
}
