% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecodata.R
\name{get_wb_variable_code}
\alias{get_wb_variable_code}
\title{Extract variable code from World Bank Data URL}
\usage{
get_wb_variable_code(text)
}
\arguments{
\item{text}{String that is the URL or partial URL from World Bank Data for a given variable}
}
\description{
For a given string from a World Bank Data URL or partial URL, extract the code for the variable, but not the location
}
