% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecodata.R
\name{ggplot_add.GeomRecessionLayer}
\alias{ggplot_add.GeomRecessionLayer}
\title{Add a geom recession layer}
\usage{
ggplot_add.GeomRecessionLayer(object, plot, object_name)
}
\description{
Define a ggplot_add method to automatically move the recession layer to the back
}
