% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecodata.R
\name{ggplot_ecodata_facet}
\alias{ggplot_ecodata_facet}
\title{ggplot_ecodata_facet}
\usage{
ggplot_ecodata_facet(
  data,
  variables = NULL,
  title = "",
  ylab = NULL,
  ncol = 4,
  scales = "free",
  color = "dodgerblue4",
  title_strlen = 60,
  strip_width = 40,
  plot.recessions = FALSE
)
}
\arguments{
\item{data}{Data frame from `get_ecodata()` that includes a variable called `Date` and the other variables to plot}

\item{variables}{Optional, vector of strings that includes the economic variables to plot. If not specified, the function will plot all the variables given in `data`, if possible.}

\item{title}{Optional, string for the title of the plot. Default is ""}

\item{ylab}{Optional, string for the y-axis label. Default is the units of the meta data for the variables to be plotted}

\item{ncol}{Optional, number of columns for the facet plot. Default is 4.}

\item{scales}{Optional, string passed to `facet_wrap()` on whether to fix or free the scales on the x and y axes}

\item{color}{Optional, color of the lines. Default is "dodgerblue4"}

\item{title_strlen}{Optional, word-wrap the length of the title by this many characters. Default = 60.}

\item{strip_width}{Optional, word-wrap variable in the title of the individual facet. Default = 40.}

\item{plot.recessions}{Optional, logical for whether or not show show NBER recession bars in the plot}
}
\value{
Returns a ggplot, faceted by each economic variable
}
\description{
Make a time series line plot of the variables in the given data frame, with each variable in its own facet.
}
