% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecodata.R
\name{ggplot_ecodata_ts}
\alias{ggplot_ecodata_ts}
\title{Create time series plot for ecodata variables}
\usage{
ggplot_ecodata_ts(
  data,
  variables = NULL,
  title = "",
  ylab = NULL,
  title_strlen = 60,
  variable_strlen = 85,
  plot.recessions = FALSE
)
}
\arguments{
\item{data}{Data frame from `get_ecodata()` that includes a variable called `Date` and the other variables to plot}

\item{variables}{Optional, vector of strings that includes the economic variables to plot. If not specified, the function will plot all the variables given in `data`, if possible.}

\item{title}{Optional, string for the title of the plot. Default is ""}

\item{ylab}{Optional, string for the y-axis label. Default is the units of the meta data for the variables to be plotted}

\item{title_strlen}{Optional, word-wrap the length of the title by this many characters. Default = 60.}

\item{variable_strlen}{Optional, word-wrap the length of the variable names by this many characters. Default = 85.}

\item{plot.recessions}{Optional, logical for whether or not show show NBER recession bars in the plot}
}
\description{
Make a time series line plot of the variables in the given data frame, in a single plot area.
The number of variables in the data frame should be between 1 and 7
}
