% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecodata.R
\name{glimpse}
\alias{glimpse}
\title{Get a glimpse of ecodata}
\usage{
glimpse(x)
}
\arguments{
\item{x}{An ecodata data frame, the return value from `get_ecodata()`}
}
\description{
A override of dplyr's glimpse() that also prints the label for the dataframe
Just prints information about the dataframe
}
\seealso{
\code{\link[dplyr]{glimpse}}
}
