% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gpa}
\alias{gpa}
\title{GPA data}
\format{
An object of class \code{data.frame} with 1200 rows and 10 columns.
}
\source{
https://m-clark.github.io/mixed-models-with-R/
}
\usage{
gpa
}
\description{
GPA data
}
\details{
From https://m-clark.github.io/mixed-models-with-R/: "The GPA data are a longitudinal data set, where 200 college students have been followed 6 consecutive semesters. In this data set, there are GPA measures on 6 consecutive occasions, with a JOB status variable (how many hours worked) for the same 6 occasions. There are two student-level explanatory variables: the sex (1= male, 2= female) and the high school GPA. There is also a dichotomous student-level outcome variable, which indicates whether a student has been admitted to the university of their choice. Since not every student applies to a university, this variable has many missing values."
}
\author{
Florian Hartig
}
\keyword{datasets}
