% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gray_seal.R
\docType{data}
\name{grayseal}
\alias{grayseal}
\title{Gray Seal Bycatch}
\format{
24 rows and 6 columns.

\itemize{
\item totalest5y: Estimated bycatch with five year rolling window.
\item totalest5yUCI: Upper confidence interval of estimated bycatch with five year rolling window.
\item totalest5yLCI: Lower confidence interval of estimated bycatch with five year rolling window.
\item totalest1y: Annual estimated bycatch.
\item pbr: Possible Biological Removals.
\item year: Time step.
}
}
\usage{
grayseal
}
\description{
Data include estimated Gray Seal bycatch on the Northeast Shelf.
}
\details{
Methods used to calculate data are available at \url{https://noaa-edab.github.io/tech-doc/harbor-porpoise-and-gray-seal-bycatch.html}.
}
\keyword{datasets}
