% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitat_diversity.R
\docType{data}
\name{habitat_diversity}
\alias{habitat_diversity}
\title{Habitat Diversity}
\format{
Time series with 400 rows and 4 columns.

\itemize{
\item Time: Year
\item Var: Diversity metrics
\item Value: Value of Variable
\item EPU: Representative Ecological Production Unit (MAB = Mid-Atlantic Bight, GB = Georges Bank,
GOM = Gulf of Maine, SS = Scotian Shelf, or All).

}
}
\usage{
habitat_diversity
}
\description{
Habitat diversity from the Northeast Regional Habitat Assessment
}
\details{
Methods used to calculate data are available at \url{https://noaa-edab.github.io/tech-doc}.
}
\keyword{datasets}
