% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitat_vulnerability.R
\docType{data}
\name{habitat_vulnerability}
\alias{habitat_vulnerability}
\title{Habitat Vulnerability}
\format{
62 rows and 9 columns.

\itemize{
\item Species: Species highly dependent on highly vulnerable habitats.
\item Habitat Name: Name of highly vulnerable habitat.
\item Habitat Vulnerability Rank (HVCA): Habitat vulnerability score.
\item Species Vulnerability Rank (FVCA): Species vulnerability score.
\item Importance of habitat by life stage (ACFHP): Life stage vulnerability to highly vulnerable habitat.

}
}
\usage{
habitat_vulnerability
}
\description{
Table includes summary of habitat vulnerability and the importance of such habitats to managed species.
}
\details{
Methods used to calculate data are available at \url{https://noaa-edab.github.io/tech-doc/habitat-vulnerability.html}.
}
\keyword{datasets}
