% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habs.R
\docType{data}
\name{habs}
\alias{habs}
\title{Harmful Algal Blooms - Alexandrium and Paralitic Shellfish Poisoning}
\format{
\itemize{
\item Var: Regions of interest.
\item Source: Harmful bloom source (PSP or Alexandrium).
\item Value: Value of variable \code{Var}.
\item EPU: Representative Ecological Production Unit (MAB = Mid-Atlantic Bight, GB = Georges Bank,
GOM = Gulf of Maine, SS = Scotian Shelf, or All).
\item Time: Years.
}
}
\usage{
habs
}
\description{
Annual time series of the Paralytic Shellfish Poisoning (PSP) and Alexandrium toxins in the Gulf of Maine.
}
\details{
\url{https://noaa-edab.github.io/tech-doc/habs_alexandrium.html}.
\url{https://noaa-edab.github.io/tech-doc/habs_psp.html}.
}
\keyword{datasets}
