% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatwave.R
\docType{data}
\name{heatwave}
\alias{heatwave}
\title{Marine Heatwaves}
\format{
648 rows and 5 columns.

\itemize{
\item EPU: Ecological Production Unit (EPU) where sampling occurred.
\item Time: Year.
\item Var: Cumulative or maximum marine heatwave intensity.
\item Value: Value of variable \code{Var}.
\item Units: Units of variable \code{Var}.
}
}
\usage{
heatwave
}
\description{
Data include cumulative intensity, maximum intensity and duration of
surface and bottom water marine heatwaves on the NES.
}
\details{
More information including processing and indicator derivation steps are available at
\url{https://noaa-edab.github.io/tech-doc/marine-heatwave.html}.
}
\keyword{datasets}
