% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatwave_peak_date.R
\docType{data}
\name{heatwave_peak_date}
\alias{heatwave_peak_date}
\title{Maximum heatwave peak day gridded}
\format{
Data set contains 6396 rows and 4 columns

\itemize{
\item Latitude: Latitude.
\item Longitude: Longitude
\item Value: Sea surface temperature Value.
\item EPU: Ecological Production Unit (EPU).
}
}
\usage{
heatwave_peak_date
}
\description{
Data include spatial sea surface temperature of peak marine heatwave day.
}
\details{
More information including processing and indicator derivation steps are available at
\url{https://noaa-edab.github.io/tech-doc/marine_heatwave.html}.
}
\keyword{datasets}
