% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hms_category.R
\docType{data}
\name{hms_category}
\alias{hms_category}
\title{Highly Migratory Species Category}
\format{
30 rows and 4 columns.

\itemize{
\item SP_CODE: Numeric species specific code.
\item COMMON_NAME: Common species name.
\item SP_CATEGORY: Grouping that the species falls into.
\item COMMON_POP: Common species name from Southwest POP.
}
}
\usage{
hms_category
}
\description{
Data include landing weight and revenue from group HMS species.
}
\details{
More information available at
\url{https://noaa-edab.github.io/tech-doc/highly-migratory-species-landings.html}.
}
\keyword{datasets}
