% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hms_cpue.R
\docType{data}
\name{hms_cpue}
\alias{hms_cpue}
\title{Highly Migratory Species Catch per Unit Effort}
\format{
3600 rows and 4 columns.

\itemize{
\item EPU: Ecological Production Unit (EPU) where sampling occurred.
\item Time: Year.
\item Var: Species.
\item Value: Value of variable \code{Var}.
}
}
\usage{
hms_cpue
}
\description{
Data include catch per unit effort for all HMS species.
}
\details{
More information available at
\url{https://noaa-edab.github.io/tech-doc/highly-migratory-species-landings.html}.
}
\keyword{datasets}
