% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hms_landings.R
\docType{data}
\name{hms_landings}
\alias{hms_landings}
\title{Highly Migratory Species Landings and Revenue}
\format{
132 rows and 5 columns.

\itemize{
\item EPU: Ecological Production Unit (EPU) where sampling occurred.
\item Time: Year.
\item Var: Species weight or revenue landings.
\item Value: Value of variable \code{Var}.
\item Units: Units of variable \code{Var}.
}
}
\usage{
hms_landings
}
\description{
Data include landing weight and revenue from group HMS species.
}
\details{
More information including processing and indicator derivation steps are available at
\url{https://noaa-edab.github.io/tech-doc/highly-migratory-species-landings.html}.
}
\keyword{datasets}
