% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hms_stock_status.R
\docType{data}
\name{hms_stock_status}
\alias{hms_stock_status}
\title{Highly Migratory species stock status}
\format{
40 rows and 4 columns

\itemize{
\item stock: Stock identifier.
\item Var: Either \code{F.Fmsy}, representing the ratio of fishery mortality to fishery mortality
at maximum sustainable yield (or proxy), or \code{B.Bmsy}, defined as the ratio of estimated stock biomass to
estimated biomass at maximum sustainable yield (or proxy).
\item Value: Value of variable \code{Var}.
\item Year: Year.
}
}
\usage{
hms_stock_status
}
\description{
Stock status for highly migratory species in North Atlantic.
}
\details{
More information about this data set may be found at \url{https://noaa-edab.github.io/tech-doc/atlantic-highly-migratory-species-stock-status.html}.
}
\keyword{datasets}
