% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hp_density.R
\docType{data}
\name{hp_density}
\alias{hp_density}
\title{hp_density}
\format{
A data set containing 6244 rows and 5 columns.

\itemize{
\item Latitude: Latitude
\item Longitude: Longitude
\item Value: Value
\item Bin: ??
\item Season: ???

}
}
\usage{
hp_density
}
\description{
What is it?
}
\details{
More information available at \url{link to techdoc}
}
\keyword{datasets}
