% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hudson_river_flow.R
\docType{data}
\name{hudson_river_flow}
\alias{hudson_river_flow}
\title{Hudson River Flow}
\format{
40 rows and 4 columns

\itemize{
\item stock: Stock identifier.
\item Var: Either \code{F.Fmsy}, representing the ratio of fishery mortality to fishery mortality
at maximum sustainable yield (or proxy), or \code{B.Bmsy}, defined as the ratio of estimated stock biomass to
estimated biomass at maximum sustainable yield (or proxy).
\item Value: Value of variable \code{Var}.
\item Year: Year.
}
}
\usage{
hudson_river_flow
}
\description{
Stock status for highly migratory species in North Atlantic.
}
\details{
More information about this data set may be found at \url{https://noaa-edab.github.io/tech-doc/hudson-river-flow.html}.
}
\keyword{datasets}
