% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hurricanes}
\alias{hurricanes}
\title{Hurricane dataset}
\format{
An object of class \code{data.frame} with 92 rows and 14 columns.
}
\source{
\url{https://www.pnas.org/doi/10.1073/pnas.1402786111}
}
\usage{
hurricanes
}
\description{
Hurricane dataset
}
\references{
Jung, K., Shavitt, S., Viswanathan, M., & Hilbe, J. M. (2014). Female hurricanes are deadlier than male hurricanes. Proceedings of the National Academy of Sciences, 111(24), 8782-8787.
}
\author{
Florian Hartig
}
\keyword{datasets}
