% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ichthyo_diversity.R
\docType{data}
\name{ichthyo_diversity}
\alias{ichthyo_diversity}
\title{Ichthyoplankton diversity}
\format{
69 rows and 5 columns

\itemize{
\item Var: Specifies variable of interest and sampling season.
\item Time: Sampling year.
\item Units: Units of variable \code{Var}.
\item EPU: Ecological Production Units (EPU) where sampling occurred. \code{All} refers to data aggregated to the level of the Northeast Continental Shelf.
\item Value: Value of variable \code{Var}.
}
}
\usage{
ichthyo_diversity
}
\description{
Data from NOAA NEFSC Oceans and Climate branch EcoMon public dataset were used to examine changes in
diversity of abundance among 45 ichthyoplankton taxa. The 45 taxa were established in Walsh et al. 2015,
and include the most abundant taxa from the 1970s to present that represent consistency in the
identification of larvae. Both species counts and estimates of Shannon-Wiener diversity are included
in this data set.
}
\details{
\code{NaN} values were included in the data where there were too few sampling stations to be included in diversity calculations.
More information about these data are available at \url{https://noaa-edab.github.io/tech-doc/ichthyoplankton-diversity.html}.
}
\references{
Walsh, H. J., Richardson, D. E., Marancik, K. E., & Hare, J. A. (2015). Long-term changes in the distributions of larval and adult fish in the northeast US shelf ecosystem. \emph{PloS one}, 10(9), e0137382.
}
\keyword{datasets}
