% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inshore_survdat.R
\docType{data}
\name{inshore_survdat}
\alias{inshore_survdat}
\title{New England inshore state survey data}
\format{A data set containing 710 rows and 5 columns.

\itemize{
\item Var: Includes SOE species grouping, season that survey occurred (fall or spring), and variable calculated.
The specific variables included are stratified mean biomass with associated confidence intervals, coefficients of
variation, and standard errors.
\item Time: Year that survey occurred.
\item Value: Value of variable \code{Var}.
\item Units: Units of variable \code{Var}.
\item EPU: Ecological Production Unit (EPU) where sampling occurred. Here \code{NE} refers to New England.
}}
\usage{
inshore_survdat
}
\description{
These data contain time series of inshore (within 12 miles) fishery-independent survey data collected by the Maine Department of Marine Resources
and New Hampshire Fish & Game Department. Time series are grouped by State of the Ecosystem feeding guild. Species assignments
within groupings are available through \code{inshore_survdat_species}.
}
\details{
Metrics for biomass were calculated by:
\enumerate{
\item All species catch weights were summed up for each tow for each group.
\item Then the average weight per tow and associated variances and standard deviation for each
survey, region, stratum, and species group was calculated.
\item The average weight for each group was then weighted by total area surveyed in each region
and stratum. (average weight * area surveyed).
\item The new weighted averages were then summed up by survey and species group then the totals
were divided by the total area of the survey (11699.831 km2) to provide the stratified mean biomass
for each species group in each survey. The coefficient of variation, standard error, and 95%
confidence intervals were also calculated for each species group and provided.
}

Read more about inshore surveys at \url{https://wildlife.state.nh.us/marine/research.html} and
\url{https://www.maine.gov/dmr/science-research/projects/trawlsurvey/index.html}
}
\keyword{datasets}
