% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecodata.R
\name{is_empty}
\alias{is_empty}
\title{Test if a string is empty}
\usage{
is_empty(x)
}
\arguments{
\item{x}{A character string}
}
\value{
Logical, whether or not the string is empty
}
\description{
Test if a string is empty
}
