% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lepanthes_rupestris}
\alias{lepanthes_rupestris}
\title{Lepanthes rupestris}
\format{
A data frame with 975 rows and 23 variables:
\describe{
  \item{UID}{}
  \item{x.coord}{}
}
}
\source{
Acevedo, Miguel; Beaudrot, Lydia; Melendez-Ackerman, Elvia; Tremblay, Raymond (2020), Local extinction risk under climate change in a neotropical asymmetrically dispersed epiphyte, Dryad, Dataset, https://doi.org/10.5061/dryad.9p8cz8wc6

\url{https://doi.org/10.5061/dryad.9p8cz8wc6}
}
\usage{
lepanthes_rupestris
}
\description{
Description goes here
}
\keyword{datasets}
