% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_line_survey.R
\docType{data}
\name{long_line_survey}
\alias{long_line_survey}
\title{Long-line Survey}
\format{
Data contain 96 rows and 8 columns.
\itemize{
\item CRUISE: Cruise Identifyer.
\item STATION: Station Identifyer.
\item YEAR: Year.
\item SEASON: Season.
\item COMMON NAME: Species common name.
\item DECDEG_BEGLAT_SET: Latitudal position.
\item DECDEG_BEGLON_SET: Longitudal position.
\item categ: Grouping category for plotting.
}
}
\usage{
long_line_survey
}
\description{
Gulf of Maine long line survey data
}
\keyword{datasets}
