% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longlinesurvey_sf.R
\docType{data}
\name{longlinesurvey_sf}
\alias{longlinesurvey_sf}
\title{Longline survey sf index ?????}
\format{
Simple feature collection with 6 features and 7 fields

\itemize{
\item geometry: yearly spatial feature for the longline survey.

}
}
\usage{
longlinesurvey_sf
}
\description{
Longline survey sf index ?????
}
\details{
More information available at \url{??}
}
\keyword{datasets}
