% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lps_sharks.R
\docType{data}
\name{lps_sharks}
\alias{lps_sharks}
\title{Large Pelagics Survey -  Shark Data}
\format{
90 rows and 5 columns

\itemize{
\item Time: Years.
\item EPU: Ecological Production Unit where data originated.
\item Var: Either \code{F.Fmsy}, representing the ratio of fishery mortality to fishery mortality
at maximum sustainable yield (or proxy), or \code{B.Bmsy}, defined as the ratio of estimated stock biomass to
estimated biomass at maximum sustainable yield (or proxy).
\item Value: Value of variable \code{Var}.
\item Units: Units of variable \code{Var}.
}
}
\usage{
lps_sharks
}
\description{
Large Pelagics Survey (LPS) shark abundance count.
}
\keyword{datasets}
