% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{marmots}
\alias{marmots}
\title{Marmots}
\format{
A 'data.frame': 896 obs. of 55 variables.
}
\usage{
marmots
}
\description{
Unfortunately no description available.
}
\author{
Florian Hartig
}
\keyword{datasets}
