% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_inshore_survey.R
\docType{data}
\name{mass_inshore_survey}
\alias{mass_inshore_survey}
\title{Massachusetts inshore bottom trawl survey}
\format{
413 rows and 5 columns

\itemize{
\item Var: Specifies variable type, including SOE species groupings and sampling season with stratified mean biomass per tow
\item Value: Value of variable \code{Var}.
\item Time: Year.
\item Units: Units of variable \code{Var}.
\item EPU: Ecological Production Unit. These data reflect inshore sampling in MA state waters.
}
}
\usage{
mass_inshore_survey
}
\description{
Data include aggregated time series of inshore fishery-independent trawl survey data from Massachusetts waters.
MA inshore surveys have been performed biannually in the spring and fall since 1978.
}
\details{
Technical details related to the MA inshore survey are available at \url{https://www.mass.gov/files/documents/2016/08/tm/tr-38.pdf}.
}
\keyword{datasets}
