% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecodata.R
\name{maximum_frequency}
\alias{maximum_frequency}
\title{Compute maximum frequency}
\usage{
maximum_frequency(vec)
}
\arguments{
\item{vec}{Vector of frequencies. All values must be in c("Annual", "Semiannual", "Quarterly", "Monthly", "Biweekly", "Weekly", "Daily")}
}
\value{
Maximum value of the frequencies
}
\description{
Compute maximum frequency of the values within a vector
}
