% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{melanoma}
\alias{melanoma}
\title{Melanoma}
\format{
A 'data.frame':	205 obs. of  7 variables:
\describe{
   \item{DC}death/censoring indicator; {The patients status at the end of the study. 1 indicates that they had died from melanoma, 2 indicates that they were still alive and 3 indicates that they had died from causes unrelated to their melanoma}
   \item{time}{Survival time in days since the operation, possibly censored}
   \item{level}level of invasion, 0, 1 or 2
   \item{ici}inflammatory cell infiltration (ICI), 0, 1, 2 or 3
   \item{ecel}presence of epithelioid cells, 1=no, 2=yes
   \item{ulcer}{Indicator of ulceration; 1=present, 0=absent.}
   \item{thickness}{Tumour thickness in mm}
   \item{sex}{The patients sex; 1=male, 0=female.}
   \item{age}{Age in years at the time pf the operation.}
}
}
\source{
\url{http://publicifsv.sund.ku.dk/~linearpredictors/?page=datasets&dataset=Melanoma}
}
\usage{
melanoma
}
\description{
Survival from Malignant Melanoma
}
\author{
Florian Hartig
}
\keyword{datasets}
