% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecodata.R
\name{minimum_frequency}
\alias{minimum_frequency}
\title{Compute minimum frequency}
\usage{
minimum_frequency(vec)
}
\arguments{
\item{vec}{Vector of frequencies. All values must be in c("Annual", "Semiannual", "Quarterly", "Monthly", "Biweekly", "Weekly", "Daily")}
}
\value{
Minimum value of the frequencies
}
\description{
Compute minimum frequency of the values within a vector
}
