% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecodata.R
\name{mutate.ecodata}
\alias{mutate.ecodata}
\title{Create and Modify Variables}
\usage{
mutate.ecodata(.data, ...)
}
\arguments{
\item{.data}{An ecodata data frame}

\item{...}{Parameters passed on to mutate}
}
\value{
An ecodata data frame with the new variables and metadata attributes
}
\description{
Wrapper to the `dplyr::mutate()` function that adds metadata attributes to the new variables.
}
\seealso{
`dplyr::mutate()`
}
