% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nao.R
\docType{data}
\name{nao}
\alias{nao}
\title{North Atlantic Oscillation}
\format{Data include 827 rows and 6 columns

\itemize{
\item Var: Variable name.
\item Value: Value of variable name \code{Var}.
\item Year: Year.
\item Month: Month.
\item Units: Units of variable \code{Var}.
\item EPU: Ecological Production Unit (EPU).

}}
\usage{
nao
}
\description{
North Atlantic Oscillation data were taken from the NOAA NWS Climate Prediction Center.
These data show the monthly NAO index time series beginning in 1950 and ending in November 2018.
}
\details{
The index is standardized by the standard deviation of the 1950-2000 reference period. More
information regarding the methodology involved in deriving the NAO and its significance is
available at http://www.cpc.ncep.noaa.gov/data/teledoc/nao.shtml.
}
\keyword{datasets}
