% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/narw.R
\docType{data}
\name{narw}
\alias{narw}
\title{North Atlantic right whale abundance}
\format{These data include 84 rows and 5 columns.

\itemize{
\item Var: Specifies variable of interest, including NARW abundance estimate median and 95% credible
intervals.
\item Value: Value of variable \code{Var}.
\item Units: Units of variable \code{Var}.
\item Time: Year.
\item EPU: Ecological Production Unit (EPU).
}}
\usage{
narw
}
\description{
Data include abundance estimates for the North Atlantic right whale (NARW) provided by the Protected Species
Branch of the NEFSC. Abundance was estimated following the methodology of Pace et. al 2017.
}
\details{
More information about these data may be found \url{https://noaa-github.io/tech-doc/right-whale-abundance.html}.
}
\references{
Pace, R. M., Corkeron, P. J., Krauss, S. D. (2017) State-space mark-recapture estimates reveal a recent decline in abundance of North Atlantic right whales.
\emph{Ecology and Evolution}, 1-17.
}
\keyword{datasets}
