% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nefsc_survey.R
\docType{data}
\name{nefsc_survey}
\alias{nefsc_survey}
\title{NEFSC Bottom Trawl Survey (aggregated)}
\format{\itemize{
\item Var: Specifies feeding guild and season in which sampling occurred.
\item Units: Units of variable \code{Var}.
\item Value: Value of variable \code{Var}.
\item EPU: Ecological Production Unit where sampling occurred (MAB = Mid-Atlantic Bight, GB = Georges Bank,
GOM = Gulf of Maine, SS = Scotian Shelf).
\item Time: Years.
}}
\usage{
nefsc_survey
}
\description{
The NEFSC Bottom Trawl Survey aggregated to the level of feeding guild, which is a simplified
grouping structure developed from those proposed by Link et al. (2006) and Garrison and Link (2000). The
feeding guild structure includes Apex Predator, Piscivore, Planktivore, Benthivore, Benthos, and Other.
Data is presented for both fall and spring bottom trawls surveys.
}
\details{
More information regarding the NEFSC Bottom Trawl Survey is available at \url{https://noaa-edab.github.io/tech-doc/survdat.html},
and further details about feeding guild structure are available at \url{https://noaa-edab.github.io/tech-doc/aggroups.html}.
}
\references{
Garrison, Lance P, and Jason S Link. 2000. “Dietary guild structure of the fish community in the Northeast United States continental shelf ecosystem.”
\emph{Marine Ecology Progress Series} 202:231–40.

Link, Jason S, Carolyn A Griswold, Elizabeth T Methratta, and Jessie Gunnard. 2006. Documentation for the energy modeling and analysis exercise (EMAX).
US Department of Commerce, National Oceanic and Atmospheric Administration.
}
\keyword{datasets}
