% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_england.R
\docType{data}
\name{new_england}
\alias{new_england}
\title{New England Spatial data}
\format{
Data set inclues 6 rows and 5 columns

\itemize{
\item STATE_NAME: Name of state in New England.
\item SUB_REGION: Ecological Production Unit (EPU).
\item STATE_ABBR: State Abbreviation.
\item SQMI: Total area in square miles.
\item geometry: polygon of state.
}
}
\usage{
new_england
}
\description{
New England Spatial data
}
\keyword{datasets}
