% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observed_sharks.R
\docType{data}
\name{observed_sharks}
\alias{observed_sharks}
\title{NE Observer program Shark data}
\format{
These data contain 232 rows and 4 columns.

\itemize{
\item Var: Shark category.
\item Value: Value of variable \code{Var}.
\item Time: Time step of \code{Var}.
\item EPU: Ecological Production Unit (EPU) where sampling occurred.
}
}
\usage{
observed_sharks
}
\description{
NE Observer program Shark data
}
\details{
Methods used to calculate data are available at \url{https://noaa-edab.github.io/tech-doc}.
}
\keyword{datasets}
