% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ocean_grid}
\alias{ocean_grid}
\title{Gridded ocean data to match for salmon dashboard project}
\format{
A data frame.
}
\usage{
ocean_grid
}
\description{
Gridded ocean data to match for salmon dashboard project
}
\keyword{datasets}
