% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_processing.R
\name{open_processing}
\alias{open_processing}
\title{Open the Rmarkdown document used for processing State of the Ecosystem indicator data.}
\usage{
open_processing(region)
}
\description{
Use this function to access the processing file for turning raw data into derived SOE indicators.
}
\examples{

open_processing()

}
