% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_soe.R
\name{open_soe}
\alias{open_soe}
\title{Open Mid-Atlantic or New England State of the Ecosystem reports}
\usage{
open_soe(region)
}
\arguments{
\item{region}{Enter "mab" or "ne".}
}
\description{
Use this function to open SOE Rmarkdown documents into your RStudio environment.
}
\examples{

open_soe(region = "mab")

}
