% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyto_size.R
\docType{data}
\name{phyto_size}
\alias{phyto_size}
\title{Phytoplankton size class}
\format{
These data contain 28197 rows and 5 columns.

\itemize{
\item Var: Phytoplankton size category.
\item Value: Value of variable \code{Var}.
\item Units: Units of variable \code{Var}.
\item Time: Time step of \code{Var}. Time format is given as YYYYMM for monthly time steps and YYYYWW for weekly data.
\item EPU: Ecological Production Unit (EPU) where sampling occurred.
}
}
\usage{
phyto_size
}
\description{
Phytoplankton size class
}
\details{
Methods used to calculate data are available at \url{https://noaa-edab.github.io/tech-doc/phytoplankton-size-class.html}.
}
\keyword{datasets}
