% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plantHeight}
\alias{plantHeight}
\title{Plant height}
\format{
An object of class \code{data.frame} with 178 rows and 35 columns.
}
\source{
https://environmentalcomputing.net/statistics/linear-models/linear-regression/
}
\usage{
plantHeight
}
\description{
A trait dataset with plant height, location and environmental covariates
}
\author{
Florian Hartig
}
\keyword{datasets}
