% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_abc.acl.R
\name{plot_abc.acl}
\alias{plot_abc.acl}
\title{plot abc.acl}
\usage{
plot_abc.acl(
  shadedRegion = shadedRegion,
  report = "MidAtlantic",
  plottype = "Stacked"
)
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10)}

\item{report}{Character string. Which SOE report ("MidAtlantic", "NewEngland")}

\item{plottype}{Character string. Which plot ("Stacked", "Catch")}
}
\value{
ggplot object
}
\description{
Plots either stacked bar of quotas/ABCs by fishery ("Stacked" plottype option)
or point comparison of catch with quota/ABC by fishery ("Catch" plottype option)
}
