% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_abc_acl.R
\name{plot_abc_acl}
\alias{plot_abc_acl}
\title{plot abc_acl}
\usage{
plot_abc_acl(shadedRegion = NULL, report = "MidAtlantic", plottype = "Stacked")
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10)}

\item{report}{Character string. Which SOE report ("MidAtlantic", "NewEngland")}

\item{plottype}{Character string. Which plot ("Stacked", "Catch")}
}
\value{
ggplot object
}
\description{
Plots either stacked bar of quotas/ABCs by fishery ("Stacked" plottype option)
or point comparison of catch with quota/ABC by fishery ("Catch" plottype option)
}
