% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_aggregate_biomass.R
\name{plot_aggregate_biomass}
\alias{plot_aggregate_biomass}
\title{plot aggregate_biomass}
\usage{
plot_aggregate_biomass(
  shadedRegion = NULL,
  report = "MidAtlantic",
  EPU = "MAB",
  n = 0
)
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10)}

\item{report}{Character string. Which SOE report ("MidAtlantic", "NewEngland")}

\item{EPU}{Character string. Which EPU for New England report ("GB", "GOM") Mid will always be MAB}

\item{n}{Numeric scalar. Number of years used (from most recent year) to estimate short term trend . Default = 0 (No trend calculated)}
}
\value{
ggplot object
}
\description{
Plots faceted spring and fall survey biomass time series by aggregate group
}
