% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_annual_chl_pp.R
\name{plot_annual_chl_pp}
\alias{plot_annual_chl_pp}
\title{plot chl_pp}
\usage{
plot_annual_chl_pp(
  shadedRegion = NULL,
  report = "MidAtlantic",
  varName = "chl",
  plottype = "mean",
  EPU = "MAB"
)
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10)}

\item{report}{Character string. Which SOE report ("MidAtlantic", "NewEngland")}

\item{varName}{Character string. Which Variable to plot ("chl","pp")}

\item{plottype}{Character string. Which plot ("mean","total")}
}
\value{
ggplot object
}
\description{
Plot time series of chlorophyll a (chl), primary production (pp), weekly, monthly,
or anomalies.
}
