% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_aquaculture.R
\name{plot_aquaculture}
\alias{plot_aquaculture}
\title{plot aquaculture}
\usage{
plot_aquaculture(shadedRegion = shadedRegion, report = "MidAtlantic")
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10)}

\item{report}{Character string. Which SOE report ("MidAtlantic", "NewEngland")}
}
\value{
ggplot object
}
\description{
Creates line plot of oyster production.
}
