% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bennet.R
\name{plot_bennet}
\alias{plot_bennet}
\title{plot Bennet indicator}
\usage{
plot_bennet(shadedRegion = NULL, report = "MidAtlantic")
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10)}

\item{report}{Character string. Which SOE report ("MidAtlantic", "NewEngland")}
}
\value{
ggplot object
}
\description{
Stacked bar charts of price and volume components of revenue by feeding guild.
}
