% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_benthic_index.R
\name{plot_benthos_index}
\alias{plot_benthos_index}
\title{plot benthic index}
\usage{
plot_benthos_index(
  shadedRegion = NULL,
  report = "MidAtlantic",
  varName = "Macrobenthos",
  plottype = "index",
  n = 0
)
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10)}

\item{report}{Character string. Which SOE report ("MidAtlantic", "NewEngland")}

\item{varName}{Character string. Benthic group name ("Macrobenthos", "Megabenthos")}

\item{plottype}{Character string. Benthic biomass index by region, or coastwide center of gravity ("index", "cog")}

\item{n}{Numeric scalar. Number of years used (from most recent year) to estimate short term trend . Default = 0 (No trend calculated)}
}
\value{
ggplot object
}
\description{
Description should be here. This needs to be reworked to uncouple GB and GOM
}
