% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_blue_runner.R
\name{plot_blue_runner}
\alias{plot_blue_runner}
\title{plot blue_runner}
\usage{
plot_blue_runner(shadedRegion = shadedRegion, report = "MidAtlantic")
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10)}

\item{report}{Character string. Which SOE report ("MidAtlantic", "NewEngland")}
}
\value{
ggplot object
}
\description{
Creates map of blue runner locations by decade
}
