% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bottom_temp.R
\name{plot_bottom_temp}
\alias{plot_bottom_temp}
\title{plot bottom anomaly temperature time series using NEFSC survey data}
\usage{
plot_bottom_temp(shadedRegion = c(2012, 2022), report = "MidAtlantic")
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10)}

\item{report}{Character string. Which SOE report ("MidAtlantic", "NewEngland")}
}
\value{
ggplot object
}
\description{
ecodata::bottom_temp
}
