% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bottom_temp_anomaly_glorys.R
\name{plot_bottom_temp_anomaly_glorys}
\alias{plot_bottom_temp_anomaly_glorys}
\title{plot bottom anomaly temperature time series using Glorys data}
\usage{
plot_bottom_temp_anomaly_glorys(
  region = NULL,
  shadedRegion = c(2012, 2022),
  shade.fill = "lightgrey",
  shade.alpha = 0.3,
  hline.size = 1,
  hline.alpha = 0.5,
  hline.lty = "dashed"
)
}
\arguments{
\item{region}{Character vector. Regional EPUs ("GB","MAB") to overly on figure. (Default = NULL, use all)}

\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot}

\item{shade.fill}{Character string. Color of shaded region. (Default = "lightgrey)}

\item{shade.alpha}{Numeric scalar. Alpha of shaded region (Default = 0.5)}

\item{hline.size}{Numeric scalar Line width of horizontal line. (Default = 1)}

\item{hline.alpha}{Numeric scalar. Alpha of horizontal line (Default = 0.5)}

\item{hline.lty}{Character string. Line type of horizontal line (Default = "dashed")}
}
\value{
ggplot object
}
\description{
ecodata::bottom_temp_glorys
}
