% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bottom_temp_anomaly_space.R
\name{plot_bottom_temp_anomaly_space}
\alias{plot_bottom_temp_anomaly_space}
\title{plot botttom temperature anomaly}
\usage{
plot_bottom_temp_anomaly_space(season = NULL, region = NULL)
}
\arguments{
\item{season}{Character string. Season to plot. (Default = NULL, plot all seasons)}

\item{region}{Character vector. Regional EPUs ("GB","MAB") to overly on figure. (Default = NULL, use all)}
}
\value{
ggplot object
}
\description{
plot botttom temperature anomaly
}
