% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bottom_temp_glorys.R
\name{plot_bottom_temp_glorys}
\alias{plot_bottom_temp_glorys}
\title{plot bottom anomaly temperature time series using GLORYS}
\usage{
plot_bottom_temp_glorys(shadedRegion = NULL, report = "MidAtlantic")
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10)}

\item{report}{Character string. Which SOE report ("MidAtlantic", "NewEngland")}
}
\value{
ggplot object
}
\description{
ecodata::bottom_temp_glorys
}
