% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bottom_temp_model_gridded.r
\name{plot_bottom_temp_model_gridded}
\alias{plot_bottom_temp_model_gridded}
\title{plot bottom temperature seasonal gridded data}
\usage{
plot_bottom_temp_model_gridded(
  shadedRegion = NULL,
  report = "MidAtlantic",
  scale = "celsius"
)
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10), passed from plot function}

\item{report}{Character string. Which SOE report ("MidAtlantic", "NewEngland"), passed from plot function}

\item{scale}{character string. celsius or fahrenheit. Default = "celsius"}
}
\value{
ggplot object
}
\description{
plots bottom_temp_model_gridded data set
}
