% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bottom_temp_seasonal_gridded.r
\name{plot_bottom_temp_seasonal_gridded}
\alias{plot_bottom_temp_seasonal_gridded}
\title{plot bottom temperature seasonal gridded data}
\usage{
plot_bottom_temp_seasonal_gridded(shadedRegion = NULL, report = "MidAtlantic")
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10), passed from plot function}

\item{report}{Character string. Which SOE report ("MidAtlantic", "NewEngland"), passed from plot function}
}
\value{
ggplot object
}
\description{
plots bottom_temp_seasonal_gridded data set
}
