% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_calanus_stage.R
\name{plot_calanus_stage}
\alias{plot_calanus_stage}
\title{plot calanus_stage}
\usage{
plot_calanus_stage(shadedRegion = shadedRegion, report = "MidAtlantic")
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10)}

\item{report}{Character string. Which SOE report ("MidAtlantic", "NewEngland")}
}
\value{
ggplot object
}
\description{
Create bar plots of Calanus zooplankton stages over time
}
