% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cetacean_dist.R
\name{plot_cetacean_dist}
\alias{plot_cetacean_dist}
\title{plot HMS species distribution}
\usage{
plot_cetacean_dist(shadedRegion = NULL, report = "MidAtlantic")
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10)}

\item{report}{Character string. Which SOE report ("MidAtlantic", "NewEngland")}
}
\value{
ggplot object
}
\description{
Plots distribution shifts for marine mammals
}
