% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ch_bay_sal.R
\name{plot_ch_bay_sal}
\alias{plot_ch_bay_sal}
\title{plot ch_bay_sal}
\usage{
plot_ch_bay_sal(shadedRegion = shadedRegion, report = "MidAtlantic")
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10)}

\item{report}{Character string. Which SOE report ("MidAtlantic" only, default)}
}
\value{
ggplot object
}
\description{
Plot Chesapeake Bay salinity daily time series.
}
