% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ches_bay_wq.R
\name{plot_ches_bay_wq}
\alias{plot_ches_bay_wq}
\title{plot ches_bay_wq}
\usage{
plot_ches_bay_wq(shadedRegion = NULL, report = "MidAtlantic", n = 0)
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10)}

\item{report}{Character string. Which SOE report ("MidAtlantic" only, default)}

\item{n}{Numeric scalar. Number of years used (from most recent year) to estimate short term trend . Default = 0 (No trend calculated)}
}
\value{
ggplot object
}
\description{
Plot time series of Chesapeake Bay water quality indicator.
}
