% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_chl_pp.R
\name{plot_chl_pp}
\alias{plot_chl_pp}
\title{plot chl_pp}
\usage{
plot_chl_pp(
  shadedRegion = NULL,
  report = "MidAtlantic",
  varName = "chl",
  plottype = "weekly",
  year = NULL
)
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10)}

\item{report}{Character string. Which SOE report ("MidAtlantic", "NewEngland")}

\item{varName}{Character string. Which Variable to plot ("chl","pp")}

\item{plottype}{Character string. Which plot ("weekly", "monthly", "anomaly")
Weekly and monthly plots are for both variables, annual anomaly plot for PP only.}

\item{year}{Numeric value. Optional. Year for weekly plot, defaults to max year in data.}
}
\value{
ggplot object
}
\description{
Plot time series of chlorophyll a (chl), primary production (pp), weekly, monthly,
or anomalies.
}
