% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cold_pool.R
\name{plot_cold_pool}
\alias{plot_cold_pool}
\title{plot cold_pool}
\usage{
plot_cold_pool(
  shadedRegion = NULL,
  report = "MidAtlantic",
  varName = "cold_pool",
  n = 0
)
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10)}

\item{report}{Character string. Which SOE report ("MidAtlantic" only, default)}

\item{varName}{Character string. Which variable to plot ("cold_pool", "persistence","extent"). Default = "cold_pool"}

\item{n}{Numeric scalar. Number of years used (from most recent year) to estimate short term trend . Default = 0 (No trend calculated)}
}
\value{
ggplot object
}
\description{
Create 3 panel plot of cold pool temperature, extent, and persistence time series.
}
