% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cold_pool_sf.R
\name{plot_cold_pool_sf}
\alias{plot_cold_pool_sf}
\title{plot cold_pool_sf}
\usage{
plot_cold_pool_sf(shadedRegion = NULL, report = "MidAtlantic")
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10)}

\item{report}{Character string. Which SOE report ("MidAtlantic" only, default)}
}
\value{
ggplot object
}
\description{
Creates a map of annual cold pool area highlighting max, min, and most recent
year, with inset time series of area.
}
