% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_comdat_profit.R
\name{plot_comdat_profit}
\alias{plot_comdat_profit}
\title{plot comdat profitability}
\usage{
plot_comdat_profit(
  shadedRegion = NULL,
  report = "MidAtlantic",
  varName = "all",
  EPU = "MAB",
  n = 0
)
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10)}

\item{report}{Character string. Which SOE report ("MidAtlantic", "NewEngland")}

\item{varName}{Character string. Which Variable to plot ("all","profit_index", "cost_index", "revenue_index")}

\item{EPU}{Character string. Which EPU for New England report ("GB", "GOM") Mid will always be "MAB"}

\item{n}{Numeric scalar. Number of years used (from most recent year) to estimate short term trend . Default = 0 (No trend calculated)}
}
\value{
ggplot object
}
\description{
Plots Geret's profitability indices
}
