% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_commercial_div.R
\name{plot_commercial_div}
\alias{plot_commercial_div}
\title{plot commercial_div}
\usage{
plot_commercial_div(
  shadedRegion = NULL,
  report = "MidAtlantic",
  varName = "Fleet count"
)
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10)}

\item{report}{Character string. Which SOE report ("MidAtlantic", "NewEngland")}

\item{varName}{Character string. Which Variable to plot ("Fleet count",
"Fleet diversity in revenue", "Permit revenue species diversity")}
}
\value{
ggplot object
}
\description{
Plot time series of commercial fleet diversity (fleet count, fleet revenue,
or species permit revenue). Mid Atlantic or New England, no EPUs.
}
