% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_community_climate_vulnerability.R
\name{plot_community_climate_vulnerability}
\alias{plot_community_climate_vulnerability}
\title{plot community climate vulnerability}
\usage{
plot_community_climate_vulnerability(
  shadedRegion = NULL,
  report = "MidAtlantic",
  varName = "vulnSum",
  plottype = "propcomm",
  n = 0
)
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10)}

\item{report}{Character string. Which SOE report ("MidAtlantic", "NewEngland")}

\item{varName}{Character string. Which variable: ocean acidification sensitivity,
temperature sensitivity, stock biomass sensitivity, total sensitivity, total vulnerability
("oaSum","tempSum","stckSum","sensSum","vulnSum")}

\item{plottype}{Character string. Which plot: proportion of communities (default)
or regional sensitivity by revenue or landings ("propcomm", "regionrev", "regionland")}

\item{n}{Numeric scalar. Number of years used (from most recent year) to estimate short term trend . Default = 0 (No trend calculated)}
}
\value{
ggplot object
}
\description{
Plots time series of proportion of communities in each climate sensitivity category, or regional sensitivities
}
