% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_demo.R
\name{plot_demo}
\alias{plot_demo}
\title{plot demo for SOE report}
\usage{
plot_demo(shadedRegion = NULL, report = "MidAtlantic", n = 0)
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10)}

\item{report}{Character string. Which SOE report ("MidAtlantic", "NewEngland")}

\item{n}{Numeric scalar. Number of years used (from most recent year) to estimate short term trend . Default = 0 (No trend calculated)}
}
\value{
ggplot object
}
\description{
Plots demo timeseries and EPU map for SOE presentations
}
