% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_energy_density.R
\name{plot_energy_density}
\alias{plot_energy_density}
\title{plot energy_density}
\usage{
plot_energy_density(shadedRegion = shadedRegion, report = "MidAtlantic")
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10)}

\item{report}{Character string. Which SOE report ("MidAtlantic", "NewEngland")}
}
\value{
ggplot object
}
\description{
Creates multiplanel plot of energy density time series by species.
}
