% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_engagement.R
\name{plot_engagement}
\alias{plot_engagement}
\title{plot fishery engagement and community social vulnerability}
\usage{
plot_engagement(
  shadedRegion = NULL,
  report = "MidAtlantic",
  varName = "Commercial"
)
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10)}

\item{report}{Character string. Which SOE report ("MidAtlantic", "NewEngland")}

\item{varName}{Character string. Which Fishery to plot ("Commercial","Recreational")}
}
\value{
ggplot object
}
\description{
Plot population relative engagement vs engagement by community, recreational or commercial fishery.
Provide table of social vulnerability indicators for highly engaged communities.
Indicator definitions https://www.fisheries.noaa.gov/national/socioeconomics/social-indicators-supporting-information
}
