% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_forage_anomaly.R
\name{plot_forage_anomaly}
\alias{plot_forage_anomaly}
\title{plot forage anomaly}
\usage{
plot_forage_anomaly(shadedRegion = NULL, report = "MidAtlantic")
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10)}

\item{report}{Character string. Which SOE report ("MidAtlantic", "NewEngland")}
}
\value{
ggplot object
}
\description{
Plots ecodata::forage_anomaly
}
