% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_function_template.R
\name{plot_function_template}
\alias{plot_function_template}
\title{plot dataset name}
\usage{
plot_function_template(shadedRegion = NULL, report = "MidAtlantic")
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10)}

\item{report}{Character string. Which SOE report ("MidAtlantic", "NewEngland")}
}
\value{
ggplot object
}
\description{
Insert description of plot function.
}
