% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gom_salmon.R
\name{plot_gom_salmon}
\alias{plot_gom_salmon}
\title{plots Gulf of Maine Salmon}
\usage{
plot_gom_salmon(shadedRegion = NULL, report = "NewEngland", n = 0)
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10)}

\item{report}{Character string. Which SOE report ("MidAtlantic", "NewEngland")}

\item{n}{Numeric scalar. Number of years used (from most recent year) to estimate short term trend . Default = 0 (No trend calculated)}
}
\value{
ggplot object
}
\description{
plots gom_salmon data set
}
