% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_grayseal.R
\name{plot_grayseal}
\alias{plot_grayseal}
\title{plot grayseal}
\usage{
plot_grayseal(shadedRegion = NULL, report = "MidAtlantic")
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10)}

\item{report}{Character string. Which SOE report ("MidAtlantic", "NewEngland")}
}
\value{
ggplot object
}
\description{
Time series of annual and 5 year rolling mean gray seal bycatch with potential
biological removal (PRB) threshold. Same plot for both regions.
}
