% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_habitat_diversity.R
\name{plot_habitat_diversity}
\alias{plot_habitat_diversity}
\title{plot habitat diversity}
\usage{
plot_habitat_diversity(
  shadedRegion = NULL,
  report = "MidAtlantic",
  varName = "Diversity"
)
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10)}

\item{report}{Character string. Which SOE report ("MidAtlantic", "NewEngland")}

\item{varName}{Character string. Which Variable to plot ("Diversity","Richness")}
}
\value{
ggplot object
}
\description{
Plots ecodata::habitat_diversity (Shannon index and species richness)
}
