% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_harborporpoise.R
\name{plot_harborporpoise}
\alias{plot_harborporpoise}
\title{plot harborporpoise}
\usage{
plot_harborporpoise(shadedRegion = shadedRegion, report = "MidAtlantic")
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10)}

\item{report}{Character string. Which SOE report ("MidAtlantic", "NewEngland")}
}
\value{
ggplot object
}
\description{
Time series of annual and 5 year rolling mean harbor porpoise bycatch with potential
biological removal (PRB) threshold. Same plot for both regions.
}
