% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatwave_peak_date.R
\name{plot_heatwave_peak_date}
\alias{plot_heatwave_peak_date}
\title{plot heatwave_peak_date}
\usage{
plot_heatwave_peak_date(shadedRegion = NULL, report = "MidAtlantic")
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10)}

\item{report}{Character string. Which SOE report ("MidAtlantic", "NewEngland")}
}
\value{
ggplot object
}
\description{
Map of SST anomaly on date of peak extreme temperature event for a given year.
NOTE DATE NOT CONTAINED IN DATASET. SHOULD BE
}
