% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatwave_year.R
\name{plot_heatwave_year}
\alias{plot_heatwave_year}
\title{plot heatwave_year}
\usage{
plot_heatwave_year(
  shadedRegion = shadedRegion,
  report = "MidAtlantic",
  varName = "Surface"
)
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10)}

\item{report}{Character string. Which SOE report ("MidAtlantic", "NewEngland")}

\item{varName}{Character string. Which Variable to plot ("Surface", "Bottom", "withtrend")
default variables Surface and Bottom are detrended; option to print surface including
climate change signal is "withtrend"}
}
\value{
ggplot object
}
\description{
Plot time series of daily detrended surface or bottom temperature for current year with
climatology in the background.
}
