% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hms_cpue.R
\name{plot_hms_cpue}
\alias{plot_hms_cpue}
\title{plot hms_cpue}
\usage{
plot_hms_cpue(shadedRegion = NULL, report = "MidAtlantic", varName = "shark")
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10)}

\item{report}{Character string. Which SOE report ("MidAtlantic" default, same plot for both)}

\item{varName}{Character string. Which Variable to plot ("shark", "tuna"). Sharks
are categorized as large coastal, pelagic, prohibited, and small coastal, while
tuna are by species.}
}
\value{
ggplot object
}
\description{
Plot time series of NEUS Highly Migratory Species (HMS) groups: sharks or tunas.
}
