% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hms_landings.R
\name{plot_hms_landings}
\alias{plot_hms_landings}
\title{plot hms_landings}
\usage{
plot_hms_landings(shadedRegion = shadedRegion, report = "MidAtlantic")
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10)}

\item{report}{Character string. Which SOE report ("MidAtlantic", "NewEngland").
HMS landings are by Mid Atlantic or New England, not EPU}
}
\value{
ggplot object
}
\description{
Plot time series of Highly Migratory Species (HMS) commercial landings.
}
