% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_HMS_species_distribution.R
\name{plot_HMS_species_distribution}
\alias{plot_HMS_species_distribution}
\title{plot HMS species distribution}
\usage{
plot_HMS_species_distribution(
  shadedRegion = shadedRegion,
  report = "MidAtlantic"
)
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10)}

\item{report}{Character string. Which SOE report ("MidAtlantic", "NewEngland")}
}
\value{
ggplot object
}
\description{
Plots distribution shifts for marine mammals
}
