% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hms_stock_status.R
\name{plot_hms_stock_status}
\alias{plot_hms_stock_status}
\title{plot hms_stock_status}
\usage{
plot_hms_stock_status(shadedRegion = NULL, report = "MidAtlantic")
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10)}

\item{report}{Character string. Which SOE report ("MidAtlantic" default, same for both)}
}
\value{
list of 2 items

\item{p}{ggplot object}
\item{unknown}{data frame listing stocks with unklnown status}
}
\description{
Kobe plot of F/Fmsy vs B/Bmsy for Highly Migratory Species (HMS) stocks.
}
