% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hudson_river_flow.R
\name{plot_hudson_river_flow}
\alias{plot_hudson_river_flow}
\title{plot flow rate of hudson river}
\usage{
plot_hudson_river_flow(shadedRegion = shadedRegion, report = "MidAtlantic")
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10)}

\item{report}{Character string. Which SOE report ("MidAtlantic", "NewEngland")}
}
\value{
ggplot object
}
\description{
Plot hudson_river_flow data set
}
