% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_long_term_sst.R
\name{plot_long_term_sst}
\alias{plot_long_term_sst}
\title{plot long term sea surface temperature}
\usage{
plot_long_term_sst(shadedRegion = shadedRegion, report = "MidAtlantic")
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10)}

\item{report}{Character string. Which SOE report ("MidAtlantic", "NewEngland")}
}
\value{
ggplot object
}
\description{
plots long_term_sst data set. This is a shelf wide product
}
