% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lps_sharks.r
\name{plot_lps_sharks}
\alias{plot_lps_sharks}
\title{plot shark fishing landings from large pelagics survey}
\usage{
plot_lps_sharks(
  shadedRegion = NULL,
  report = "MidAtlantic",
  varName = "Total",
  n = 0
)
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10)}

\item{report}{Character string. Which SOE report ("MidAtlantic", "NewEngland")}

\item{varName}{Character string. Which variable to plot ("Total", "Blue_Shark", "Common_Thresher", "Shortfin_Mako")
Default plot is Total pelagic sharks.}

\item{n}{Numeric scalar. Number of years used (from most recent year) to estimate short term trend . Default = 0 (No trend calculated)}
}
\value{
ggplot object
}
\description{
Plots lps_sharks data set
}
