% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_narw.R
\name{plot_narw}
\alias{plot_narw}
\title{plot north atlantic right whale abundance}
\usage{
plot_narw(
  shadedRegion = shadedRegion,
  report = "MidAtlantic",
  varName = "adult"
)
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10)}

\item{report}{Character string. Which SOE report ("MidAtlantic", "NewEngland")}

\item{varName}{Character string. Which variable to plot ("adult","calf")}
}
\value{
ggplot object
}
\description{
plots narw data set. calf and adult abundance. This is a shelf wide product
}
