% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_phyto_size.R
\name{plot_phyto_size}
\alias{plot_phyto_size}
\title{plot phytoplankton by size class}
\usage{
plot_phyto_size(shadedRegion = NULL, report = "MidAtlantic", epu = "MAB")
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10)}

\item{report}{Character string. Which SOE report ("MidAtlantic", "NewEngland")}

\item{epu}{Character string. Specify the EPU to plot within the report (Only affects NewEngland)}
}
\value{
ggplot object
}
\description{
plots phyto_size data set.
}
