% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_productivity_anomaly.R
\name{plot_productivity_anomaly}
\alias{plot_productivity_anomaly}
\title{plot productivity_anomaly}
\usage{
plot_productivity_anomaly(
  shadedRegion = NULL,
  report = "MidAtlantic",
  varName = "anomaly",
  EPU = "MAB"
)
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10)}

\item{report}{Character string. Which SOE report ("MidAtlantic", "NewEngland")}

\item{varName}{Character string. Which variable to plot ("anomaly","assessment")}

\item{EPU}{Character string. Which EPU for New England report ("GB", "GOM") Mid will always be MAB}
}
\value{
ggplot object
}
\description{
plots the productivity anomaly dataset
}
