% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rec_hms.r
\name{plot_rec_hms}
\alias{plot_rec_hms}
\title{plot recreational shark fishing landings}
\usage{
plot_rec_hms(shadedRegion = shadedRegion, report = "MidAtlantic")
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10)}

\item{report}{Character string. Which SOE report ("MidAtlantic", "NewEngland")}
}
\value{
ggplot object
}
\description{
Plots rec_hms data set
}
