% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_SAV.R
\name{plot_SAV}
\alias{plot_SAV}
\title{plot SAV}
\usage{
plot_SAV(shadedRegion = NULL, report = "MidAtlantic")
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10)}

\item{report}{Character string. Which SOE report ("MidAtlantic", "NewEngland")}
}
\value{
ggplot object
}
\description{
Plots Mid-Atlantic Submerged Aquatic Vegetation SAV time series
}
