% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_seabird_ne.R
\name{plot_seabird_ne}
\alias{plot_seabird_ne}
\title{plot New England sea bird population}
\usage{
plot_seabird_ne(
  shadedRegion = NULL,
  report = "NewEngland",
  varName = "diversity"
)
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10)}

\item{report}{Character string. Which SOE report ("MidAtlantic", "NewEngland")}

\item{varName}{Character string. Variable to plot ("diversity","productivity","prey")}
}
\value{
ggplot object
}
\description{
Plots seabird_ne data set.
Diet diversity, prey composition,
}
