% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_seal_pups.R
\name{plot_seal_pups}
\alias{plot_seal_pups}
\title{plot Gray seal pup populations}
\usage{
plot_seal_pups(shadedRegion = shadedRegion, report = "MidAtlantic")
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10)}

\item{report}{Character string. Which SOE report ("MidAtlantic", "NewEngland")}
}
\value{
ggplot object
}
\description{
Plots seal_pups data set
}
