% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_seasonal_bt_anomaly_gridded.R
\name{plot_seasonal_bt_anomaly_gridded}
\alias{plot_seasonal_bt_anomaly_gridded}
\title{plot botttom temperature anomaly}
\usage{
plot_seasonal_bt_anomaly_gridded(shadedRegion = NULL, report = "MidAtlantic")
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10), passed from plot function}

\item{report}{Character string. Which SOE report ("MidAtlantic", "NewEngland"), passed from plot function}
}
\value{
ggplot object
}
\description{
plot botttom temperature anomaly
}
