% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_seasonal_oisst_anom.R
\name{plot_seasonal_oisst_anom}
\alias{plot_seasonal_oisst_anom}
\title{plot SST anomaly (OISST)}
\usage{
plot_seasonal_oisst_anom(shadedRegion = shadedRegion, report = "MidAtlantic")
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10), passed from plot function}

\item{report}{Character string. Which SOE report ("MidAtlantic", "NewEngland"), passed from plot function}
}
\value{
ggplot object
}
\description{
plot SST anomaly (OISST)
}
