% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_seasonal_oisst_anom_gridded.r
\name{plot_seasonal_oisst_anom_gridded}
\alias{plot_seasonal_oisst_anom_gridded}
\title{plot seasonal_oisst_anom_gridded}
\usage{
plot_seasonal_oisst_anom_gridded(
  shadedRegion = NULL,
  report = "MidAtlantic",
  scale = "celsius"
)
}
\arguments{
\item{scale}{character string. celsius or fahrenheit. Default = "celsius"}

\item{season}{Character string. Season to plot. (Default = NULL, plot all seasons)}

\item{region}{Character vector. Regional EPUs ("GB","MAB") to overly on figure. (Default = NULL, use all)}
}
\value{
ggplot object
}
\description{
plot seasonal_oisst_anom_gridded
}
