% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_seasonal_sst_anomaly_gridded.r
\name{plot_seasonal_sst_anomaly_gridded}
\alias{plot_seasonal_sst_anomaly_gridded}
\title{plot seasonal_sst_anomaly_gridded}
\usage{
plot_seasonal_sst_anomaly_gridded(
  shadedRegion = shadedRegion,
  report = "MidAtlantic"
)
}
\arguments{
\item{season}{Character string. Season to plot. (Default = NULL, plot all seasons)}

\item{region}{Character vector. Regional EPUs ("GB","MAB") to overly on figure. (Default = NULL, use all)}
}
\value{
ggplot object
}
\description{
plot seasonal_sst_anomaly_gridded
}
