% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_setup.R
\name{plot_setup}
\alias{plot_setup}
\title{Setup all plot defaults}
\usage{
plot_setup(shadedRegion = c(2014, 2023), report = report)
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot}

\item{EPUs}{Character string. Which SOE report ("GB","MAB")}

\item{shade.fill}{Character string. Color of shaded region. (Default = "lightgrey)}

\item{shade.alpha}{Numeric scalar. Alpha of shaded region (Default = 0.5)}
}
\value{
list of settings
}
\description{
Single function setting up plot defaults for SOE, catalog, presentations, and other products
}
