% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_slopewater.R
\name{plot_slopewater}
\alias{plot_slopewater}
\title{plot slopewater}
\usage{
plot_slopewater(shadedRegion = NULL, report = "MidAtlantic", n = 0)
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10)}

\item{report}{Character string. Which SOE report ("MidAtlantic", "NewEngland")}

\item{n}{Numeric scalar. Number of years used (from most recent year) to estimate short term trend . Default = 0 (No trend calculated)}
}
\value{
ggplot object
}
\description{
plots warm and labridor slopewater entering the shelf. This is a shelfwide indicator
}
