% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spawn_timing.R
\name{plot_spawn_timing}
\alias{plot_spawn_timing}
\title{plot spawn_timing}
\usage{
plot_spawn_timing(
  shadedRegion = NULL,
  report = "MidAtlantic",
  varName = "Resting",
  n = 0
)
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10)}

\item{report}{Character string. Which SOE report ("MidAtlantic", "NewEngland")}

\item{varName}{Character string. Which variable to plot:
maturity stage ("Resting", "Ripe", "Spent", "Developing"),
number of mature females ("MF"),
mean sampled bottom temperature ("meanTEMP"),
or mean sampled day of year ("meanJDAY")}

\item{n}{Numeric scalar. Number of years used (from most recent year) to estimate short term trend . Default = 0 (No trend calculated)}
}
\value{
ggplot object
}
\description{
Plots time series of maturity stage and associated data for available stocks
}
