% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sst.R
\name{plot_sst}
\alias{plot_sst}
\title{plot SST}
\usage{
plot_sst(season = NULL, region = NULL)
}
\arguments{
\item{season}{Character string. Season to plot. (Default = NULL, plot all seasons)}

\item{region}{Character vector. Regional EPUs ("GB","MAB") to overly on figure. (Default = NULL, use all)}
}
\value{
ggplot object
}
\description{
plot SST
}
