% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sst_transition_dates.R
\name{plot_sst_transition_dates}
\alias{plot_sst_transition_dates}
\title{plot SST transition_dates}
\usage{
plot_sst_transition_dates(season = NULL, region = NULL)
}
\arguments{
\item{season}{Character string. Season to plot. (Default = NULL, plot all seasons)}

\item{region}{Character vector. Regional EPUs ("GB","MAB") to overly on figure. (Default = NULL, use all)}
}
\value{
ggplot object
}
\description{
plot SST transition_dates
}
