% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sst_transition_dates_length.R
\name{plot_sst_transition_dates_length}
\alias{plot_sst_transition_dates_length}
\title{plot SST transition_dates}
\usage{
plot_sst_transition_dates_length(
  season = NULL,
  region = NULL,
  shadedRegion = c(2012, 2022),
  shade.fill = "lightgrey",
  shade.alpha = 0.3,
  hline.size = 1,
  hline.alpha = 0.5,
  hline.lty = "dashed"
)
}
\arguments{
\item{season}{Character string. Season to plot. (Default = NULL, plot all seasons)}

\item{region}{Character vector. Regional EPUs ("GB","MAB") to overly on figure. (Default = NULL, use all)}

\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot}

\item{shade.fill}{Character string. Color of shaded region. (Default = "lightgrey)}

\item{shade.alpha}{Numeric scalar. Alpha of shaded region (Default = 0.5)}

\item{hline.size}{Numeric scalar Line width of horizontal line. (Default = 1)}

\item{hline.alpha}{Numeric scalar. Alpha of horizontal line (Default = 0.5)}

\item{hline.lty}{Character string. Line type of horizontal line (Default = "dashed")}
}
\value{
ggplot object
}
\description{
plot SST transition_dates
}
