% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stock_status.R
\name{plot_stock_status}
\alias{plot_stock_status}
\title{plot stock_Status}
\usage{
plot_stock_status(shadedRegion = NULL, report = "MidAtlantic")
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10)}

\item{report}{Character string. Which SOE report ("MidAtlantic", "NewEngland")}
}
\value{
list of 2 items

\item{p}{ggplot object}
\item{unknown}{data frame listing stocks with unklnown status}
}
\description{
Kobe plots of regional stock status
}
