% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stom_fullness.R
\name{plot_stom_fullness}
\alias{plot_stom_fullness}
\title{plot stomach fullness}
\usage{
plot_stom_fullness(shadedRegion = NULL, report = "MidAtlantic", EPU = "MAB")
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10)}

\item{report}{Character string. Which SOE report ("MidAtlantic", "NewEngland")}
}
\value{
ggplot object
}
\description{
plots \link{stom_fullness}
}
