% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_survey_shannon.R
\name{plot_survey_shannon}
\alias{plot_survey_shannon}
\title{plot dataset name}
\usage{
plot_survey_shannon(shadedRegion = shadedRegion, report = "MidAtlantic")
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10)}

\item{report}{Character string. Which SOE report ("MidAtlantic", "NewEngland")}
}
\value{
ggplot object
}
\description{
Insert description of plot function.
}
