% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_thermal_habitat_gridded.R
\name{plot_thermal_habitat_gridded}
\alias{plot_thermal_habitat_gridded}
\title{plot thermal habitat persistence}
\usage{
plot_thermal_habitat_gridded(
  shadedRegion = NULL,
  report = "MidAtlantic",
  year = NULL
)
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10)}

\item{report}{Character string. Which SOE report ("MidAtlantic", "NewEngland")}
}
\value{
ggplot object
}
\description{
plots thermal_habitat_gridded data set. Current SOE Year only
}
