% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wcr.R
\name{plot_wcr}
\alias{plot_wcr}
\title{plot Warm core rings}
\usage{
plot_wcr(shadedRegion = NULL, report = "MidAtlantic")
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10)}

\item{report}{Character string. Which SOE report ("MidAtlantic", "NewEngland")}
}
\value{
ggplot object
}
\description{
plots wcr dataset
}
