% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wea_landings_rev.R
\name{plot_wea_landings_rev}
\alias{plot_wea_landings_rev}
\title{plot revenue from top 10 landings}
\usage{
plot_wea_landings_rev(shadedRegion = NULL, report = "MidAtlantic", n = 0)
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10)}

\item{report}{Character string. Which SOE report ("MidAtlantic", "NewEngland")}

\item{n}{numeric scalar. The number of species to show (default = n = 0, all species)}
}
\value{
flextable object
}
\description{
plot wea_landings_rev.
}
