% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wind_dev_speed.R
\name{plot_wind_dev_speed}
\alias{plot_wind_dev_speed}
\title{plot wind lease cumulative areas}
\usage{
plot_wind_dev_speed(shadedRegion = shadedRegion, report = "MidAtlantic")
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10)}

\item{report}{Character string. Which SOE report ("MidAtlantic", "NewEngland")}
}
\value{
ggplot object
}
\description{
plot wind_dev_speed. This is shelf wide
}
