% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wind_port.R
\name{plot_wind_port}
\alias{plot_wind_port}
\title{plot revenue by port from wind energy}
\usage{
plot_wind_port(shadedRegion = NULL, report = "MidAtlantic")
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10)}

\item{report}{Character string. Which SOE report ("MidAtlantic", "NewEngland")}
}
\value{
ggplot object
}
\description{
plot wind_port data
}
