% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wind_revenue.R
\name{plot_wind_revenue}
\alias{plot_wind_revenue}
\title{plot wind revenue}
\usage{
plot_wind_revenue(
  shadedRegion = NULL,
  report = "MidAtlantic",
  varName = "landing",
  plottype = "facets",
  n = 0
)
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10)}

\item{report}{Character string. Which SOE report ("MidAtlantic", "NewEngland")}

\item{varName}{Character string. Which Variable to plot ("landing","value")}

\item{plottype}{Character string. Which plot ("facets", "nofacets")}

\item{n}{Numeric scalar. Number of years used (from most recent year) to estimate short term trend . Default = 0 (No trend calculated)}
}
\value{
ggplot object
}
\description{
plot wind_revenue data set. Only 5 species are plotted.
}
