% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_zoo_abundance_anom.R
\name{plot_zoo_abundance_anom}
\alias{plot_zoo_abundance_anom}
\title{plot zooplankton abundance anomaly}
\usage{
plot_zoo_abundance_anom(
  shadedRegion = NULL,
  report = "MidAtlantic",
  varName = "copepod"
)
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10)}

\item{report}{Character string. Which SOE report ("MidAtlantic", "NewEngland")}

\item{varName}{Character string. Which variable to plot ("copepod","euphausid")}
}
\value{
ggplot object
}
\description{
Plots zoo_abundance_anom data set
}
