% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_zoo_regime.R
\name{plot_zoo_regime}
\alias{plot_zoo_regime}
\title{plot zoo_regime}
\usage{
plot_zoo_regime(shadedRegion = NULL, report = "MidAtlantic")
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10)}

\item{report}{Character string. Which SOE report ("MidAtlantic", "NewEngland")}
}
\value{
ggplot object
}
\description{
Plots Zooplankton Abundance anomaly
}
