% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_zoo_strat_abun.R
\name{plot_zoo_strat_abun}
\alias{plot_zoo_strat_abun}
\title{plot abundance of Euphasids and Cnids}
\usage{
plot_zoo_strat_abun(
  shadedRegion = NULL,
  report = "MidAtlantic",
  EPU = "MAB",
  n = 0
)
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10)}

\item{report}{Character string. Which SOE report ("MidAtlantic" only, default)}

\item{EPU}{Character string. Which EPU in the report ("MAB" only, default)}

\item{n}{Numeric scalar. Number of years used (from most recent year) to estimate short term trend . Default = 0 (No trend calculated)}
}
\value{
ggplot object
}
\description{
Plots zoo_abund_strat data set
}
