% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppr.R
\docType{data}
\name{ppr}
\alias{ppr}
\title{Ecosystem Overfishing Indicators}
\format{
550 rows and 5 columns.

\itemize{
\item EPU: Ecological Production Unit (EPU) where sampling occurred.
\item Time: Year.
\item Var: Ecosystem Overfishing Indicator.
\item Value: Value of variable \code{Var}.
\item Units: Units of variable \code{Var}.
}
}
\usage{
ppr
}
\description{
Data include ecosystem overfishing indicators (Primary Production Required,
Ryther, Fogarty).
}
\details{
More information including processing and indicator derivation steps are available at
\url{https://noaa-edab.github.io/tech-doc/ecosystem-overfishing.html}.
}
\keyword{datasets}
