% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rec_hms.R
\docType{data}
\name{rec_hms}
\alias{rec_hms}
\title{Recreational Higly Migratory Species}
\format{
These data contain 225 rows and 4 columns.

\itemize{
\item Var: HMS species category.
\item Value: Value of variable \code{Var}.
\item Time: Time step of \code{Var}.
\item EPU: Ecological Production Unit (EPU) where sampling occurred.
}
}
\usage{
rec_hms
}
\description{
Recreational Higly Migratory Species
}
\details{
Methods used to calculate data are available
at \url{https://noaa-edab.github.io/tech-doc/highly-migratory-species-landings.html}.
}
\keyword{datasets}
