% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regrowth}
\alias{regrowth}
\title{Regrowth}
\format{
A 'data.frame': 40 obs. of 3 variables:
\describe{
 \item{Root}{Root biomass (numeric).}
 \item{Fruit}{Fruit biomass (numeric).}
 \item{Grazing}{Grazed or not (factor).}
}
}
\usage{
regrowth
}
\description{
Plant regrowth observations
}
\examples{
regrowth_df <- regrowth
head(regrowth_df)

str(regrowth_df)

# factorize variables -------------------------------------------------
regrowth_df$Grazing = as.factor(regrowth_df$Grazing )


plot(regrowth_df$Grazing, regrowth_df$Root)
plot(regrowth_df$Grazing, regrowth_df$Fruit)


library(glmmTMB)
# binomial model -------------------------------------------------
fit <-  glmmTMB(Grazing ~  Root ,family = binomial , data = regrowth_df)
summary(fit)


library(DHARMa)
res <- simulateResiduals(fit)
plot(res)


DHARMa::testQuantiles(res)




}
\author{
Florian Hartig
}
\keyword{datasets}
