% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandlance.R
\docType{data}
\name{sandlance}
\alias{sandlance}
\title{Stellwagen Bank Sandlance}
\format{
Data set contains 61 rows and 4 columns

\itemize{
\item Value: Value of variable \code{Var}.
\item Units: Units of variable \code{Var}.
\item EPU: Ecological Production Unit (EPU) where sampling occurred. All data were sampled in southern Gulf of Maine.
\item Time: Year of sampling.
}

More information including processing and indicator derivation steps are available at
\url{https://noaa-edab.github.io/tech-doc/sandlance.html}.
}
\usage{
sandlance
}
\description{
These data are time series sandlance collection and co-location.
}
\keyword{datasets}
