% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sav.R
\docType{data}
\name{SAV}
\alias{SAV}
\title{Chesapeake Bay Submerged Aquatic Vegetation}
\format{
Data set contains 185 rows and 4 columns

\itemize{
\item Value: Value of variable \code{Var}.
\item Units: Units of variable \code{Var}.
\item EPU: Ecological Production Unit (EPU) where sampling occurred. All data were sampled in southern Gulf of Maine.
\item Time: Year of sampling.
}

More information including processing and indicator derivation steps are available at
\url{https://noaa-edab.github.io/tech-doc/submerged-aquatic-vegetation.html}.
}
\usage{
SAV
}
\description{
These data are time series Submerged Aquatic Vegetation (SAV) in the
Chesapeake Bay.
}
\keyword{datasets}
