% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scoutingAnts}
\alias{scoutingAnts}
\title{Scouting Ants data}
\format{
An object of class \code{data.frame} with 3785 rows and 9 columns.
}
\usage{
scoutingAnts
}
\description{
An experiment to test if there are scouting ants in the species Lasius niger.
}
\details{
The idea of this experiment was to see if there are certain individuals that do not follow pheromone trails (scouts). The idea for this originated because pheromone following behavior in Lasius niger is only around 70\%, unusually low for ants. The question is thus if they just often don't follow on average, or if there are specialized individuals that do not follow pheromone trails.

In the experiments, ants were put 2 times in a y-maze. We want to see if their behaviour in the first visit predicts their behavior in the second visit. If that was the case, we would conclude that there are individuals that predictably do not follow pheromone trails (which may indicate that they are scouts, but could of course also mean that their sense of smell is impaired).

Variables:
\itemize{
\item ant_group - for the first visit, whether the ant followed the pheromone trail or not, for the second visit, whether there was a switch or not
\item ant_group_main - aggregation of ant_group - for observations of the first visit, just "Firstvisit", for the second visit, split by decision
\item first.visit	a logical column returning 1 if this is the first visit, and 0 if it is not
\item went.phero	a logical column returning 1 if the ant went towards the pheromone, and 0 if it did not
\item phero.orientation	whether the Y maze arm marked with a pheromone trail was facing into the room, or towards the wall
\item where.phero	whether the Y maze arm marked with a pheromone trail was on the right or lef, from the perspective of the Y-maze stem
\item Orientation	Whether the left arm was facing towards the wall or towards the room
\item Treatment	The direct of the pheromone trail (L or R) at the first or second test
\item Colony	a unique ID for each colony
}
}
\note{
It may make sense to re-define factors / contrasts. For example, you could create a factor to see whether ants went left or right, or whether they switched their direction on the second visit.
}
\examples{

# to undestand the variables better, have a look at the options
table(scoutingAnts$ant_group_main)
table(scoutingAnts$ant_group_main)

# probably you want to convert to factor before continuing the analysis
}
\references{
Koch, Alexandra, and Tomer J. Czaczkes. "No specialist pheromone-ignoring ants in Lasius niger." Ecological Entomology 46 (2020): 677-680.
}
\author{
Florian Hartig
}
\keyword{datasets}
