% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seabird_mab.R
\docType{data}
\name{seabird_mab}
\alias{seabird_mab}
\title{Virginia waterbor productivity}
\format{
Data set contains 24 rows and 5 columns

\itemize{
\item Group: Group of waterbirds.
\item Value: Value of variable \code{Var}.
\item Units: Units of variable \code{Var}.
\item EPU: Ecological Production Unit (EPU) where sampling occurred. All data were sampled in southern Gulf of Maine.
\item Time: Year of sampling.
}

More information including processing and indicator derivation steps are available at
\url{https://noaa-edab.github.io/tech-doc/ma-waterbird-productivity.html}.
}
\usage{
seabird_mab
}
\description{
These data are time series of productivity for groups of Virginia waterbirds
}
\keyword{datasets}
