% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seal_pups.R
\docType{data}
\name{seal_pups}
\alias{seal_pups}
\title{Seal Pups counts in New England}
\format{
Data set contains 340 rows and 4 columns

\itemize{
\item Time: Year.
\item Var: Variable and colony name.
\item Value: Value of variable \code{Var}.
\item EPU: Ecological Production Unit (EPU) where sampling occurred. All data were sampled in southern Gulf of Maine.
}
}
\usage{
seal_pups
}
\description{
Data include timeseries of seal pup count on four colonies.
}
\details{
More information including processing and indicator derivation steps are available at
\url{https://noaa-edab.github.io/tech-doc.html}.
}
\keyword{datasets}
