% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal_sst_anomaly.R
\docType{data}
\name{seasonal_sst_anomaly}
\alias{seasonal_sst_anomaly}
\title{Sea-surface temperature anomaly}
\format{Data set contains 15536 rows and 4 columns

\itemize{
\item Latitude
\item Longitude: Longitude (-180 to 180 degrees)
\item Value: Mean SST anomaly (degrees C)
\item Season: Season

}}
\source{
Source data for this indicator are available at
\url{https://www.esrl.noaa.gov/psd/data/gridded/data.noaa.oisst.v2.highres.html}
}
\usage{
seasonal_sst_anomaly
}
\description{
Seasonally-averaged smoothed sea-surface temperature anomalies for the Northeast
Continental Shelf; derived from the NOAA optimum interpolation SST
high resolution data set (NOAA OISST V2). The data extend from 1981 to present,
and provide a 0.25 x 0.25 global grid of SST measurements. The 1982-2010
climatology to calculate anomalies.
}
\details{
Seasons are defined as: Fall = October, November,
December; Winter = January, February, March; Spring = April, May, June;
Summer = July, August, September. The CRS used to manipulate these data was
"+proj=longlat +lat_1=35 +lat_2=45 +lat_0=40 +lon_0=-77 +x_0=0 +y_0=0 +datum=NAD83
+no_defs +ellps=GRS80 +towgs84=0,0,0"
}
\keyword{datasets}
