% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slopewater.R
\docType{data}
\name{slopewater}
\alias{slopewater}
\title{Northeast Channel slopewater proportions}
\format{
80 rows and 5 columns

\itemize{
\item Var: Specifies slopewater source (either Labrador Slope Water (LSLW) or Warm Slope Water (WSW))
\item Value: Proportion of slopewater derived from distinct source.
\item Units: Units of variable \code{Var}; proportion.
\item Time: Year.
\item EPU: Ecological Production Unit where data originated.
}
}
\usage{
slopewater
}
\description{
Percent total of water type observed in the deep Northeast Channel (150-200 m water depth).
}
\details{
Read more about the roles of different slopewater types in the Gulf of Maine
at \url{https://www.integratedecosystemassessment.noaa.gov/regions/northeast/components/other-physical-drivers}.
}
\keyword{datasets}
