% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soe.R
\docType{package}
\name{soe}
\alias{soe}
\alias{soe-package}
\title{Package Title}
\description{
A description of the data package
}
\details{
soe
A data package for soe.

Use \code{data(package='soe')$results[, 3]} tosee a list of availabledata sets in this data package
and/or DataPackageR::load_all
_datasets() to load them.
}
\seealso{
NA
\link{soe}
}
