% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{soep}
\alias{soep}
\title{Data from the german socio-economic panel}
\format{
An object of class \code{data.frame} with 23522 rows and 15 columns.
}
\source{
https://environmentalcomputing.net/statistics/linear-models/linear-regression/
}
\usage{
soep
}
\description{
Data from the german socio-economic panel
}
\details{
The following information is copied from the origina data provider at https://www.diw.de/en/diw_01.c.836543.en/soep_practice_dataset.html:

This dataset  is based on the original SOEP data, but provides the data in significantly altered and fully anonymous form. This means that the practice dataset can be used without the need for any contracts or user agreements. The practice dataset consists of original variables, covers five time points, and is available in the “long” format. The dataset is provided in German and English.

Dataset Information

Title: SOEP practice dataset

DOI info : 10.5684/soep.practice.v36
Collection period: 2015-2019
Publication Date: 31.01.2022

Population: Persons living in private households in Germany.

Number of individuals: 6.355


\itemize{
  \item id		Person ID 
  \item syear		Survey Year
  \item sex		Sex
  \item alter		Age of Individual
  \item anz_pers		Number of Persons in Household
  \item anz_kind		Number of Children in Household
  \item bildung		Number of Years of Education 
  \item erwerb		Employment Status
  \item branche		Industry Current Occupation
  \item einkommenj1		Gross Income from Main Job/Year
  \item einkommenj2		Gross Income from Secondary Employment/Year
  \item einkommenm1		Gross Income from Main Job/Month
  \item einkommenm2		Gross Income from Secondary Employment/Month
  \item gesund_org		subj. Health
  \item lebensz_org		Life Satisfaction
}

Selection: To anonymize the variables, an algorithm was used that largely maintains the longitudinal consistency in the original data.

Usability: The practice dataset is suited to calculating panel-specific univariate statistics (intra- and inter-individual correlation patterns, transition rates) in classes on descriptive methods. In the context of multivariate analysis, the dataset is useful for teaching (panel) regression techniques

Citation of data: SOEP Practice Dataset, Data from  2015-2019. 2022. DOI: 10.5684/soep.practice.v36

This data is licenced under CC BY-SA 4.0
}
\author{
Florian Hartig
}
\keyword{datasets}
