% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{solomonislands}
\alias{solomonislands}
\title{Solomon Islands - Island Biogeography}
\format{
A 'data.frame':	40 obs. of  3 variables:
\describe{
  \item{Island}{Island name (chr)}
  \item{Species}{Species richness on the respective island (num)}
  \item{Area}{Island area (num)}
  \item{Distance}{Distance to mainland (num)}
}
}
\usage{
solomonislands
}
\description{
A classical dataset of island biogeography - number of species found on the Solomon islands, as a function of distance to mainland and island area
}
\examples{
solomonIsland_df <- solomonislands
head(solomonIsland_df)


# factorize variables -------------------------------------------------
solomonIsland_df$Island = as.factor(solomonIsland_df$Island)


# plotting variables -------------------------------------------------
plot(solomonIsland_df$Species, 1:nrow(solomonIsland_df))
plot(solomonIsland_df$Island, solomonIsland_df$Species)

fit <- lm(Species ~ Area + Distance, data=solomonIsland_df)
summary(fit)
image(xtabs(residuals(fit)~Area + Distance, data=solomonIsland_df))
}
\author{
Florian Hartig
}
\keyword{datasets}
