% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_dist.R
\docType{data}
\name{species_dist}
\alias{species_dist}
\title{Species distribution on the NES.}
\format{
Data set contains 275 rows and 5 columns

\itemize{
\item Time: Year.
\item Var: Depth or Distance along shelf or Distance from shelf.
\item Value: Value of variable  \code{Var}.
\item Units: Unit of Variable  \code{Var}.
\item EPU: Ecological Production Unit (EPU) where sampling occurred.
}
}
\usage{
species_dist
}
\description{
Data include timeseries of depth, distance from shelf and distance along shelf.
}
\details{
More information including processing and indicator derivation steps are available at
\url{https://noaa-edab.github.io/tech-doc/species-distribution-indicators.html}.
}
\keyword{datasets}
