% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecodata.R
\name{standardize_frequency}
\alias{standardize_frequency}
\title{Function to standardize time series frequency label}
\usage{
standardize_frequency(freq)
}
\arguments{
\item{freq}{Character string or vector of strings, each describing the frequency of a time series}
}
\value{
Vector of the same length of `freq`, with each value equal to a standard string to describe the frequency.
Return value will be one of c("Annual", "Semiannual", "Quarterly", "Monthly", "Biweekly", "Weekly", "Daily")
}
\description{
To standardize the frequency description across different sources, this function takes as an input some frequency description,
attempts to decipher it, and returns on the following frequencies:
c("Annual", "Semiannual", "Quarterly", "Monthly", "Biweekly", "Weekly", "Daily")
}
