% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecodata.R
\name{standardize_frequency_single}
\alias{standardize_frequency_single}
\title{Function to standardize time series frequency label}
\usage{
standardize_frequency_single(freq)
}
\arguments{
\item{freq}{Character string, describing the frequency of a time series}
}
\value{
Standardized character string to describe the frequency, one of c("Annual", "Semiannual", "Quarterly", "Monthly", "Biweekly", "Weekly", "Daily")
}
\description{
To standardize the frequency description across different sources, this function takes as an input some frequency description,
attempts to decipher it, and returns on the following frequencies:
c("Annual", "Semiannual", "Quarterly", "Monthly", "Biweekly", "Weekly", "Daily")
}
