% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatLM.R
\docType{data}
\name{StatLM}
\alias{StatLM}
\title{StatLM Stat protto based on ecodata StatGLS
Need to pass arguments}
\format{
An object of class \code{StatLM} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 4.
}
\usage{
StatLM
}
\description{
nBootSamples, pVal significance, number datapoints
}
\keyword{datasets}
