% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock_status.R
\docType{data}
\name{stock_status}
\alias{stock_status}
\title{Managed species stock status}
\format{
104 rows and 7 columns

\itemize{
\item Stock: Stock identifier.
\item Last assessment: Last year that a stock assessment was performed for a given stock.
\item Council: Management body (MAFMC, NEFMC, or Both).
\item Code: Shortened identifier for stocks.
\item Var: Either \code{F.Fmsy}, representing the ratio of fishery mortality to fishery mortality
at maximum sustainable yield (or proxy), or \code{B.Bmsy}, defined as the ratio of estimated stock biomass to
estimated biomass at maximum sustainable yield (or proxy).
\item Value: Value of variable \code{Var}.
\item Units: Units of variable \code{Var}.
}
}
\usage{
stock_status
}
\description{
Stock status for MAFMC and NEFMC managed species in New England and the Mid-Atlantic.
}
\details{
More information about this data set may be found at \url{https://noaa-edab.github.io/tech-doc/stockstatus.html}.
}
\keyword{datasets}
