% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stom_fullness.R
\docType{data}
\name{stom_fullness}
\alias{stom_fullness}
\title{Stomach fullness index calculated from NEFSC fall bottom trawl surveys}
\format{
Data set contains 2068 rows and 5 columns

\itemize{
\item Time: Year sampled.
\item Var: Species sampled.
\item Value: Value of variable  \code{Var}.
\item Units: Unit of Variable  \code{Var}.
\item EPU: Ecological Production Unit (EPU) where sampling occurred.
}
}
\usage{
stom_fullness
}
\description{
Data include timeseires of stomach fullness index for 22 species sampled in the NEFSC fall bottom trawl survey.
}
\details{
More information including processing and indicator derivation steps are available at
\url{https://noaa-edab.github.io/tech-doc/stomach-fullness.html}.
}
\keyword{datasets}
