% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecodata.R
\name{str_detect}
\alias{str_detect}
\title{`str_detect()`
Wrapper for `stringr::str_detect()`}
\usage{
str_detect(...)
}
\description{
`str_detect()`
Wrapper for `stringr::str_detect()`
}
\seealso{
[stringr::str_detect()]
}
