% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecodata.R
\name{str_extract}
\alias{str_extract}
\title{Extract a pattern from a string}
\usage{
str_extract(...)
}
\description{
Wrapper for `stringr::str_extract()`
}
\seealso{
[stringr::str_extract()]
}
\keyword{internal}
