% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{streams}
\alias{streams}
\title{Streams}
\format{
A 'data.frame':	16 obs. of  2 variables:
\describe{
  \item{down}{after the water treatment filter}
  \item{up}{before the water treatment filter}
}
}
\source{
\url{https://raw.githubusercontent.com/biometry/APES/master/Data/Simone/streams.txt}
}
\usage{
streams
}
\description{
Data on water measurements of 16 rivers taken before and after a water treatment filter.
}
\author{
Florian Hartig
}
\keyword{datasets}
