% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecodata.R
\name{string_compare}
\alias{string_compare}
\title{Compare two strings, ignoring case and whitespace}
\usage{
string_compare(str1, str2)
}
\arguments{
\item{str1}{String to compare}

\item{str2}{Another string to compare}
}
\value{
Boolen on whether the strings match
}
\description{
Compare two strings, ignoring case and whitespace
}
