% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecodata.R
\name{string_detect}
\alias{string_detect}
\title{`string_detect()`
Detect a pattern in a string, ignoring case}
\usage{
string_detect(str, pattern)
}
\arguments{
\item{str}{String to search}

\item{pattern}{Pattern to look for}
}
\value{
Logical indicating whether or not the pattern was found
}
\description{
`string_detect()`
Detect a pattern in a string, ignoring case
}
