% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecodata.R
\name{string_empty}
\alias{string_empty}
\title{Test if String is Empty}
\usage{
string_empty(text)
}
\arguments{
\item{text}{Text to test if it is empty}
}
\value{
Logical, whether or not the string is empty
}
\description{
Just return TRUE or FALSE if string given is empty ("empty" can include NULL, NA, whitespace only, "")
}
