% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecodata.R
\name{string_not_empty}
\alias{string_not_empty}
\title{`string_not_empty(text)`
Just return TRUE or FALSE if string given is not empty (i.e. not NULL, not NA, not whitespace, not "")}
\usage{
string_not_empty(text)
}
\arguments{
\item{text}{Text to test if it is not empty}
}
\value{
Logical, whether or not the string exists and has characters in it
}
\description{
`string_not_empty(text)`
Just return TRUE or FALSE if string given is not empty (i.e. not NULL, not NA, not whitespace, not "")
}
