% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecodata.R
\name{string_not_empty_single}
\alias{string_not_empty_single}
\title{Test if String Not Empty}
\usage{
string_not_empty_single(text)
}
\arguments{
\item{text}{Character string, text to test if it is not empty}
}
\value{
Logical, whether or not the string exists and has characters in it
}
\description{
Just return TRUE or FALSE if string given is not empty (i.e. not NULL, not NA, not whitespace, not "")
}
