% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{students}
\alias{students}
\title{Students}
\format{
A 'data.frame': 382 obs. of 53 variables:
\describe{
\item{school}{Student's school (binary: 'GP' - Gabriel Pereira or 'MS' - Mousinho da Silveira).}
\item{sex}{Student's sex (binary: 'F' - female or 'M' - male).}
\item{age}{Student's age (numeric: from 15 to 22).}
\item{address}{Student's home address type (binary: 'U' - urban or 'R' - rural).}
\item{famsize}{Family size (binary: 'LE3' - less or equal to 3 or 'GT3' - greater than 3).}
\item{Pstatus}{Parent's cohabitation status (binary: 'T' - living together or 'A' - apart).}
\item{Medu}{Mother's education (numeric: 0: none, 1: primary education (4th grade), 2: 5th to 9th grade, 3: secondary education or 4: higher education).}
\item{Fedu}{Father's education (numeric: 0: none, 1: primary education (4th grade), 2: 5th to 9th grade, 3: secondary education or 4: higher education).}
\item{Mjob}{Mother's job (nominal: 'teacher', 'health' care related, civil 'services' (e.g. administrative or police), 'at_home' or 'other').}
\item{Fjob}{Father's job (nominal: 'teacher', 'health' care related, civil 'services' (e.g. administrative or police), 'at_home' or 'other').}
\item{reason}{Reason to choose this school (nominal: close to 'home', school 'reputation', 'course' preference or 'other').}
\item{guardian}{Student's guardian (nominal: 'mother', 'father' or 'other').}
\item{traveltime}{Home to school travel time (numeric: 1: <15 min., 2: 15 to 30 min., 3: 30 min. to 1 hour, or 4: >1 hour).}
\item{studytime}{Weekly study time (numeric: 1: <2 hours, 2: 2 to 5 hours, 3: 5 to 10 hours, or 4: >10 hours).}
\item{failures}{Number of past class failures (numeric: n if 1 <= n < 3, else 4).}
\item{schoolsup}{Extra educational support (binary: yes or no).}
\item{famsup}{Family educational support (binary: yes or no).}
\item{paid}{Extra paid classes within the course subject (Math or Portuguese) (binary: yes or no).}
\item{activities}{Extra-curricular activities (binary: yes or no).}
\item{nursery}{Attended nursery school (binary: yes or no).}
\item{higher}{Wants to take higher education (binary: yes or no).}
\item{internet}{Internet access at home (binary: yes or no).}
\item{romantic}{With a romantic relationship (binary: yes or no).}
\item{famrel}{Quality of family relationships (numeric: from 1 - very bad to 5 - excellent).}
\item{freetime}{Free time after school (numeric: from 1 - very low to 5 - very high).}
\item{goou}{Going out with friends (numeric: from 1 - very low to 5 - very high).}
\item{Dalc}{Workday alcohol consumption (numeric: from 1 - very low to 5 - very high).}
\item{Walc}{Weekend alcohol consumption (numeric: from 1 - very low to 5 - very high).}
\item{health}{Current health status (numeric: from 1 - very bad to 5 - very good).}
\item{absences}{Number of school absences (numeric: from 0 to 93).}

These grades are related with the course subject, Math or Portuguese:

\item{G1}{First period grade (numeric: from 0 to 20)}
\item{G2}{Second period grade (numeric: from 0 to 20)}
\item{G3}{Final grade (numeric: from 0 to 20, output target)}
}
}
\source{
\url{https://www.kaggle.com/uciml/student-alcohol-consumption}
}
\usage{
students
}
\description{
The data was obtained in a survey of students of math and portuguese language courses in secondary school. It contains a lot of interesting social, gender and study information about students. You can use it for some EDA or try to predict students final grade.
}
\note{
There are several (382) students that belong to both data sets. These students can be identified by searching for identical attributes that characterize each student, as shown in the annexed R file.

Source Information

P. Cortez and A. Silva. Using Data Mining to Predict Secondary School Student Performance. In A. Brito and J. Teixeira Eds., Proceedings of 5th FUture BUsiness TEChnology Conference (FUBUTEC 2008) pp. 5-12, Porto, Portugal, April, 2008, EUROSIS, ISBN 978-9077381-39-7.

Fabio Pagnotta, Hossain Mohammad Amran.

https://archive.ics.uci.edu/ml/datasets/STUDENT+ALCOHOL+CONSUMPTION
Released Under CC0: Public Domain License
}
\examples{
students_df <- students
head(students_df)


# factorize variables -------------------------------------------------
students_df$school = as.factor(students_df$school)
students_df$sex = as.factor(students_df$sex)
students_df$age = as.factor(students_df$age)



# plotting variables -------------------------------------------------
plot(students_df$G1.x, students_df$G1.y)
abline(a = 1, b = 1, col = "red")

plot(students_df$G2.x, students_df$G2.y)
abline(a = 1, b = 1, col = "red")

plot(students_df$G3.x, students_df$G3.y)
abline(a = 1, b = 1, col = "red")


library(glmmTMB)
# gaussian model -------------------------------------------------
fit <-  glmmTMB(G2.x ~  school + sex + age, family = gaussian ,data = students_df)
summary(fit)


library(DHARMa)
res <- simulateResiduals(fit)
plot(res)


DHARMa::testDispersion(res)

DHARMa::testQuantiles(res)
}
\author{
Florian Hartig
}
\keyword{datasets}
