% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_shannon.R
\docType{data}
\name{survey_shannon}
\alias{survey_shannon}
\title{Shannon Diversity of NEFSC Bottom Trawl Survey}
\format{
436 rows and 5 columns.

\itemize{
\item EPU: Ecological Production Unit (EPU) where sampling occurred.
\item Time: Year.
\item Var: Variable describing shannon diversity.
\item Value: Value of variable \code{Var}.
\item Units: Units of variable \code{Var}.
}
}
\usage{
survey_shannon
}
\description{
Data include shannon diversity values for bottom trawl survey catch.
}
\details{
More information available at
\url{https://noaa-edab.github.io/tech-doc/survdat.html}.
}
\keyword{datasets}
