% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_facet.R
\name{theme_facet}
\alias{theme_facet}
\title{A ggplot2 theme for facetted SOE time series figures}
\usage{
theme_facet(...)
}
\description{
A ggplot2 theme for facetted SOE time series figures
}
\examples{
library(ggplot2)
data <- data.frame(x = rep(1:10,2),
                   y = rnorm(20),
                   Var = rep(c("group 1","group 2"), each = 10))

#Plot series with trend and SOE plot theme
ggplot(data = data) +
  geom_line(aes(x = x, y = y)) +
  facet_wrap(Var~.)+
  theme_facet()
  
  
}
