% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_title.R
\name{theme_title}
\alias{theme_title}
\title{A ggplot2 ggtitle formmatting for SOE figures}
\usage{
theme_title(...)
}
\description{
A ggplot2 ggtitle formmatting for SOE figures
}
\examples{
library(ggplot2)
data <- data.frame(x = rep(1:10,2),
                   y = rnorm(20),
                   Var = rep(c("group 1","group 2"), each = 10))

#Plot series with trend and SOE plot theme
ggplot(data = data) +
  geom_line(aes(x = x, y = y)) +
  theme_title()


}
