% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ts.R
\name{theme_ts}
\alias{theme_ts}
\title{A ggplot2 theme for SOE time series figures}
\usage{
theme_ts(...)
}
\description{
A ggplot2 theme for SOE time series figures
}
\examples{
library(ggplot2)
m <- 0.1
x <- 1:30
y <-  m*x + rnorm(30, sd = 0.35)

data <- data.frame(x = x,
                  y = y)

#Plot series with trend and SOE plot theme
ggplot(data = data) +
  geom_line(aes(x = x, y = y)) +
  stat_gls(aes(x = x, y = y, color = stat(col)))+
  theme_ts()
}
