% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{thickness}
\alias{thickness}
\title{thickness of coal}
\format{
An object of class \code{data.frame} with 75 rows and 6 columns.
}
\source{
\url{https://stats.idre.ucla.edu/r/faq/how-do-i-model-a-spatially-autocorrelated-outcome/}
}
\usage{
thickness
}
\description{
Measured thickness of coal seams at different coordinates from 
\describe{
\item{ID}{ID}
\item{east}{east coordinates}
\item{north}{north coordinates}
\item{thick}{thickness of coal seams}
\item{soil}{measuring the soil quality}
\item{dummy}{dummy grouping variable}
}
}
\examples{
thickness_df <- thickness
head(thickness_df)

str(thickness_df)


# ------------------------------------------------------------------------------
library(glmmTMB)

fit <- glmmTMB(thick ~ soil + (1|dummy), data = thickness_df)
summary(fit)

library(DHARMa)
res <- simulateResiduals(fit)
plot(res)


# adding spatial correlation    ------------------------------------------------
fit2 <- glmmTMB(thick ~ soil + (1|dummy) , dispformula = ~ east + north, data = thickness_df)
summary(fit2)

res <- simulateResiduals(fit2)
plot(res)

DHARMa::testQuantiles(res)
DHARMa::testDispersion(res)

DHARMa::testZeroInflation(res)
}
\author{
Andreas Ettner, Florian Hartig
}
\keyword{datasets}
