% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{titanic}
\alias{titanic}
\title{Titanic data}
\format{
A 'data.frame':	1310 obs. of  14 variables:
\describe{
  \item{pclass}{Passenger class (int)}
  \item{survived}{Did the passenger survive (int, 1 = survived)}
  \item{name}{Pasenger name (character)}
  \item{sex}{Passenger sex (character)}
  \item{age}{Passenger age (numeric)}
  \item{sibsp}{number of siblings / spouses aboard the Titanic (int)}
  \item{parch}{number of parents / children aboard the Titanic (int)}
  \item{ticket}{Ticket number (character)}
  \item{fare}{Passenger fare (numeric)}
  \item{cabin}{cabin number (character)}
  \item{embarked}{Port of Embarkation (character)}
  \item{boat}{Passenger rescue boat number (character)}
  \item{body}{Passenger body size (int)}
  \item{home.dest}{Passenger home destination (character)}
}
}
\source{
\url{http://biostat.mc.vanderbilt.edu/wiki/Main/DataSets}
}
\usage{
titanic
}
\description{
Survival of passengers on the MS titanic
}
\examples{
titanic_df <- titanic
head(titanic_df)

str(titanic_df)

# factorize variables -------------------------------------------------
titanic_df$survived = as.factor(titanic_df$survived)
titanic_df$pclass = as.factor(titanic_df$pclass)
titanic_df$sex = as.factor(titanic_df$sex)


# plotting variables -------------------------------------------------
plot( titanic_df$sex , titanic_df$survived,
      ylab = "survived 1=yes/ 0=no", xlab= "sex")

plot( titanic_df$pclass, titanic_df$survived,
      ylab = "passenger class", xlab= "sex")




library(glmmTMB)
# binomial model -------------------------------------------------
fit <-  glmmTMB(survived ~ pclass + sex + age + sibsp ,
                family = binomial , data = titanic_df)
summary(fit)



library(DHARMa)
res <- simulateResiduals(fit)
plot(res)


DHARMa::testQuantiles(res)
DHARMa::testDispersion(res)

DHARMa::testZeroInflation(res)

# adding random effect -------------------------------------------------
fit2 <-  glmmTMB(survived ~ pclass + sex + age + sibsp + (1|name),
                 family = binomial , data = titanic_df)
summary(fit)


res2 <- simulateResiduals(fit2)
plot(res2)
}
\author{
Florian Hartig
}
\keyword{datasets}
