% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{titanic_ml}
\alias{titanic_ml}
\title{Titanic Data for Machine Learning}
\format{
A 'data.frame': 1309 obs. of 14 variables:
\describe{
\item{pclass}{Passenger class (int).}
\item{survived}{Did the passenger survive (int, 1 = survived).}
\item{name}{Pasenger name (character).}
\item{sex}{Passenger sex (character).}
\item{age}{Passenger age (numeric).}
\item{sibsp}{Number of siblings / spouses aboard the Titanic (int).}
\item{parch}{Number of parents / children aboard the Titanic (int).}
\item{ticket}{Ticket number (character).}
\item{fare}{Passenger fare (numeric).}
\item{cabin}{Cabin number (character).}
\item{embarked}{Port of Embarkation (character).}
\item{body}{Passenger body size (int).}
\item{home.dest}{Passenger home destination (character).}
}
Survival status is unknown for 50\% of the passengers and must be predicted.
}
\source{
\url{http://biostat.mc.vanderbilt.edu/wiki/Main/DataSets} \url{https://www.kaggle.com/c/titanic}
#' @example inst/examples/titanic_ml.R
}
\usage{
titanic_ml
}
\description{
Survival of passengers on the MS titanic
}
\author{
Florian Hartig, Maximilian Pichler
}
\keyword{datasets}
