% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_dates.R
\docType{data}
\name{trans_dates}
\alias{trans_dates}
\title{Sea surface temperature seasonal transiition dates}
\format{
1599 rows and 4 columns.

\itemize{
\item EPU: Ecological Production Unit (EPU) where sampling occurred.
\item Time: Year.
\item Var: Variable describing shannon diversity.
\item Value: Value of variable \code{Var}.
}
}
\usage{
trans_dates
}
\description{
Data include transition day of year.
}
\details{
More information available at
\url{https://noaa-edab.github.io/tech-doc.html}.
}
\keyword{datasets}
