% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_gls.R
\name{ts_gls}
\alias{ts_gls}
\title{GLS model selection for time series}
\usage{
ts_gls(data, ...)
}
\arguments{
\item{data}{Input series to be analyzed.}

\item{...}{Other arguments may be passed to the gls.}
}
\value{
gls model object
}
\description{
GLS model selection for time series
}
\examples{
#Generate series

m <- 0.1
x <- 1:30
y <-  m*x + rnorm(30, sd = 0.35)

data <- data.frame(x = x,
                  y = y)

summary(ts_gls(data = data))
}
