% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_all.R
\name{update_all}
\alias{update_all}
\title{update_all is a wrapper for updating all public datasets}
\usage{
update_all()
}
\value{
null
}
\description{
update_all is a wrapper for updating all public datasets
}
\examples{
\dontrun{
update_all()
}
}
