% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{volcanoisland}
\alias{volcanoisland}
\title{Volcano island}
\format{A 'data.frame':	1000 obs. of  21 variables:}
\usage{
volcanoisland
}
\description{
Ecological observatins on a lonely volcano island. See vignette for details
}
\author{
Florian Hartig
}
\keyword{datasets}
