% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcr.R
\docType{data}
\name{wcr}
\alias{wcr}
\title{Warm core rings spun off of gulf stream}
\format{
Data set contains 40 rows and 5 columns

\itemize{
\item Time: Year.
\item Var: Number of warm core rings.
\item Value: Value of variable  \code{Var}.
\item Units: Unit of Variable  \code{Var}.
\item EPU: Ecological Production Unit (EPU) where sampling occurred.
}
}
\usage{
wcr
}
\description{
Data include timeseires of stomach fullness index for 22 species sampled in the NEFSC fall bottom trawl survey.
}
\details{
More information including processing and indicator derivation steps are available at
\url{https://noaa-edab.github.io/tech-doc/warm-core-rings.html}.
}
\keyword{datasets}
