% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wea_landings_rev.R
\docType{data}
\name{wea_landings_rev}
\alias{wea_landings_rev}
\title{Fisheries max and min landings and revenue from wind energy areas (WEAs)}
\format{
Data set contains 606 rows and 5 columns

\itemize{
\item GARFO and ASMFC Managed Species: Managed Fishery.
\item perc_landings_max: Max Landings Percent from WEA.
\item perc_revenue_max: Max Revenue Percent from WEA.
\item perc_landings_min: Minimum Landings Percent from WEA.
\item perc_revenue_min: Minimum Revenue Percent from WEA.
}
}
\usage{
wea_landings_rev
}
\description{
Data include Top ten species Landings and Revenue from Wind Energy Areas.
}
\details{
More information including processing and indicator derivation steps are available at
\url{https://noaa-edab.github.io/tech-doc/fisheries-revenue-in-wind-development-areas.htm}.
}
\keyword{datasets}
