% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind.R
\docType{data}
\name{wind}
\alias{wind}
\title{NCEP NARR Wind}
\format{Data set inclues 2385 rows and 5 columns

\itemize{
\item Var: Specifies variable and season of interest.
\item Value: Value of variable \code{Var}.
\item Units: Units of variable \code{Var}.
\item EPU: Ecological Production Unit (EPU) from which data were collected.
\item Time: Year.
}}
\source{
\url{https://www.esrl.noaa.gov/psd/data/gridded/data.narr.monolevel.html}
}
\usage{
wind
}
\description{
These time series include mean surface wind speed and direction ("uwnd" and "vwnd" respectively),
in the Northeast Large Marine Ecosystem.
}
\details{
Indicator for total wind speed is calculated as TWS = sqrt(u^2 + v^2), where u = surface wind speed
and v = surface wind direction. Seasons are defined as: Fall = October, November, December;
Winter = January, February, March; Spring = April, May, June; Summer = July, August, September.
}
\keyword{datasets}
