% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind_dev_speed.R
\docType{data}
\name{wind_dev_speed}
\alias{wind_dev_speed}
\title{Wind energy development speed}
\format{
Data set contains 9 rows and 4 columns

\itemize{
\item Time: Year.
\item Var: Cumulative wind lease area.
\item Value: Value of variable  \code{Var}.
\item EPU: Ecological Production Unit (EPU) where sampling occurred.
}
}
\usage{
wind_dev_speed
}
\description{
Data include the planned wind development areas on the Northeast US shelf.
}
\details{
More information including processing and indicator derivation steps are available at
\url{https://noaa-edab.github.io/tech-doc/wind-energy-delvelopment-speed.html}.
}
\keyword{datasets}
