% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind_occupancy.R
\docType{data}
\name{wind_occupancy}
\alias{wind_occupancy}
\title{Wind lease area and habitat occupancy model overlap}
\format{Data set contains 50 rows and 5 columns

\itemize{
\item Season: Fall or Spring.
\item Species: Top 5 species.
\item Mean: Mean.
\item Trend: Trend of species in or out of wind lease area.
\item Area: Existing or proposed.
}}
\usage{
wind_occupancy
}
\description{
Data include a list of species most likely to be found in existing and propsed wind lease areas.
}
\details{
More information including processing and indicator derivation steps are available at
\url{https://noaa-edab.github.io/tech-doc/wind-lease-areas-and-habitat-occupancy-overlap.html}.
}
\keyword{datasets}
