% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind_port.R
\docType{data}
\name{wind_port}
\alias{wind_port}
\title{Community Port Landings and Revenue from Wind Energy Areas (WEAs)}
\format{
Data set contains 330 rows and 5 columns

\itemize{
\item City: Fishing Community.
\item State: State that fishing community is located.
\item Var: Revenue percenatage and social vulnerability score.
\item Value: Value: Value of variable  \code{Var}.
\item EPU: Ecological Production Unit (EPU) where fishing community is located.
}
}
\usage{
wind_port
}
\description{
Data include important fishing communities and the percent landing and revenue from wind energy areas.
Data also include social vulnerability scores for the fishing communities.
}
\details{
More information including processing and indicator derivation steps are available at
\url{https://noaa-edab.github.io/tech-doc/wea-fishing-port-landings.html}.
}
\keyword{datasets}
