% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind_revenue.R
\docType{data}
\name{wind_revenue}
\alias{wind_revenue}
\title{Fisheries revenue from wind lease areas}
\format{
Data set contains 126 rows and 5 columns

\itemize{
\item Time: Year.
\item Var: Top 5 Species Revenue from Wind Development Areas.
\item Value: Value of variable  \code{Var}.
\item Units: Unit of Variable  \code{Var}.
\item EPU: Ecological Production Unit (EPU) where sampling occurred.
}
}
\usage{
wind_revenue
}
\description{
Data include Top 5 Species Revenue from Wind Development Areas
}
\details{
More information including processing and indicator derivation steps are available at
\url{https://noaa-edab.github.io/tech-doc/fisheries-revenue-in-wind-development-areas.html}.
}
\keyword{datasets}
