% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wine}
\alias{wine}
\title{Wine}
\format{
An object of class \code{data.frame} with 1599 rows and 12 columns.
}
\usage{
wine
}
\description{
Wine data from \url{https://archive.ics.uci.edu/ml/datasets/wine+quality}.
}
\details{
The dataset is a collection of wines of different qualities. The wines are described by physochemical variables.

\describe{
\item{quality}{response variable, ranges between 1 and 8}
}
}
\author{
Maximilian Pichler
}
\keyword{datasets}
