% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearly_hw.R
\docType{data}
\name{yearly_hw}
\alias{yearly_hw}
\title{Marine heatwaves within 2019}
\format{Data set contains 1029 rows and 5 columns

\itemize{
\item Time: Date.
\item Var: Marine heatwave value
\item Value: Value of variable  \code{Var}.
\item Units: Unit of Variable  \code{Var}.
\item EPU: Ecological Production Unit (EPU) where sampling occurred.
}}
\usage{
yearly_hw
}
\description{
Data include daily heatwave anomoly for 2019.
}
\details{
More information including processing and indicator derivation steps are available at
\url{https://noaa-edab.github.io/tech-doc/marine_heatwave.html}.
}
\keyword{datasets}
