% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoo_community.R, R/zoo_diversity.R
\docType{data}
\name{zoo_diversity}
\alias{zoo_diversity}
\title{Zooplankton Communities}
\format{
284 rows and 5 columns.

\itemize{
\item Var: Principal Components
\item Value: Value of variable \code{Var}.
\item Units: Units of variable \code{Var}.
\item EPU: Ecological Production Units where data originated.
\item Time: Year.
}

129 rows and 5 columns.

\itemize{
\item Var: Shannon diversity.
\item Value: Value of variable \code{Var}.
\item Units: Units of variable \code{Var}.
\item EPU: Ecological Production Units where data originated.
\item Time: Year.
}
}
\usage{
zoo_diversity

zoo_diversity
}
\description{
EcoMon Zooplankton Principal Components: Zooplankton community composition on the Northeast US Continental Shelf

Estimated zooplankton diversity on the Northeast US Continental Shelf.
}
\details{
These data represent an ecosystem-level look at zooplankton community composition using the dimension reduction technique
of principal component analysis.

These data were derived from bimonthly Ecosystem Monitoring (EcoMon) cruises throughout the Northeast US Continental Shelf,
and represent a different estimation method than what is found in the zooplankton anomaly data set. Specifically, abundance
anomalies were interpolated following an ordinary kriging approach by season across the Northeast Shelf Ecosystem. Exact methods
used to derived these data are available at \url{https://noaa-edab.github.io/ECSA/}.
}
\keyword{datasets}
