% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoo_sli_anom.R
\docType{data}
\name{zoo_sli_anom}
\alias{zoo_sli_anom}
\title{Zooplankton small large index}
\format{
344 rows and 5 columns.

\itemize{
\item Var: Shannon diversity.
\item Value: Value of variable \code{Var}.
\item Units: Units of variable \code{Var}.
\item EPU: Ecological Production Units where data originated.
\item Time: Year.
}
}
\usage{
zoo_sli_anom
}
\description{
Estimated zooplankton small large index on the Northeast US Continental Shelf.
}
\details{
More information including processing and indicator derivation steps are available at
\url{https://noaa-edab.github.io/tech-doc/zooabund.html}.
}
\keyword{datasets}
