% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoo_strat_abun.R
\docType{data}
\name{zoo_strat_abun}
\alias{zoo_strat_abun}
\title{Zooplankton small large absolute values}
\format{
516 rows and 5 columns.

\itemize{
\item Var: Shannon diversity.
\item Value: Value of variable \code{Var}.
\item Units: Units of variable \code{Var}.
\item EPU: Ecological Production Units where data originated.
\item Time: Year.
}
}
\usage{
zoo_strat_abun
}
\description{
Estimated zooplankton small large numbers on the Northeast US Continental Shelf.
}
\details{
More information including processing and indicator derivation steps are available at
\url{https://noaa-edab.github.io/tech-doc/zooabund.html}.
}
\keyword{datasets}
