% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{arthritis}
\alias{arthritis}
\title{Arthritis}
\format{
A 'data.frame':	84 obs. of  5 variables:
\describe{
  \item{ID}{patient ID.}
  \item{Treatment}{factor indicating treatment (Placebo, Treated).}
  \item{Sex}{factor indicating sex (Female, Male).}
  \item{Age}{age of patient.}
  \item{Improved}{ordered factor indicatinf treatment outcome (None, Some, Marked).}
}
}
\source{
\url{https://www.picostat.com/dataset/r-dataset-package-vcd-arthritis}
}
\usage{
arthritis
}
\description{
Arthritis Treatment Data
}
\examples{
arthritis_df <- arthritis

arthritis_df$Treatment = as.factor(arthritis_df$Treatment)
arthritis_df$Sex = as.factor(arthritis_df$Sex)
arthritis_df$Improved = as.factor(arthritis_df$Improved)


str(arthritis_df)
plot(arthritis_df$Improved , c(1:84))
plot(arthritis_df$Treatment , c(1:84))


library(glmmTMB)
# binomial model -------------------------------------------------
fit <-  glmmTMB(Treatment ~ Improved + Age + Sex, family = binomial, data = arthritis_df)
summary(fit)



library(DHARMa)
res <- simulateResiduals(fit)
plot(res)


DHARMa::testDispersion(res)

DHARMa::testOutliers(res)
}
\author{
Florian Hartig
}
\keyword{datasets}
