% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{arthritis}
\alias{arthritis}
\title{Arthritis}
\format{
A 'data.frame':	84 obs. of  5 variables:
\describe{
  \item{ID}{patient ID.}
  \item{Treatment}{factor indicating treatment (Placebo, Treated).}
  \item{Sex}{factor indicating sex (Female, Male).}
  \item{Age}{age of patient.}
  \item{Improved}{ordered factor indicatinf treatment outcome (None, Some, Marked).}
}
}
\source{
\url{https://www.picostat.com/dataset/r-dataset-package-vcd-arthritis}
}
\usage{
arthritis
}
\description{
Arthritis Treatment Data
}
\author{
Florian Hartig
}
\keyword{datasets}
