% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{barbetData}
\alias{barbetData}
\title{barbetData}
\format{
A 'data.frame': 33 obs. of 8 variables:
\describe{
 \item{wing}{ (numeric).}
 \item{Lnalt}{Altitude at which a species is found (numeric).}
 \item{patch}{ (numeric).}
 \item{colour}{ (numeric).}
 \item{Frequency}{ (numeric).}
 \item{Length}{ (numeric).}
 \item{Lnote}{Length of the note in the bird's respective song (numeric).}
}
}
\source{
\url{http://www.phytools.org/Cordoba2017/data/Barbetdata.csv}
}
\usage{
barbetData
}
\description{
This data set comes from \url{http://www.phytools.org/Cordoba2017} and is about fictional birds and their attributes.
}
\details{
This is the data set belonging to the phylogenetic tree "barbetTree".
}
\examples{
library(ape)
library(effects)
library(lme4)


dat = read.csv(url("http://www.phytools.org/Cordoba2017/data/Barbetdata.csv"),
               header = TRUE, row.names = 1)
tree = read.nexus(url("http://www.phytools.org/Cordoba2017/data/BarbetTree.nex"))

dat$species = as.factor(row.names(dat))
plot(tree)

tree$tip.label \%in\% rownames(dat)


model = lm(log(Lnote) ~ Lnalt + I(Lnalt^.5) + I(Lnalt^2), data = dat)
summary(model)
plot(model)

#plot(allEffects(model, partial.residuals = T), selection = 1)
plot(allEffects(model, partial.residuals = T))

res = simulateResiduals(model)
plot(res)
testDispersion(res)

}
\author{
Florian Hartig
}
\keyword{datasets}
