% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bees}
\alias{bees}
\title{Bee data}
\format{
An object of class \code{data.frame} with 72 rows and 1 columns.
}
\usage{
bees
}
\description{
Bee data
}
\author{
Florian Hartig
}
\keyword{datasets}
