% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{birdabundance}
\alias{birdabundance}
\title{Birdabundance}
\format{
A 'data.frame':	56 obs. of  8 variables:
\describe{
  \item{SITE}{site ID}
  \item{ABUND}{Bird abundance}
  \item{AREA}{Area}
  \item{DIST}{Distance to the next forest fragment}
  \item{LDIST}{Distance to the next large forest fragment}
  \item{YR.ISOL}{Years since the site was isolated / fragmented}
  \item{GRAZE}{Grazing intensity in the surrounding}
  \item{ALT}{Altitude}
}
}
\source{
\url{http://biometry.github.io/APES/}
}
\usage{
birdabundance
}
\description{
Abundance of birds in forest fragments in Australia
}
\author{
Florian Hartig
}
\keyword{datasets}
