% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calanus_stage.R
\docType{data}
\name{calanus_stage}
\alias{calanus_stage}
\title{Calanus Stage}
\format{
A data set containing 1006 rows and 5 columns.

\itemize{
\item Year: Time
\item EPU: Ecological Production Unit (EPU) where data originated. Here "MAB" refers to the Mid-Atlantic Bight.
\item Value: Value of variable.
\item Var: calanus stage
\item Units: Number of calanus/meter cubed.

}
}
\usage{
calanus_stage
}
\description{
Time series of calanus abundance by life stage.
}
\details{
More information available at \url{https://noaa-edab.github.io/tech-doc/calanus_stage.html}
}
\keyword{datasets}
