% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_bay_sal.R
\docType{data}
\name{ch_bay_sal}
\alias{ch_bay_sal}
\title{Chesapeake Bay 2018 - 2019 salinity with historical trends for comapirson}
\format{
A data set containing 2555 rows and 5 columns.

\itemize{
\item Time: Day
\item Value: Value of variable \code{Var}
\item Var: Salinity measurement
\item EPU: Ecological Production Unit (EPU) where data originated. Here "MAB" refers to the Mid-Atlantic Bight.
\item Units: Units of variable \code{Var}.

}
}
\usage{
ch_bay_sal
}
\description{
Chesapeake Bay 2018 - 2019 salinity with historical trends for comapirson
}
\details{
More information available at \url{https://noaa-edab.github.io/tech-doc/chesapeake-bay-salinity.html}
}
\keyword{datasets}
