% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{dataset_flower}
\alias{dataset_flower}
\title{Flower Image Recognition Data}
\usage{
dataset_flower()
}
\value{
\itemize{
 \item train - Images for training
 \item labels - Labels for train split, 0 - 4
 \item test - Images for testing
}
}
\description{
Downsampled (80x80 pixels) images of 5 flower species from \url{https://www.kaggle.com/alxmamaev/flowers-recognition}.
}
\details{
The function will return a list with three elements:
}
\author{
Maximilian Pichler
}
