% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecodata.R
\docType{package}
\name{ecodata-package}
\alias{ecodata}
\alias{ecodata-package}
\title{ecodata: indicator data for reporting ecosystem health}
\description{
ecodata is developed by the Ecosystems Dynamics and Assessment Branch of the Northeast Fisheries Science Center for use in State of the Ecosystem (SOE) reporting.  SOE reports are high-level overviews of ecosystem indicator status and trends occurring on the Northeast Continental Shelf. Unless otherwise stated, data are representative of specific Ecological Production Units (EPUs), referring to the Mid-Atlantic Bight (MAB), Georges Bank (GB), Gulf of Maine (GOM), and Scotian Shelf (SS). SOE reports are developed for US Fishery Management Councils (FMCs), and therefore indicator data for Scotian Shelf  are included when available, but this is not always the case.
}
\details{
\itemize{
The main goals are:
\item To create a centralized repository for indicator data
\item To facilitate creation of annual reports
}

To learn more about using \code{ecodata}, start with the vignette: \code{browseVignettes(package="ecodata")} or click the index link below
}
\seealso{
Useful links:
\itemize{
  \item \url{https://noaa-edab.github.io/ecodata/},\url{https://github.com/NOAA-EDAB/ecodata}
}

}
\author{
\strong{Maintainer}: Brandon Beltz \email{brandon.beltz@noaa.gov} (\href{https://orcid.org/0009-0002-7385-5950}{ORCID})

Authors:
\itemize{
  \item Brandon Beltz \email{brandon.beltz@noaa.gov} (\href{https://orcid.org/0009-0002-7385-5950}{ORCID})
  \item Andy Beet \email{andrew.beet@noaa.gov} (\href{https://orcid.org/0000-0001-8270-7090}{ORCID})
  \item Kimberly Bastille \email{kimberly.bastille@noaa.gov}
  \item Sean Hardison
}

}
