% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecodata.R
\name{ecodata}
\alias{ecodata}
\alias{ecodata,missing,missing,missing,missing-method}
\alias{ecodata,character,missing,missing,missing-method}
\alias{ecodata,character,character,missing,missing-method}
\alias{ecodata,character,character,character,missing-method}
\alias{ecodata,character,character,character,character-method}
\title{Load and/or save a dataset}
\usage{
ecodata(subdiscipline, topic, dataset, path)

\S4method{ecodata}{missing,missing,missing,missing}()

\S4method{ecodata}{character,missing,missing,missing}(subdiscipline)

\S4method{ecodata}{character,character,missing,missing}(subdiscipline, topic)

\S4method{ecodata}{character,character,character,missing}(subdiscipline, topic, dataset)

\S4method{ecodata}{character,character,character,character}(subdiscipline, topic, dataset, path)
}
\arguments{
\item{subdiscipline}{A string containing the name of an ecological discipline}

\item{topic}{A string containing the name of an analytical topic}

\item{dataset}{A string containing the name of a dataset}

\item{path}{A string containing the path to save the data file}
}
\value{
An R object
}
\description{
Load a dataset directly to an R object, save it as file, or both.
}
\details{
This function is used to load one of the static or dynamic datasets provided
by the package. It can be either loaded directly into R as an object, or it
can be saved to a file by specifying the path parameter, or both.
}
\examples{
# Try running the ecodata() without any arguments. You'll get some information
# about valid options for the first parameter.
data <- ecodata::ecodata()

# Note that the function returns NULL at this point
print(data)

# Repeat this process until we get data
data <- ecodata::ecodata("Quantitative Ecology")

# Instead of NULl this time, we get the information that was printed out. It may
# or many not be useful to store in a variable for later reference without having
# to run the function again.
head(data)

# Repeat this process until we get data
data <- ecodata::ecodata("Quantitative Ecology", "Distributions")
head(data)
data <- ecodata::ecodata("Quantitative Ecology", "Distributions", "Normal Random Sets")
head(data)

# This particular data set is actually dynamically generated. This means that
# running it again should produce different results.
data <- ecodata::ecodata("Quantitative Ecology", "Distributions", "Normal Random Sets")
head(data)
}
