% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EcoData.R
\docType{package}
\name{EcoData}
\alias{EcoData}
\title{EcoData - A collection of ecological data sets for teaching}
\description{
A collection of ecological data sets for teaching
}
\details{
See index / vignette for details
}
\examples{
vignette("EcoData", package="EcoData")
}
\author{
Florian Hartig
}
