% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecodata.R
\name{ecodata_cite_table}
\alias{ecodata_cite_table}
\title{`ecodata_cite_table(data)`
Get a table of citation information for every variable in the given ecodata data frame}
\usage{
ecodata_cite_table(data)
}
\arguments{
\item{data}{Data frame from `get_ecodata()` that includes units information in the meta data}
}
\value{
Returns a flextable with a row for each variable in the given data frame, with a column for how to cite each variable
}
\description{
`ecodata_cite_table(data)`
Get a table of citation information for every variable in the given ecodata data frame
}
\seealso{
[ecodata_cite()]
}
