% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{elephant}
\alias{elephant}
\title{Elephant}
\format{
An object of class \code{list} of length 2.
}
\usage{
elephant
}
\description{
Species distribution and prediction data for the African elephant (Loxodonta).
}
\details{
This dataset consists of two datasets, which are presented as a list
\enumerate{
\item elephant$occurenceData contains presence / absence data as well as bioclim variables (environmental predictors) for the African elephant
\item elephant$predictionData data with environmental predictors for spatial predictions
}
}
\author{
Maximilian Pichler, Florian Hartig
}
\keyword{datasets}
