% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_density.R
\docType{data}
\name{energy_density}
\alias{energy_density}
\title{Forage fish energy density}
\format{
A data set containing 72 rows and 6 columns.

\itemize{
\item Species: Species included in analysis.
\item Year: Time.
\item Season: Survey season.
\item N: Number of samples included in the analysis.
\item Energy.Density_Mean: Mean forage fish energy density.
\item Energy.Density_SD: Standard Deviation of forage fish energy density.
}
}
\usage{
energy_density
}
\description{
Forage fish energy density comparison between studies in the 80s and 90s.
}
\details{
More information available at \url{https://noaa-edab.github.io/tech-doc/forage-fish-energy-density.html}
}
\keyword{datasets}
