% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_density.R
\docType{data}
\name{energy_density}
\alias{energy_density}
\title{Forage fish energy density}
\format{A data set containing 8 rows and 13 columns.

\itemize{
\item Species: Species included in analysis.
\item Mean.ED..SD.: Mean and standard deviation of spring 2017.
\item n: Number of samples taken in spring 2017.
\item Mean.ED..SD..1.: Mean and standard deviation of fall 2017.
\item n..1: Number of samples taken in fall 2017.
\item Mean.ED..SD..2: Mean and standard deviation of spring 2018.
\item n.2: Number of samples taken in spring 2018.
\item Mean.ED..SD..3.: Mean and standard deviation of fall 2018.
\item X: Number of samples taken in fall 2018.
\item Mean.ED..SD..4.: Mean and standard deviation of total.
\item n.3: Number of samples taken in total.
\item Mean.ED: Mean found in Stiemle and Terranove 1985.
\item Mean.ED..SD..5: Mean and standard deviation found in Lawson et al. 1998.

}}
\usage{
energy_density
}
\description{
Forage fish energy density
}
\details{
More information available at \url{https://noaa-edab.github.io/tech-doc/forage-fish-energy-density.html}
}
\keyword{datasets}
