% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_n.R
\docType{data}
\name{exp_n}
\alias{exp_n}
\title{Expected Number of Species per 1000 Individuals}
\format{
909 rows and 6 columns.

\itemize{
\item Value: Value of variable \code{Var}.
\item Units: Units of variable \code{Var}.
\item Time: Year.
\item EPU: Ecological Production Unit (EPU) where sampling occurred.
\item Var: Vessel and Survey season.
}
}
\usage{
exp_n
}
\description{
Data include mean and standard deviation of expected number of species caught
on the NEFSC bottom trawl survey per 1000 individuals collected for both Fall and Spring.
}
\details{
Methods used to calculate data are available at \url{https://noaa-edab.github.io/tech-doc/expected-number-of-species.html}.
}
\keyword{datasets}
