% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecodata.R
\name{get_ecodata_variable_fred}
\alias{get_ecodata_variable_fred}
\title{`get_ecodata_variable_fred()`
Downloads data from FRED for a given variable code or URL}
\usage{
get_ecodata_variable_fred(varcode, varname = NULL, recessions = FALSE)
}
\arguments{
\item{varcode}{String that contains the variable code or URL to the data}

\item{varname}{Optional, string for the variable name}

\item{recessions}{Optional, logical for whether to include an NBER recession dummy variable. Default = FALSE.}
}
\value{
Data frame time series with two variables, Date and the variable requested. The data frame will also include all relevant meta data describing the data and citing its source.
}
\description{
`get_ecodata_variable_fred()`
Downloads data from FRED for a given variable code or URL
}
