% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecodata.R
\name{get_state_abbr}
\alias{get_state_abbr}
\title{Get the two-letter abbreviation for a state}
\usage{
get_state_abbr(state)
}
\arguments{
\item{state}{String that is a name of a state or the FIPS code}
}
\value{
String that is the two-letter abbreviation for the state
}
\description{
Get the two-letter abbreviation for a state, given the name of the state or the FIPS code
}
