% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecodata.R
\name{get_state_name}
\alias{get_state_name}
\title{Get name of a U.S. State}
\usage{
get_state_name(state)
}
\arguments{
\item{state}{String that is a FIPS code or two-letter abbreviation}
}
\value{
String that is the name of the state
}
\description{
Get the state name for the given abbreviation or FIPS code
}
