% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecodata.R
\name{get_varcode_url_worldbank}
\alias{get_varcode_url_worldbank}
\title{`get_varcode_url_worldbank()`
Extract a variable code from a Workbank URL}
\usage{
get_varcode_url_worldbank(url)
}
\arguments{
\item{url}{String that is a URL to a World Bank data series}
}
\value{
String that is the variable code
}
\description{
`get_varcode_url_worldbank()`
Extract a variable code from a Workbank URL
}
