% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gpa}
\alias{gpa}
\title{GPA data}
\format{
An object of class \code{data.frame} with 1200 rows and 10 columns.
}
\usage{
gpa
}
\description{
The GPA (college grade point average) data is a longitudinal data set in which 200 college students and their GPA have been followed 6 consecutive semesters.
}
\details{
Variables:
\itemize{
\item student - student ID
\item occass - time
\item gpa - the students gpa
\item job - how much the student worked
\item sex - sex
\item highgpa - high school gpa
\item admitted - whether a student has been admitted to the university of their choice. Since not every student applies to a university, this variable has many missing values.
}

For convenience, there are additionally the variables
\itemize{
\item year
\item semester
\item occasion
}

which are essentially a re-coding of ocasss
}
\examples{
require(EcoData)
fit <- lm(gpa ~ occasion, data = gpa)
summary(fit)
}
\author{
Florian Hartig
}
\keyword{datasets}
