% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gray_seal.R
\docType{data}
\name{grayseal}
\alias{grayseal}
\title{Gray Seal Bycatch}
\format{
A time series with 130 rows and 4 columns.

\itemize{
\item Time: Year
\item Var: Bycath variable
\item Value: Value of variable
\item EPU: Shelfwide indicator
}
}
\usage{
grayseal
}
\description{
Data include estimated Gray Seal bycatch on the Northeast Shelf.
}
\details{
Methods used to calculate data are available at \url{https://noaa-edab.github.io/tech-doc/harbor-porpoise-and-gray-seal-bycatch.html}.
}
\keyword{datasets}
