% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harborporpoise.R
\docType{data}
\name{harborporpoise}
\alias{harborporpoise}
\title{Harbor porpoise bycatch}
\format{
Data set includes 120 rows and 5 columns.

\itemize{
\item Var: Specifies variable of interest, including bycatch estimates (five year moving average) with
coefficient of variation and confidence intervals. Also included is the potential biological removal (PBR) for
harbor porpoise.
\item Value: Value of variable \code{Var}.
\item Time: Year.
\item Units: Units of variable \code{Var}.
\item EPU: Ecological Production Unit (EPU) where sampling occurred.
}
}
\usage{
harborporpoise
}
\description{
Data included here show the 5 year moving average for harbor porpoise bycatch estimates and potential
biological removal (PBR) levels in the Northeast United States. These data were extracted from NEFSC Center
Reference Documents and marine mammal stock assessments.
}
\details{
Links to source data, further data definitions, and extended methodology are available at
\url{https://noaa-edab.github.io/tech-doc/harbor-porpoise-bycatch.html}
}
\keyword{datasets}
