% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harborporpoise.R
\docType{data}
\name{harborporpoise}
\alias{harborporpoise}
\title{Harbor porpoise bycatch}
\format{
160 rows and 4 columns.

\itemize{
\item Time: Year
\item Var: Bycatch variable
\item Value: Value of variable
\item EPU: Shelfwide indicator
}
}
\usage{
harborporpoise
}
\description{
Data included here show the 5 year moving average for harbor porpoise bycatch estimates and potential
biological removal (PBR) levels in the Northeast United States. These data were extracted from NEFSC Center
Reference Documents and marine mammal stock assessments.
}
\details{
Links to source data, further data definitions, and extended methodology are available at
\url{https://noaa-edab.github.io/tech-doc/harbor-porpoise-bycatch.html}
}
\keyword{datasets}
