% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hp_density.R
\docType{data}
\name{hp_density}
\alias{hp_density}
\title{Harbor porpoise density estimates}
\format{
6244 rows and 5 columns

\itemize{
\item Latitude: Decimal latitude.
\item Longitude: Decimal longitude.
\item Value: Density (Animals/km^2)
\item Bin: Density bins for plotting.
\item Season: Season; \code{fall} or \code{spring}.
}
}
\source{
Data were sourced from \url{https://www.nefsc.noaa.gov/AMAPPSviewer/} on 12/14/2018.
}
\usage{
hp_density
}
\description{
Data include seasonal density estimates for harbor porpoise on the Northeast US Continental Shelf.
}
\details{
From the AMAPPS Marine Mammal Model Viewer Site: "Distribution and abundance of wildlife is largely driven by physical and biological environmental
factors, including climate, habitat characteristics and prey distribution (Ainley et al. 2005). To
account for this, these spatially- and temporally-explicit density maps were based on animal
density - environmental generalized additive statistical models that were fit to visual shipboard
and aerial survey line-transect data, associated survey conditions, animal group characteristics,
spatially- and temporally-explicit static and dynamic environmental data, and species-specific
availability bias correction factors. For a more completed description see Palka et al. 2017."
}
\references{
Ainley D.G., L. Spear, C.T. Tynan, J.A. Barth, S.D. Pierce, R. Glenn Ford, T. Cowles. 2005. Physical
and biological variables affecting seabird distributions during the upwelling season of the northern
California Current. Deep Sea Research II 52(2):123–143.

Palka, D.L., S. Chavez-Rosales, E. Josephson, D. Cholewiak, H.L. Haas, L. Garrison, M. Jones, D. Sigourney,
G. Waring (retired), M. Jech, E. Broughton, M. Soldevilla, G. Davis, A. DeAngelis, C.R. Sasso, M.V.Winton,
R.J. Smolowitz, G. Fay, E. LaBrecque, J.B. Leiness, Dettloff, M. Warden, K. Murray, and C. Orphanides.
2017. Atlantic Marine Assessment Program for Protected Species: 2010-2014. US Dept. of the Interior,
Bureau of Ocean Energy Management, Atlantic OCS Region, Washington, DC. OCS Study BOEM 2017-071. 211 pp
}
\keyword{datasets}
