% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hudson_river_flow.R
\docType{data}
\name{hudson_river_flow}
\alias{hudson_river_flow}
\title{Hudson River Flow}
\format{
40 rows and 4 columns

\itemize{
\item Time: Year.
\item Var: Flowrate of Hudson River
\item Value: Value of variable \code{Var}.
\item EPU: Ecological Production Unit (EPU) where sampling occurred.
}
}
\usage{
hudson_river_flow
}
\description{
Flow from Hudson River into the Atlantic.
}
\details{
More information about this data set may be found at \url{https://noaa-edab.github.io/tech-doc/hudson-river-flow.html}.
}
\keyword{datasets}
