% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecodata.R
\name{is_valid_url}
\alias{is_valid_url}
\title{Test if a string is a valid URL}
\usage{
is_valid_url(url)
}
\arguments{
\item{url}{String, that is maybe a URL}
}
\value{
Logical, whether or not the string was actually a valid URL
}
\description{
Test whether a string is a valid URL to an actual web address
}
