% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_line_survey.R
\docType{data}
\name{long_line_survey}
\alias{long_line_survey}
\title{longline survey}
\format{
A data set containing 96 rows and 8 columns.

\itemize
\item Cruise: Latitude
\item Station: Longitude
\item Year: Value
\item Common_name: ??
\item Deceg_beglat_set: ???
\item Deceg_beglon_set: ???
\item categ: ???
}
\usage{
long_line_survey
}
\description{
What is it?
}
\details{
More information available at \url link to techdoc
}
\keyword{datasets}
