% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_line_survey_sf.R
\docType{data}
\name{longlinesurvey_sf}
\alias{longlinesurvey_sf}
\title{Long-Line Survey Strata Shapefile}
\format{
Data contain 6 rows and 6 columns.
\itemize{
\item OBJECTID: Polygon Identifyer.
\item STR2: Abbreviated strata code.
\item STRATUMA: Strata Code.
\item NUMPICKS:
\item CALCULATED: Calculated strata size.
\item Shape_Leng: Length of strata shape.
\item Shape_Area: Area of strata shape.
\item geometry: Shape discription.
}
}
\usage{
longlinesurvey_sf
}
\description{
Gulf of Maine long-line survey strata.
}
\details{
More information including processing and indicator derivation steps are available at
\url{https://noaa-edab.github.io/tech-doc.html}.
}
\keyword{datasets}
