% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lps_sharks.R
\docType{data}
\name{lps_sharks}
\alias{lps_sharks}
\title{lps_sharks ???}
\format{
A data set containing 90 rows and 5 columns.

\itemize
\item Time: ???
\item Var: ???
\item Value: ???
\item EPU: ??
\item Units: ???
}
\usage{
lps_sharks
}
\description{
What is it?
}
\details{
More information available at \url link to techdoc
}
\keyword{datasets}
