% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{marmots}
\alias{marmots}
\title{Marmot data}
\format{
An object of class \code{data.frame} with 896 rows and 55 columns.
}
\usage{
marmots
}
\description{
Marmot data
}
\author{
Florian Hartig
}
\keyword{datasets}
