% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{melanoma}
\alias{melanoma}
\title{Melanoma}
\format{
A 'data.frame':	205 obs. of  7 variables:
\describe{
  \item{DC}{death/censoring indicator; 1 = died from melanoma, 2 =still alive and 3 = died from causes unrelated to their melanoma}
  \item{time}{Survival time in days since the operation, possibly censored}
  \item{level}{level of invasion, 0, 1 or 2}
  \item{ici}{inflammatory cell infiltration (ICI), 0, 1, 2 or 3}
  \item{ecel}{presence of epithelioid cells, 1=no, 2=yes}
  \item{ulcer}{Indicator of ulceration; 1=present, 0=absent.}
  \item{thickness}{Tumour thickness in mm}
  \item{sex}{The patients sex; 1=male, 0=female.}
  \item{age}{Age in years at the time pf the operation.}
}
}
\source{
\url{http://publicifsv.sund.ku.dk/~linearpredictors/?page=datasets&dataset=Melanoma}
}
\usage{
melanoma
}
\description{
Survival from Malignant Melanoma
}
\examples{
melanoma_df <- melanoma
head(melanoma_df)

str(melanoma_df)

# factorize variables -------------------------------------------------
melanoma_df$DC = as.factor(melanoma_df$DC)
melanoma_df$level = as.factor(melanoma_df$level)
melanoma_df$ici = as.factor(melanoma_df$ici)
melanoma_df$ecel = as.factor(melanoma_df$ecel)
melanoma_df$ulcer = as.factor(melanoma_df$ulcer)
melanoma_df$sex = as.factor(melanoma_df$sex)


# plotting variables -------------------------------------------------
plot(melanoma_df$DC, 1:nrow(melanoma_df))
plot(melanoma_df$time, melanoma_df$thickness)


melanoma_df$time = scale(melanoma_df$time)
melanoma_df$thickness = scale(melanoma_df$thickness)



library(glmmTMB)
# binomial model -------------------------------------------------
fit <-  glmmTMB(ecel ~  level + ici + DC + ulcer + thickness
                + (1|sex), family = binomial , data = melanoma_df)
summary(fit)



library(DHARMa)
res <- simulateResiduals(fit)
plot(res)


DHARMa::testQuantiles(res)
DHARMa::testDispersion(res)

DHARMa::testZeroInflation(res)

}
\author{
Florian Hartig
}
\keyword{datasets}
