% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{melanoma}
\alias{melanoma}
\title{Melanoma}
\format{
A 'data.frame':	205 obs. of  7 variables:
\describe{
  \item{time}{Survival time in days since the operation, possibly censored}
  \item{status}{The patients status at the end of the study. 1 indicates that they had died from melanoma, 2 indicates that they were still alive and 3 indicates that they had died from causes unrelated to their melanoma}
  \item{sex}{The patients sex; 1=male, 0=female.}
  \item{age}{Age in years at the time pf the operation.}
  \item{year}{Year of operation.}
  \item{thickness}{Tumour thickness in mm}
  \item{ulcer}{Indicator of ulceration; 1=present, 0=absent.}
}
}
\source{
\url{http://publicifsv.sund.ku.dk/~linearpredictors/?page=datasets&dataset=Melanoma}
}
\usage{
melanoma
}
\description{
Survival from Malignant Melanoma
}
\author{
Florian Hartig
}
\keyword{datasets}
