% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nasa}
\alias{nasa}
\title{Nasa asteroids}
\format{
An object of class \code{data.frame} with 4687 rows and 40 columns.
}
\usage{
nasa
}
\description{
A collection about asteroids and their characteristics from \url{https://www.kaggle.com/shrutimehta/nasa-asteroids-classification}. The aim is to predict whether the asteroids are hazardous or not. 
\describe{
\item{Hazardous}{response variable, 0 or 1} 
}
}
\author{
Maximilian Pichler
}
\keyword{datasets}
