% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nitrofen}
\alias{nitrofen}
\title{Nitrofen}
\format{
A 'data.frame':	50 obs. of  5 variables:
\describe{
  \item{conc}{The nitrofen concentration in the solution (mug/litre)}
  \item{brood1}{The number of live offspring in the first brood}
  \item{brood2}{The number of live offspring in the second brood}
  \item{brood3}{The number of live offspring in the thrid brood}
  \item{total}{The total number of live offspring in the first three broods}
}
}
\source{
\url{http://statwww.epfl.ch/davison/BMA/Data4BMA/}
}
\usage{
nitrofen
}
\description{
Toxicity of Nitrofen in Aquatic Systems
}
\author{
Florian Hartig
}
\keyword{datasets}
