% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plantcounts}
\alias{plantcounts}
\title{Plantcounts}
\format{
An object of class \code{data.frame} with 285 rows and 6 columns.
}
\usage{
plantcounts
}
\description{
Number of plant species in a grid of quadrants around Regensburg.
Variables
}
\details{
\describe{
  \item{richness}{plant "species" richness. No. of plant taxa records (species, including some aggregates) in map quadrants (Topographische Karte, "Viertelmesstischblatt"), ca. 1980. from Mergenthaler, O. 1982: Verbreitungsatlas zur Flora von Regensburg. In: Hoppea, Denkschriften der Regensburgischen Botanischen Gesellschaft. Bd. 40, S. V-XII, 1-297. http://digital.bib-bvb.de/view/bvbmets/viewer.0.6.4.jsp?folder_id=0&dvs=1597093892683~24&pid=1537325&locale=de&usePid1=true&usePid2=true}
  \item{tk}{quadrant id (Topographische Karte)}
  \item{area}{quadrant area}
  \item{agrarea}{fraction of agricultural area per quadrant (agricultural landuse classes without forest, from https://doi.pangaea.de/10.1594/PANGAEA.910837)}
}
}
\examples{

plants_sf <- plantcounts
plants_sf$agrarea_scaled <- scale(plants_sf$agrarea)

plants_sf$longitude <- plants_sf$lon
plants_sf$latitude <- plants_sf$lat
library(sf)
plants_sf <- sf::st_as_sf(plants_sf, coords = c('longitude', 'latitude'), crs = st_crs("+proj=longlat +ellps=bessel +towgs84=606,23,413,0,0,0,0 +no_defs"))

library(mapview)
mapview(plants_sf["agrarea"])
mapview(plants_sf["richness"], map.types = "OpenTopoMap")

library(glmmTMB)
# Negative binomial model -------------------------------------------------
fit <-  glmmTMB(richness ~ agrarea_scaled + offset(log(area)), family = nbinom1, data = plants_sf)
summary(fit)

library(DHARMa)
res <- simulateResiduals(fit)
plot(res)
testSpatialAutocorrelation(res, x = plants_sf$lon, y =  plants_sf$lat)

library(gstat)
vario <- gstat::variogram(residuals(res) ~ 1, loc = ~ lon + lat, data = plantcounts, alpha=c(0,45,90,135))
plot(vario)
}
\author{
Lukas Heiland, Florian Hartig
}
\keyword{datasets}
