% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plantPollinator_df}
\alias{plantPollinator_df}
\title{Plant Pollinators}
\format{
A 'data.frame': 20,480 obs. of 18 variables:
\describe{
 \item{crop}{Plant name.}
 \item{insect}{Insect name.}
 \item{type}{Type of plant (arboreous or herbaceous).}
 \item{season}{Flowering season.}
 \item{diameter}{Flower diameter in mm.}
 \item{corolla}{Corolla shape (open, campanulate, or tubular).}
 \item{colour}{Flower colour.}
 \item{nectar}{Nectar (yes or no).}
 \item{b.symstem}{Bloom system (type of pollination: insects, insects / wind, or insects / birds).}
 \item{s.pollination}{Self pollination (yes or no).}
 \item{inflorescence}{Inflorescence (yes, solitary, solitary/pairs, solitary / clusters).}
 \item{composite}{Composite flowers (yes or no).}
 \item{guild}{Five pollinator traits: Guild (bumblebees, butterflies etc.).}
 \item{tongue}{Tongue length in mm.}
 \item{body}{Body size in mm.}
 \item{sociality}{Soziality (yes or no).}
 \item{feeding}{Feeding behaviour (oligolectic, polylectic, or parasitic).}
 \item{interaction}{Interaction between insect and plant (yes or no).}
}
}
\source{
\url{https://besjournals.onlinelibrary.wiley.com/doi/full/10.1111/2041-210X.13329}
}
\usage{
plantPollinator_df
}
\description{
plant pollination by different insects.
This data set is a machine learning version of the original one. Containing variable NAs in the "interaction" variable {plantPollinations$interactions}. Generated with the example code.
}
\examples{
library(EcoData)
plant_poll = plantPollinators

plant_poll = plant_poll
levels(plant_poll$interaction) = c(0, 1)
plant_poll$interaction

str(plant_poll)
summary(plant_poll)

unique(plant_poll$crop)[1:10]
unique(plant_poll$insect)[1:20]

nrow(  plant_poll[( plant_poll$crop \%in\% unique(plant_poll$crop)[1:15] )
      | ( plant_poll$insect \%in\% unique(plant_poll$insect)[1:30] ), ]  )


library(glmmTMB)
# binomial model -------------------------------------------------
fit <-  glmmTMB(interaction ~ season  + crop + nectar + guild + tongue
                + body + sociality, family = binomial ,data = plant_poll)
summary(fit)

library(DHARMa)
res <- simulateResiduals(fit)
plot(res)


DHARMa::testQuantiles(res)
DHARMa::testDispersion(res)

DHARMa::testZeroInflation(res)
}
\author{
Andreas Ettner
}
\keyword{datasets}
