% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_advection.R
\name{plot_advection}
\alias{plot_advection}
\title{plot shelf break advection index}
\usage{
plot_advection(shadedRegion = NULL, report = "MidAtlantic", varName = 3, n = 0)
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10)}

\item{report}{Character string. Which SOE report ("MidAtlantic", "NewEngland")}

\item{varName}{Numeric scalar. Which variable (months) to plot (3 = 'March', 4 = 'April', 5 = 'May', 6 = 'June'). Can be a vector}

\item{n}{Numeric scalar. Number of years used (from most recent year) to estimate short term trend . Default = 0 (No trend calculated)}
}
\value{
ggplot object
}
\description{
plots integrated mass transport through transects along the shelf break
}
