% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bottom_temp.R
\name{plot_bottom_temp}
\alias{plot_bottom_temp}
\title{plot bottom anomaly temperature time series using NEFSC survey data}
\usage{
plot_bottom_temp(shadedRegion = NULL, report = "MidAtlantic", n = 0)
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10)}

\item{report}{Character string. Which SOE report ("MidAtlantic", "NewEngland")}

\item{n}{Numeric scalar. Number of years used (from most recent year) to estimate short term trend . Default = 0 (No trend calculated)}
}
\value{
ggplot object
}
\description{
ecodata::bottom_temp
}
