% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bottom_temp_comp.R
\name{plot_bottom_temp_comp}
\alias{plot_bottom_temp_comp}
\title{plot bottom anomaly temperature time series}
\usage{
plot_bottom_temp_comp(shadedRegion = c(2012, 2022), report = "MidAtlantic")
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10)}

\item{report}{Character string. Which SOE report ("MidAtlantic", "NewEngland")}
}
\value{
ggplot object
}
\description{
plot bottom anomaly temperature time series
}
