% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cold_pool.R
\name{plot_cold_pool}
\alias{plot_cold_pool}
\title{plot cold_pool}
\usage{
plot_cold_pool(shadedRegion = NULL, report = "MidAtlantic")
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10)}

\item{report}{Character string. Which SOE report ("MidAtlantic" only, default)}
}
\value{
ggplot object
}
\description{
Create 3 panel plot of cold pool temperature, extent, and persistence time series.
}
