% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_comdat.R
\name{plot_comdat}
\alias{plot_comdat}
\title{plot comdat}
\usage{
plot_comdat(
  shadedRegion = NULL,
  report = "MidAtlantic",
  varName = "landings",
  plottype = "total",
  NAFOyear = 2019,
  n = 0
)
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10)}

\item{report}{Character string. Which SOE report ("MidAtlantic", "NewEngland")}

\item{varName}{Character string. Which Variable to plot ("landings", "revenue")}

\item{plottype}{Character string. Which plot ("total", "guild")}

\item{NAFOyear}{Numeric value. Year that NAFO landings are separated.
Defaults to 2019, as NAFO data was missing 2019-2021 in 2022.}

\item{n}{Numeric scalar. Number of years used (from most recent year) to estimate short term trend . Default = 0 (No trend calculated)}
}
\value{
ggplot object
}
\description{
Plot time series of commercial landings or revenue for various aggregations.
}
