% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_community_factors.R
\name{plot_community_factors}
\alias{plot_community_factors}
\title{table of community factor indicators}
\usage{
plot_community_factors(
  report = "MidAtlantic",
  varName = "Social",
  plottype = "Commercial",
  n = 10
)
}
\arguments{
\item{report}{Character string. Which SOE report ("MidAtlantic", "NewEngland")}

\item{varName}{Character string. Which variable sets to plot ("Social","Economic","Gentrification")}

\item{plottype}{Character string. Which Fishery to plot ("Commercial","Recreational")}

\item{n}{Numeric scalar. Number of rows to return (top in plottype category port activity or recreational engagement)}
}
\value{
dataframe
}
\description{
returns community indicators with highest engagement or port activity
}
