% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_condition.R
\name{plot_condition}
\alias{plot_condition}
\title{plots Fish Condition}
\usage{
plot_condition(shadedRegion = NULL, report = "MidAtlantic", epu = "MAB")
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10)}

\item{report}{Character string. Which SOE report ("MidAtlantic", "NewEngland")}

\item{epu}{Character string. Which EPU in the report ("GB", "GOM", "MAB")}
}
\value{
ggplot object
}
\description{
plots the condition data set
}
