% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_engagement.R
\name{plot_engagement}
\alias{plot_engagement}
\title{plot engagement}
\usage{
plot_engagement(
  shadedRegion = shadedRegion,
  report = "MidAtlantic",
  varName = "Commercial"
)
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10)}

\item{report}{Character string. Which SOE report ("MidAtlantic", "NewEngland")}

\item{varName}{Character string. Which Fishery to plot ("Commercial","Recreational")}
}
\value{
ggplot object
}
\description{
Plot reliance vs engagement by community, recreational or commercial fishery.
}
