% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_exp_n.R
\name{plot_exp_n}
\alias{plot_exp_n}
\title{plot exp_n}
\usage{
plot_exp_n(
  shadedRegion = shadedRegion,
  report = "MidAtlantic",
  varName = "fall"
)
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10)}

\item{report}{Character string. Which SOE report ("MidAtlantic", "NewEngland")}

\item{varName}{Character string. Which Season to plot ("fall", "spring")}
}
\value{
ggplot object
}
\description{
Plots time series of expected number of species from NEFSC bottom trawl survey.
Plot by season
}
