% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_forage_index.R
\name{plot_forage_index}
\alias{plot_forage_index}
\title{plot forage index}
\usage{
plot_forage_index(
  shadedRegion = NULL,
  report = "MidAtlantic",
  varName = "index"
)
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10)}

\item{report}{Character string. Which SOE report ("MidAtlantic", "NewEngland")}

\item{varName}{Character string. Forage biomass index by region, or coastwide center of gravity ("index", "cog")}
}
\value{
ggplot object
}
\description{
Description should be here. This needs to be reworked to uncouple GB and GOM
}
