% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_forage_index.R
\name{plot_forage_index}
\alias{plot_forage_index}
\title{plot forage index}
\usage{
plot_forage_index(
  EPUs = "MAB",
  shadedRegion = c(2012, 2022),
  shade.fill = "lightgrey",
  shade.alpha = 0.3
)
}
\arguments{
\item{EPUs}{Character string. Which SOE report ("GB","MAB")}

\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot}

\item{shade.fill}{Character string. Color of shaded region. (Default = "lightgrey)}

\item{shade.alpha}{Numeric scalar. Alpha of shaded region (Default = 0.5)}
}
\value{
ggplot object
}
\description{
Description should be here. This needs to be reworked to uncouple GB and GOM
}
