% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gom_salmon.R
\name{plot_gom_salmon}
\alias{plot_gom_salmon}
\title{plots Gulf of Maine Salmon}
\usage{
plot_gom_salmon(shadedRegion = NULL, report = "NewEngland")
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10)}

\item{report}{Character string. Which SOE report ("MidAtlantic", "NewEngland")}
}
\value{
ggplot object
}
\description{
plots gom_salmon data set
}
