% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gsi.R
\name{plot_gsi}
\alias{plot_gsi}
\title{plot gulf stream index}
\usage{
plot_gsi(shadedRegion = NULL, report = "MidAtlantic", varName = "gsi", n = 10)
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10)}

\item{report}{Character string. Which SOE report ("MidAtlantic", "NewEngland")}

\item{varName}{Character string. Which variable to plot ("gsi","westgsi")}

\item{n}{Numeric scalar. Number of values used in short term trend}
}
\value{
ggplot object
}
\description{
plots gsi dataset. This is a shelfwide indicator
}
