% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_function_habs.R
\name{plot_habs}
\alias{plot_habs}
\title{plot harmful algal bloom}
\usage{
plot_habs(
  shadedRegion = shadedRegion,
  report = "MidAtlantic",
  varName = "Alexandrium"
)
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10)}

\item{report}{Character string. Which SOE report ("MidAtlantic", "NewEngland")}

\item{varName}{Character string. Which Variable to plot ("Alexandrium", "PSP")}
}
\value{
ggplot object
}
\description{
Plots alexandrium counts, PSP
}
