% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hudson_river_flow.R
\name{plot_hudson_river_flow}
\alias{plot_hudson_river_flow}
\title{plot flow rate of hudson river}
\usage{
plot_hudson_river_flow(shadedRegion = NULL, report = "MidAtlantic", n = 0)
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10)}

\item{report}{Character string. Which SOE report ("MidAtlantic", "NewEngland")}

\item{n}{Numeric scalar. Number of years used (from most recent year) to estimate short term trend . Default = 0 (No trend calculated)}
}
\value{
ggplot object
}
\description{
Plot hudson_river_flow data set
}
