% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mass_inshore_survey.R
\name{plot_mass_inshore_survey}
\alias{plot_mass_inshore_survey}
\title{plot massachusetts inshore data}
\usage{
plot_mass_inshore_survey(shadedRegion = shadedRegion, report = "MidAtlantic")
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10)}

\item{report}{Character string. Which SOE report ("MidAtlantic", "NewEngland")}
}
\value{
ggplot object
}
\description{
Plots mass_inshore_survey
}
