% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_narw.R
\name{plot_narw}
\alias{plot_narw}
\title{plot north atlantic right whale abundance}
\usage{
plot_narw(
  shadedRegion = NULL,
  report = "MidAtlantic",
  varName = "adult",
  n = 0
)
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10)}

\item{report}{Character string. Which SOE report ("MidAtlantic", "NewEngland")}

\item{varName}{Character string. Which variable to plot ("adult","calf")}

\item{n}{Numeric scalar. Number of years used (from most recent year) to estimate short term trend . Default = 0 (No trend calculated)}
}
\value{
ggplot object
}
\description{
plots narw data set. calf and adult abundance. This is a shelf wide product
}
