% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ne_inshore_survey.R
\name{plot_ne_inshore_survey}
\alias{plot_ne_inshore_survey}
\title{plot ME/NH inshore bottom trawl survey}
\usage{
plot_ne_inshore_survey(shadedRegion = NULL, report = "MidAtlantic", n = 0)
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10)}

\item{report}{Character string. Which SOE report ("MidAtlantic", "NewEngland")}

\item{n}{Numeric scalar. Number of years used (from most recent year) to estimate short term trend . Default = 0 (No trend calculated)}
}
\value{
ggplot object
}
\description{
Plot ne_inshore_survey.
}
