% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_observed_sharks.R
\name{plot_observed_sharks}
\alias{plot_observed_sharks}
\title{plot NE observer program shark data}
\usage{
plot_observed_sharks(shadedRegion = NULL, report = "MidAtlantic")
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10)}

\item{report}{Character string. Which SOE report ("MidAtlantic", "NewEngland")}
}
\value{
ggplot object
}
\description{
plots observed_shark data set
}
