% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ppr.R
\name{plot_ppr}
\alias{plot_ppr}
\title{Plot ecosystem overfishing indices}
\usage{
plot_ppr(
  shadedRegion = NULL,
  report = "MidAtlantic",
  EPU = "MAB",
  varName = "ryther",
  threshold = "global"
)
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10)}

\item{report}{Character string. Which SOE report ("MidAtlantic", "NewEngland")}

\item{EPU}{Character string. Which EPU for New England report ("GB", "GOM") Mid will always be "MAB"}

\item{varName}{Character string. Variable to plot ("pp","fogarty","ryther","ppr","mtl")}

\item{threshold}{Character string. Select how thresholds are calculated ("global","regional")}
}
\value{
ggplot object
}
\description{
Plots ppr dataset. Primary production (PP), Fogarty and Ryther indices. Primary production required scaled by Primary production and mean trophic level index
}
