% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ppr.R
\name{plot_ppr}
\alias{plot_ppr}
\title{plot ecosystem overfishing indices}
\usage{
plot_ppr(
  shadedRegion = NULL,
  report = "MidAtlantic",
  varName = "ryther",
  threshold = "global"
)
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10)}

\item{report}{Character string. Which SOE report ("MidAtlantic", "NewEngland")}

\item{varName}{Character string. Variable to plot ("pp","fogarty","ryther")}

\item{threshold}{Character string. Select how thresholds are calculated ("global","regional")}
}
\value{
ggplot object
}
\description{
plots ppr dataset. Primary production (PP), Fogarty, Ryther indices
}
