% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_recdat.R
\name{plot_recdat}
\alias{plot_recdat}
\title{plot recreational data}
\usage{
plot_recdat(
  shadedRegion = NULL,
  report = "MidAtlantic",
  varName = "landings",
  n = 0
)
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10)}

\item{report}{Character string. Which SOE report ("MidAtlantic", "NewEngland")}

\item{varName}{Character string. Variable to plot ("landings","effortdiversity","catchdiversity","effort")}

\item{n}{Numeric scalar. Number of years used (from most recent year) to estimate short term trend . Default = 0 (No trend calculated)}
}
\value{
ggplot object
}
\description{
Plot recdat data set, effort, diversity, and landings
}
