% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sandlance.R
\name{plot_sandlance}
\alias{plot_sandlance}
\title{plot sandlance numbers}
\usage{
plot_sandlance(shadedRegion = shadedRegion, report = "MidAtlantic")
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10)}

\item{report}{Character string. Which SOE report ("MidAtlantic", "NewEngland")}
}
\value{
ggplot object
}
\description{
Plots sandlance data.
}
