% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_setup.R
\name{plot_setup}
\alias{plot_setup}
\title{Setup all plot defaults}
\usage{
plot_setup(shadedRegion = shadedRegion, report = report)
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10), passed from plot function}

\item{report}{Character string. Which SOE report ("MidAtlantic", "NewEngland"), passed from plot function}
}
\value{
list of settings
}
\description{
Single function setting up plot defaults for SOE, catalog, presentations, and other products
}
