% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_species_groupings.R
\name{plot_species_groupings}
\alias{plot_species_groupings}
\title{plot species groupings by SOE year}
\usage{
plot_species_groupings(
  shadedRegion = NULL,
  report = "MidAtlantic",
  varName = "SOE.24"
)
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10)}

\item{report}{Character string. Which SOE report ("MidAtlantic", "NewEngland")}

\item{varName}{Character string. Which year of SOE report ("SOE.24", "SOE.20","SOE.18","SOE.17")}
}
\value{
flextable object
}
\description{
plots species_grouping dataset
}
