% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_thermal_habitiat_area.R
\name{plot_thermal_habitat_area}
\alias{plot_thermal_habitat_area}
\title{plot thermal habitat area}
\usage{
plot_thermal_habitat_area(
  shadedRegion = NULL,
  report = "MidAtlantic",
  EPU = "MAB"
)
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10)}

\item{report}{Character string. Which SOE report ("MidAtlantic", "NewEngland")}

\item{EPU}{Character string. Which EPU for New England report ("GB", "GOM") Mid will always be MAB}

\item{plottype}{character string. Which type of plots should be made. "daily" for daily area proportions or "annual"}
}
\value{
ggplot object
}
\description{
plots thermal_habitat_area data set.
}
