% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_thermal_habitat_gridded.R
\name{plot_thermal_habitat_gridded}
\alias{plot_thermal_habitat_gridded}
\title{plot thermal habitat persistence}
\usage{
plot_thermal_habitat_gridded(
  shadedRegion = NULL,
  report = "MidAtlantic",
  year = NULL,
  thresholds = c(15, 24),
  depths = "AllDepths"
)
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10)}

\item{report}{Character string. Which SOE report ("MidAtlantic", "NewEngland")}

\item{thresholds}{numeric vector for the temperature thresholds (0.5 degree increments from 0.5 to 30)}

\item{depths}{Character string. Which depth bands do you want to plot ('0-25m','25-100m','100-300m','AllDepths')}
}
\value{
ggplot object
}
\description{
plots thermal_habitat_gridded data set. Current SOE Year only
}
