% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trans_dates.R
\name{plot_trans_dates}
\alias{plot_trans_dates}
\title{plot SST transition dates and timing}
\usage{
plot_trans_dates(
  shadedRegion = shadedRegion,
  report = "MidAtlantic",
  varName = "timing"
)
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10)}

\item{report}{Character string. Which SOE report ("MidAtlantic", "NewEngland")}

\item{varName}{Character string. Which Variable to plot ("timing","length")}
}
\value{
ggplot object
}
\description{
uses ecodata::trans_dates
}
