% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wind_occupancy.R
\name{plot_wind_occupancy}
\alias{plot_wind_occupancy}
\title{plot probability of species occupancy in wind area}
\usage{
plot_wind_occupancy(shadedRegion = shadedRegion, report = "MidAtlantic")
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10)}

\item{report}{Character string. Which SOE report ("MidAtlantic", "NewEngland")}
}
\value{
kable object
}
\description{
plot wind_occupancy data. This is not region specific?
}
