% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wind_revenue.R
\name{plot_wind_revenue}
\alias{plot_wind_revenue}
\title{plot wind revenue}
\usage{
plot_wind_revenue(
  shadedRegion = NULL,
  report = "MidAtlantic",
  newVar = "landing"
)
}
\arguments{
\item{shadedRegion}{Numeric vector. Years denoting the shaded region of the plot (most recent 10)}

\item{report}{Character string. Which SOE report ("MidAtlantic", "NewEngland")}

\item{varName}{Character string. Which Variable to plot ("landing","value")}
}
\value{
ggplot object
}
\description{
plot wind_revenue data set. Only 5 species are plotted.
}
