% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppr.R
\docType{data}
\name{ppr}
\alias{ppr}
\title{Primary Production required}
\format{
63 rows and 5 columns.

\itemize{
\item EPU: Ecological Production Unit (EPU) where sampling occurred.
\item Time: Year.
\item Var: Primary Production Required.
\item Value: Value of variable \code{Var}.
\item Units: Units of variable \code{Var}.
}
}
\usage{
ppr
}
\description{
Data include percent of primary production required to support reported landings.
}
\details{
More information including processing and indicator derivation steps are available at
\url{https://noaa-edab.github.io/tech-doc/primary-production-required.html}.
}
\keyword{datasets}
