% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rec_hms.R
\docType{data}
\name{rec_hms}
\alias{rec_hms}
\title{Recreational shark fishing landings}
\format{
These data contain 246 rows and 4 columns.

\itemize{
\item Var: HMS species category.
\item Value: Value of variable \code{Var}.
\item Time: Time step of \code{Var}.
\item EPU: Ecological Production Unit (EPU) where sampling occurred.
}
}
\usage{
rec_hms
}
\description{
Shark landings reported using Marine Recreational Information Program (MRIP).
}
\details{
Methods used to calculate data are available
at \url{https://noaa-edab.github.io/tech-doc/recreational-shark-fishing-indicators.html}.
}
\keyword{datasets}
