% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{redCards}
\alias{redCards}
\title{redCards}
\format{
A 'data.frame':	146,028 obs. of  28 variables:
\describe{
  \item{playershort}{short player ID}
  \item{player}{player name}
  \item{club}{player club}
  \item{leagueCountry}{country of player club (England, Germany, France, and Spain)}
  \item{birthday}{player birthday}
  \item{height}{player height (in cm)}
  \item{weight}{player weight (in kg)}
  \item{position}{detailed player position}
  \item{games}{number of games in the player-referee dyad}
  \item{victories}{victories in the player-referee dyad}
  \item{ties}{ties in the player-referee dyad}
  \item{defeats}{losses in the player-referee dyad}
  \item{goals}{goals scored by a player in the player-referee dyad}
  \item{yellowCards}{number of yellow cards player received from referee}
  \item{yellowReds}{number of yellow-red cards player received from referee}
  \item{redCards}{number of red cards player received from referee}
  \item{photoID}{ID of player photo (if available)}
  \item{rater1}{skin rating of photo by rater 1 (5-point scale ranging from very light skin, to very dark skin)}
  \item{rater2}{skin rating of photo by rater 2 (5-point scale ranging from very light skin, to very dark skin)}
  \item{refNum}{unique referee ID number (referee name removed for anonymizing purposes)}
  \item{refCountry}{unique referee country ID number (country name removed for anonymizing purposes)}
  \item{Alpha_3}{unique referee country ID number (country name removed for anonymizing purposes)}
  \item{meanIAT}{mean implicit bias score (using the race IAT) for referee country, higher values correspond to faster white | good, black | bad associations}
  \item{nIAT}{sample size for race IAT in that particular country}
  \item{seIAT}{standard error for mean estimate of race IAT }
  \item{meanExp}{mean explicit bias score (using a racial thermometer task) for referee country, higher values correspond to greater feelings of warmth toward whites versus blacks}
  \item{nExp}{sample size for explicit bias in that particular country}
  \item{seExp}{standard error for mean estimate of explicit bias measure}
}
}
\source{
\url{https://journals.sagepub.com/doi/10.1177/2515245917747646}
}
\usage{
redCards
}
\description{
soccer redCards dataset Data
}
\examples{
redCards_df <- redCards
head(redCards_df)

str(redCards_df)

# preparing data -------------------------------------------------

redCards_df$mrater <- apply(FUN=mean, X=redCards_df[,which(grepl(x=colnames(redCards_df),pattern = "rater"))], MARGIN = 1)

redCards_df <- redCards_df[which(!is.na(redCards_df$mrater)),]
redCards_df$refNum = as.factor(redCards_df$refNum)

## ..................................

fit <- glmmTMB(redCards ~ scale(mrater) + (1|playerShort)  + (1|refNum),
               data = redCards_df, family= nbinom2,
               dispformula = ~ 1,
               ziformula = ~ 1)

summary(fit)


## ..................................
library(DHARMa)
res <- simulateResiduals(fit)
res


testQuantiles(res)
testOutliers(res)
testZeroInflation(res)
}
\author{
Andreas Ettner
}
\keyword{datasets}
