% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regrowth}
\alias{regrowth}
\title{Regrowth}
\format{
A 'data.frame':	40 obs. of  3 variables:
\describe{
  \item{Root}{Root biomass (num)}
  \item{Fruit}{Fruit biomass (num)}
  \item{Grazing}{Grazed or not (factor)}
}
}
\usage{
regrowth
}
\description{
Plant regrowth observations
}
\author{
Florian Hartig
}
\keyword{datasets}
