% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rikz}
\alias{rikz}
\title{RIKZ}
\format{
A 'data.frame':	45 obs. of  17 variables:
\describe{
  \item{Sample}{Index of samping station}
  \item{Polychaeta}{Number of species that were found of this faunal group}
  \item{Crustacea}{Number of species that were found of this faunal group}
  \item{Mollusca}{Number of species that were found of this faunal group}
  \item{Insecta}{Number of species that were found of this faunal group}
  \item{week}{In which of 4 weeks was this plot probed}
  \item{angle1}{Angle of the station}
  \item{angle2}{Slope of the beach at the plot}
  \item{exposure}{Index composed of wave action etc.}
  \item{salinity}{Salinity at the plot}
  \item{temperature}{}
  \item{NAP}{Altitude of the plot compared to the mean sea level}
  \item{penetrability}{}
  \item{grainsize}{Average diameter of sand grains}
  \item{humus}{Fraction of organic material}
  \item{chalk}{}
  \item{sorting1}{Variable from observational design}

}
}
\source{
\url{http://evol.bio.lmu.de/_statgen/Multivariate/11SS/RIKZGroups.txt}
}
\usage{
rikz
}
\description{
Data from the dutch National Institute for Coastal and Marine Management (RIKZ: Rijksinstituut voor Kust en Zee) to investigate the relationship between environmentalfactors and how they might affect benthic fauna.
}
\details{
Downloaded from https://www.uni-koblenz-landau.de/en/campus-landau/faculty7/environmental-sciences/landscape-ecology/Teaching/RIKZ_data/
}
\author{
Florian Hartig
}
\keyword{datasets}
