% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seal_pups.R
\docType{data}
\name{seal_pups}
\alias{seal_pups}
\title{Seal Pups}
\format{
Data set contains 77 rows and 6 columns

\itemize{
\item Time: Year.
\item Count: Number of seal pups.
\item Colony: Colony name..
\item fitted: GLM modelled values.
\item upper: Upper CV.
\item lower: Lower CV.
}
}
\usage{
seal_pups
}
\description{
Data include timeseires of seal pup count on four colonies.
}
\details{
More information including processing and indicator derivation steps are available at
\url{https://noaa-edab.github.io/tech-doc.html}.
}
\keyword{datasets}
