% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seedBank}
\alias{seedBank}
\title{Seed bank data}
\format{
An object of class \code{data.frame} with 2213 rows and 36 columns.
}
\usage{
seedBank
}
\description{
Data for species with seed traits, environmental factors, and information about their ability to form natural seed banks.
}
\details{
The dataset comes from a study from Rosbakh, 2022a (environmental factors were obtained from Rosbakh, 2022b) and includes information about plant species, their vegetative and seed traits, environmental factors, and their ability to form natural seed banks.

The objective of this dataset is to identify which traits and environmental factors are responsible if species can naturally produce a seed bank. The ability to form seed banks is encoded as presence-absence and as a density. Data was collected at 17 sites for 286 species.

\strong{Task:} Which environmental factors and traits explain the density or occurrence of a seed bank?

\strong{Tip:}\tabular{ll}{
   Effects \tab Variables \cr
   Response \tab SBPA, SBDensity \cr
   Fixed effects \tab Altitude, SeedMass, SeedShape, SeedN, SeedPr, DormRank, Temp, Hum, Nitrogen, Grazing \cr
   Random effects \tab Site, species \cr
}


\strong{Variables:}

Species information:

\itemize{
\item Species species
\item SBPA presence or absence of seed bank
\item SBDensity seed bank density
}

Environmental factors:

\itemize{
\item Site Site id (17 sites)
\item Altitude Altitude
\item Temp Landolt's temperature
\item Humidity Landolt's F
\item Nitrogen Landolt's N
\item Grazing Grazing
}

Vegetative traits:

\itemize{
\item SLA specific leaf aera
\item CanopyHeight canopy height
\item LeafN leaf nitrogen content
\item LeafC foliar frost sensitivity
}

Seed traits:

\itemize{
\item SeedMass	Seed morphology
\item SeedShape	Seed morphology
\item SeedN	Seed morphology
\item SeedPr	Seed morphology
\item SRH	Seed releasing height
\item TV	Terminal velocity
\item WingLoad	Wing loading
\item EpiStein	Attachement potential (Steinschaf)
\item EpiCow	Attachement potential (Cow)
\item EpiMerino	Attachement potential (Merino)
\item Endozoo	Potential for endozoochorous dispersal
\item DormRank	Dormancy 'depth'
\item T5	Base germination temperature
\item T50	Temperature for 50\% germination
\item T95	Temperature for 95\% germination
\item Jwidth	Germination niche width (temperature)
\item MGT	Mean germination time
\item MGR	Mean germiation rate
\item GSP	Germination speed
\item UNC	Synchrony of germination
\item SYN	Synchrony of germination
\item Endosperm Endosperm
}
}
\note{
\itemize{
\item start with lme4 (model checks are easier because of conditional simulations and the plotting option for residuals ~ random effects to identify random slopes (\verb{plot(model, fitted(.)~sAltitude | Site))})
\item use \code{dispformula}in glmmTMB if there is heteroskedasticity
\item scale your variables
}
}
\examples{
# SB density (for lm or lmm):
hist(seedBank$SBDensity)

# SB presence-absence (for glm or glmm)
table(seedBank$SBPA)

# Scale and remove NA:
data = as.data.frame(seedBank)
data$sAltitude = scale(data$Altitude)
data$sSeedMass = scale(data$SeedMass)
data$sSeedShape = scale(data$SeedShape)
data$sSeedN = scale(data$SeedN)
data$sSeedPr = scale(data$SeedPr)
data$sDormRank = scale(data$DormRank)
data$sTemp = scale(data$Temp)
data$sHum = scale(data$Humidity)
data$sNitro = scale(data$Nitrogen)
data$sGrazing = scale(data$Grazing)

rows = rownames(model.matrix(~sAltitude + sSeedMass + sSeedShape + sSeedN +
                               sSeedPr + sDormRank + sTemp + sHum + sNitro + 
                               sGrazing + Site + Species, data = data))
data = data[rows, ]
}
\references{
Rosbakh, S., Chalmandrier, L., Phartyal, S., & Poschlod, P. (2022). Inferring community assembly processes from functional seed trait variation along elevation gradient. Journal of Ecology.

Rosbakh, S., Pichler, M., & Poschlod, P. (2022). Machine‐learning algorithms predict soil seed bank persistence from easily available traits. Applied Vegetation Science, 25(2), e12660.
}
\author{
Maximilian Pichler
}
\keyword{datasets}
