% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{snails}
\alias{snails}
\title{Snails data}
\format{
An object of class \code{data.frame} with 2928 rows and 43 columns.
}
\usage{
snails
}
\description{
Temporal and spatial distribution of (infected) 4 freshwater snail species from 92 sites
}
\details{
The dataset comes from a study from Rabone et al., 2019 and includes information about the distribution of four snail species (healthy and infected) in the Niger River Valley. Freshwater snails are known to be intermediate hosts of schistosomiasis (a parasit).

The objective of Rabone et al., 2019 was to model the total abundance of the snail species as well as the proporation of the infected individuals. The hypothesis is that the total abundance and the infection depends on environmental and on seasonal factors. Data were collected at 92 (20 localities) sites over a 6-year period (2011-2016).

\strong{Task:}
Rabone et al., 2019 found differences for the species. Start with one species (total abundance + infection rate) and if you have time, go for other species.

\strong{Tip:}\tabular{ll}{
   Effects \tab Variables \cr
   Response \tab Bulinus_tot, BP_tot, BF_tot, BT_tot total \cr
   Fixed effects \tab Temp_Water, pH, water_speed_ms, water_depth, Cond, wmo_prec, locality, site_type, year \cr
   Random effects \tab Locality, site_irn, coll_date \cr
   offset \tab duration \cr
}


\strong{Variables:}

\itemize{
\item locality    nearby village. 20 in total
\item site_irn    site code unique in dataset
\item visit_no    number allocated to each monthly survey- e.g. 1 for first field mission in late july 2011- to 48 for last one in jan 2016
\item coll_date    collection date
\item month    month
\item Bulinus_tot    total of all Bulinus spp. collected
\item Bulinus_pos_tot    total of all Bulinus spp. positive for infection collected
\item BP_tot    Biomphalaria pfeifferi total count/collected
\item BP_pos_tot    Biomphalaria pfeifferi positive for infection total no collected
\item BF_tot    total of Bulinus forskalii collected
\item BF_pos_tot    total of Bulinus forskalii positive for infection collected
\item BF_prev    B. forskalii total positive/total collected *100 to give prevalence
\item BT_tot    total of Bulinus truncatus collected
\item BT_pos_tot    total of Bulinus truncatus positive for infection collected
\item BT_prev    B. truncatus total positive/total collected *100 to give prevalence
\item bp_pres    Biomphalaria pfeifferi present- n/y
\item bt_pres    Bulinus truncatus present- y/n
\item site_type    can.2= secondary irrigation canal, canal.3= tertiary irrigation canal, rice.p = rice paddy
\item L_tot    total of Lymnaea natalensis collected
\item Temp_Air    air temperature
\item Temp_Water    water temperature
\item water_speed_ms    water speed in m per sec
\item water_depth    water depth in cm
\item pH    pH reading from flow metre
\item Cond    conductivity in ppm
\item PPM    ppm water param data value
\item Latitude.verb    latitude verbatim
\item Longitude.verb    longitude verbatim
\item Stagnante    no flow y/n
\item wmo_min_temp    weather station day minimum air temp
\item wmo_max_temp    weather station day max air temp
\item wmo_av_temp    weather station day average air temp
\item wmo_prec    weather station precipication in cm
\item seas_wmo    season value adjusted to make may as part of wet season as often rain in this month (i.e. season acc to month of year AND WMO data).  wet season MAY- oct, dry nov- april)
\item duration    cleaned duration data- combining durat_verb and duration.amina
\item Heure    time of survey
\item water_level.v    water level as recorded in original dataset (low, med, high, in flood)
\item water_level3    water level edited to 3 levels only (some of in flood records the water depth is recorded as low)
\item tz tz
\item year year
\item Latitude latitude
\item Longitude longitude
}
}
\note{
\itemize{
\item start with lme4 (model checks are easier because of conditional simulations and the plotting option for residuals \~ random effects to identify random slopes (\verb{plot(model, fitted(.)~sTemp_Water | coll_date))})
\item check for underdispersion
\item check for temporal or spatial trends/autocorrelation
\item scale your variables
}
}
\examples{
# Abundance of bulinus
hist(snails$Bulinus_tot)

# Infection rates:
hist(snails$Bulinus_pos_tot / snails$Bulinus_tot)

# Scale and remove NA:
data = snails 
data$sTemp_Water = scale(data$Temp_Water)
data$spH = scale(data$pH)
data$swater_speed_ms = scale(data$water_speed_ms)
data$swater_depth = scale(data$water_depth)
data$sCond = scale(data$Cond)
data$swmo_prec = scale(data$wmo_prec)
data$syear = scale(data$year)
data$sLat = scale(data$Latitude)
data$sLon = scale(data$Longitude)

rows = rownames(model.matrix(~sTemp_Water + spH + sLat + sLon + sCond + 
                               seas_wmo+ swmo_prec + swater_speed_ms + 
                               swater_depth + syear + duration + locality + 
                               site_irn + coll_date, data = data))
data = data[rows, ]
}
\references{
Rabone, M., Wiethase, J. H., Allan, F., Gouvras, A. N., Pennance, T., Hamidou, A. A., ... & Rollinson, D. (2019). Freshwater snails of biomedical importance in the Niger River Valley: evidence of temporal and spatial patterns in abundance, distribution and infection with Schistosoma spp. Parasites & vectors, 12(1), 1-20.
}
\author{
Maximilian Pichler
}
\keyword{datasets}
