% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{solomonislands}
\alias{solomonislands}
\title{Solomon Islands - Island Biogeography}
\format{
A 'data.frame': 40 obs. of 3 variables:
\describe{
 \item{Island}{Island name (character).}
 \item{Species}{Species richness on the respective island (numeric).}
 \item{Area}{Island area (numeric).}
 \item{Distance}{Distance to mainland (numeric).}
}
}
\usage{
solomonislands
}
\description{
A classical data set of island biogeography - Number of species found on the Solomon islands, as a function of distance to mainland and island area
}
\examples{
solomonIsland_df <- solomonislands
head(solomonIsland_df)


# factorize variables -------------------------------------------------
solomonIsland_df$Island = as.factor(solomonIsland_df$Island)


# plotting variables -------------------------------------------------
plot(solomonIsland_df$Species, 1:nrow(solomonIsland_df))
plot(solomonIsland_df$Island, solomonIsland_df$Species)

fit <- glm(Species ~ Area + Distance, data=solomonIsland_df, family = poisson)
summary(fit)
image(xtabs(residuals(fit)~Area + Distance, data=solomonIsland_df))
}
\author{
Florian Hartig
}
\keyword{datasets}
