% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_groupings.R
\docType{data}
\name{species_groupings}
\alias{species_groupings}
\title{Species groupings}
\format{
1022 rows and 15 columns

\itemize{
\item COMNAME: Common name of listed species.
\item SVSPP: NEFSC BTS database species codes.
\item ITISSPP: Integrated Taxonomic Information System (ITIS) species identifier.
\item NESPP3: Species codes used in CFDBS.
\item SCINAME: Scientific name of listed species.
\item Fed.managed: Management body. One of \code{NEFMC}, \code{MAFMC}, \code{JOINT} (MAFMC/NEFMC), or \code{NA}.
\item SOE.17: Feeding guild groupings used in 2017-2018 SOE reports.
\item EMAX: Groupings adapted from Link et al. 2006.
\item RPATH: Groupings from Rpath - the R version of the Ecopath with Ecosim model.
\item SOE.18: Feeding guild groupings used in 2018-2019 SOE reports.
\item SizeCat: Size category.
\item Min.size: Minimum size.
\item Garrison.Link: Groupings adapted from Garrison and Link 2000.
\item NEIEA: Simplified feeding guilds for NE-IEA website.
\item SOE.20: Feeding guild groupings used in 2020+ SOE reports.
}
}
\usage{
species_groupings
}
\description{
Species groupings from a variety of NEFSC EDAB products are included in this data set.
}
\details{
More information about these species groupings may be found at \url{https://noaa-edab.github.io/tech-doc/aggroups.html}.
}
\references{
Garrison, Lance P, and Jason S Link. 2000. “Dietary guild structure of the fish community in the Northeast United
States continental shelf ecosystem.” \emph{Marine Ecology Progress Series} 202:231–40.

Link, Jason S, Carolyn A Griswold, Elizabeth T Methratta, and Jessie Gunnard. 2006. Documentation
for the energy modeling and analysis exercise (EMAX). US Department of Commerce,
National Oceanic and Atmospheric Administration.

Lucey, S. 2019. Rpath - the R version of the Ecopath with Ecosim model. US Department of Commerce,
National Oceanic and Atmospheric Administration. \url{https://github.com/slucey/RpathDev}.
}
\keyword{datasets}
