% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecodata.R
\name{str_extract}
\alias{str_extract}
\title{`str_extract()`
Wrapper for `stringr::str_extract()`}
\usage{
str_extract(...)
}
\description{
`str_extract()`
Wrapper for `stringr::str_extract()`
}
\seealso{
[stringr::str_extract()]
}
