% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecodata.R
\name{string_detect_any}
\alias{string_detect_any}
\title{`string_detect_any()`
Discover whether any of the strings in `str` contain `pattern`}
\usage{
string_detect_any(str, pattern)
}
\arguments{
\item{str}{Vector of strings to search through}

\item{pattern}{Pattern to search for}
}
\value{
Logical on whether the pattern was found in any of the strings
}
\description{
`string_detect_any()`
Discover whether any of the strings in `str` contain `pattern`
}
