% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecodata.R
\name{string_which}
\alias{string_which}
\title{Get index into string where a pattern if found}
\usage{
string_which(str, pattern)
}
\arguments{
\item{str}{String or vector of strings to search through}

\item{pattern}{Pattern to look for}
}
\value{
Integer index into the string that the pattern was found
}
\description{
Wrapper for `str_which()` while ignoring case, returns what index into the string the pattern is found
}
