% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_all.R
\name{update_all}
\alias{update_all}
\title{get_buoy is a wrapper for rnoaa::buoy, to access the National Buoy Data Center}
\usage{
update_all()
}
\value{
null
}
\description{
get_buoy is a wrapper for rnoaa::buoy, to access the National Buoy Data Center
}
\examples{
\dontrun{
update_all()
}
}
