% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{volcanoisland}
\alias{volcanoisland}
\title{Volcano island}
\format{
A 'data.frame':	1000 obs. of  21 variables:
}
\usage{
volcanoisland
}
\description{
Ecological observatins on a lonely volcano island.
}
\details{
We are looking at the Maunga Whau (Mt Eden), one of about 50 volcanos in the Auckland volcanic field.

A crazy ecologist has created 100 permanent beetle observation plots arcross the volcano (random), and monitored them for 10 years, so you have one observation each year. Below a visualization of the terrain

We have also a bunch of other things going on. For example,

* Actually, we have two types of beetles. The second type is much more complicated, I'm curious if you find out what's going on with them

* There is the wind, which might depend on other variables

* There are lizards on the island that eat beetles. The lizards always live at the same place, however, they are hard to spot, so that the ecologist didn't always notice them. However, every time they were spotted, they were also observed.

* Moreover, the ecologist also did a beetle survival study each year at each site. Don't ask me how he did it, but you have get for each year and each site how many of 20 beetles survived.

The main question of the study was to understand how beetle abudance depends on altitude. Here is an overview of counts per site and year.
}
\examples{
oldpar <- par(mfrow = c(1,2))
contour(x = 1:ncol(volcano), y = 1:nrow(volcano), t(volcano), asp = 1)
points(volcanoisland$x, volcanoisland$y, cex = volcanoisland$wind)
persp(volcano)
par(oldpar)

with(volcanoisland, {
  
  plot(year, altitude, cex = beetles/5, pch =2, main = "Beetle counts across altitudinal gradient, triangle is proportional to counts")
  #lines(plot, altitude * 1000)
  #points(unique(plot), unique(altitude * 1000) , pch = 3)
  #text(unique(plot), unique(altitude * 1000) - 50, unique(plot), cex = 0.7 )
})
}
\author{
Florian Hartig
}
\keyword{datasets}
