% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcr.R
\docType{data}
\name{wcr}
\alias{wcr}
\title{Warm core ring births}
\format{
Data set contains 43 rows and 5 columns

\itemize{
\item Time: Year.
\item Var: Number of warm core rings.
\item Value: Value of variable  \code{Var}.
\item Units: Unit of Variable  \code{Var}.
\item EPU: Ecological Production Unit (EPU) where sampling occurred.
}
}
\usage{
wcr
}
\description{
Data include timeseries of warm core ring births on the Northeast U.S. Shelf.
}
\details{
More information including processing and indicator derivation steps are available at
\url{https://noaa-edab.github.io/tech-doc/warm-core-rings.html}.
}
\keyword{datasets}
