% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wea_landings_rev.R
\docType{data}
\name{wea_landings_rev}
\alias{wea_landings_rev}
\title{Fisheries maximum and minimum landings and revenue from wind energy areas (WEAs).}
\format{
Data set contains 32 rows and 5 columns

\itemize{
\item NEFMC, MAFMC, and ASMFC Managed Species: Managed Fishery.
\item perc_landings_max: Max Landings Percent from WEA.
\item perc_revenue_max: Max Revenue Percent from WEA.
\item Units: Units for landings or revenue.
\item Council: New England (NEFMC) or Mid Atlantic (MAFMC)
}
}
\usage{
wea_landings_rev
}
\description{
Data include top ten species Landings and Revenue from Wind Energy Areas.
}
\details{
More information including processing and indicator derivation steps are available at
\url{https://noaa-edab.github.io/tech-doc/fisheries-revenue-in-wind-development-areas.htm}.
}
\keyword{datasets}
