% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wine}
\alias{wine}
\title{Wine}
\format{
An object of class \code{data.frame} with 1599 rows and 12 columns.
}
\usage{
wine
}
\description{
Wine data from \url{https://archive.ics.uci.edu/ml/datasets/wine+quality}.
}
\details{
The data set is a collection of wines of different qualities. The wines are described by physiochemical variables.

\describe{
 \item{fixed.acidity}{Fixed acidity (numeric).}
 \item{volatile.acidity}{Volatile acidity (numeric).}
 \item{citric.acid}{Citric acid (numeric).}
 \item{residual.sugar}{Residual sugar (numeric).}
 \item{chlorides}{Amount of chlorides measured (numeric).}
 \item{free.sulfur.dioxide}{Amount of free sulfur dioxide measured (numeric).}
 \item{total.sulfur.dioxide}{Amount of total sulfur dioxide measured (numeric).}
 \item{density}{Density.}
 \item{pH}{pH of the wine.}
 \item{sulphates}{Sulphates in the wine.}
 \item{alcohol}{Amount of alcohol measured.}
 \item{quality}{Response variable, ranges between 1 and 8.}
}
}
\author{
Maximilian Pichler
}
\keyword{datasets}
