% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zooplankton_index.R
\docType{data}
\name{zooplankton_index}
\alias{zooplankton_index}
\title{Zooplankton Index}
\format{
n rows and 5 columns.

\itemize{
\item Value: Value of variable \code{Var}.
\item Var: Variable calculated.
\item Time: Year.
\item EPU: Ecological Production Unit (EPU) where sampling occurred.
\item Units: grams per stomach.
}
}
\usage{
zooplankton_index
}
\description{
Data include Zooplankton (Calanus finmarchicus, Eupausiids, Smallcopesoe, Lgcopeall, Zoopvol, Smallcopeall) biomass index from predator diet data from NEFSC bottom trawl survey.
}
\details{
Methods used to calculate data are available at \url{https://noaa-edab.github.io/tech-doc/zooplankton_index.html}.
}
\keyword{datasets}
